% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_Al2O3C_ITC.R
\name{analyse_Al2O3_ITC}
\alias{analyse_Al2O3_ITC}
\title{Al2O3 Irradiation Time Correction Analysis}
\usage{
analyse_Al2O3_ITC(object, signal_integral = NULL, dose_points = c(2, 4, 8,
  12, 16), method_control = NULL, verbose = TRUE, plot = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{RLum.Analysis}} \bold{(required)}: measurement input}

\item{signal_integral}{\code{\link{numeric}} (optional): signal integral, used for the signal
and the background. If nothing is provided the full range is used}

\item{dose_points}{\code{\link{numeric}} (with default): vector with dose points, if dose points
are repeated, only the general pattern needs to be provided. Default values follow the suggestions
made by Kreutzer et al., 2017}

\item{method_control}{\code{\link{list}} (optional): optional parameters to control the calculation.
See details for further explanations}

\item{verbose}{\code{\link{logical}} (with default): enable/disable verbose mode}

\item{plot}{\code{\link{logical}} (with default): enable/disable plot output}

\item{...}{further aruments that can be passed to the plot output}
}
\value{
Function returns results numerically and graphically:\cr

-----------------------------------\cr
[ NUMERICAL OUTPUT ]\cr
-----------------------------------\cr
\bold{\code{RLum.Reuslts}}-object\cr

\bold{slot:} \bold{\code{@data}}\cr
\tabular{lll}{
\bold{Element} \tab \bold{Type} \tab \bold{Description}\cr
 \code{$data} \tab \code{data.frame} \tab correction value and error \cr
 \code{$table} \tab \code{data.frame} \tab table used for plotting  \cr
 \code{$table_mean} \tab \code{data.frame} \tab table used for fitting
}

\bold{slot:} \bold{\code{@info}}\cr

The original function call\cr

------------------------\cr
[ PLOT OUTPUT ]\cr
------------------------\cr

\itemize{
 \item A dose response curve with the marked correction values

}
}
\description{
The function provides a very particular analysis to correct the irradiation time while irradiating Al2O3:C
pellets in a luminescence reader.
}
\details{
Background: Due to their high dose sensitivity Al2O3:C pellets are usually irradiated for only a very short duration or
under the closed beta-source within a luminescence reader. However, due to its high dose sensitivity, the movement
itself below the source induces an apparent luminescence signal, which can be translated to an irradiation time.
Based on measurements following a protocol suggested by Kreutzer et al., 2017, a dose response curve is constructed
and the intersection with the time axis is taken as real irradiation time.
}
\section{Function version}{
 0.1.0
}

\examples{

##nothing so far TODO ... add tests with example

} 

\section{How to cite}{
Kreutzer, S. (2017). analyse_Al2O3_ITC(): Al2O3 Irradiation Time Correction Analysis. Function version 0.1.0. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence
}

\references{
TODO
}
\seealso{
\code{\link{plot_GrowthCurve}}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\keyword{datagen}
