% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_SAR.NCF.R
\name{analyse_SAR.NCF}
\alias{analyse_SAR.NCF}
\title{Compute SAR palaeodoses using natural sensitivity correction (NCF)}
\usage{
analyse_SAR.NCF(object, TL_peak_range = 15, method_control = NULL, ...)
}
\arguments{
\item{object}{\link[=RLum.Analysis-class]{RLum.Analysis} or \link{list} of
\link[=RLum.Analysis-class]{RLum.Analysis} objects (\strong{required}):
input object containing data for analysis.
If a \link{list} is provided the functions tries to iterate over each element
in the list.}

\item{TL_peak_range}{\link{numeric} (\strong{with default}):
size of the integration window in deg. C on each side of the identified peak.}

\item{method_control}{\link{list} (\emph{optional}):
parameters to control the peak-finding step (see Details).}

\item{...}{further arguments that will be passed to
\link{fit_DoseResponseCurve}, \link{plot_DoseResponseCurve}
or \link{calc_OSLLxTxRatio} (the latter only supports
\code{background.count.distribution}, \code{sigmab}, \code{sig0}, \code{od_rates}).
Additionally, supported are \code{legend.cex} and \code{legend.pch} to modify the
legend symbols.
\strong{Note:} If you consider using the early light subtraction method,
\code{sigmab} should be provided.

\strong{Note:} \code{od_rates} can be used to treat uncertainties in Lx/Tx according
to Bluszcz et al. (2015) instead of the standard approach of Galbraith
(2002, 2014). See \link{calc_OSLLxTxRatio} for details.}
}
\value{
Plots (\emph{optional}) and an \link[=RLum.Results-class]{RLum.Results} object is
returned containing the following elements:

\tabular{lll}{
\strong{DATA.OBJECT} \tab \strong{TYPE} \tab \strong{DESCRIPTION} \cr
\code{..$data} : \tab  \code{data.frame} \tab Table with corrected \eqn{D_e} values \cr
\code{..$LnLxTnTx.table} : \tab \code{data.frame} \tab Corrected \code{LnLxTnTx} values \cr
\code{..$rejection.criteria} : \tab \code{data.frame} \tab Rejection criteria \cr
\code{..$Formula} : \tab list \tab Function used for fitting of the dose response curve \cr
\code{..$data_uncor} : \tab  \code{data.frame} \tab Table with the uncorrected data \cr
\code{..$LnLxTnTx.table_uncor} : \tab \code{data.frame} \tab Uncorrected \code{LnLxTnTx} values\cr
\code{..$NCF_settings} : \tab list \tab Setting used for TL peak finding \cr
}

This output is similar to the output of \link{analyse_SAR.CWOSL}
extended by a few more elements.
}
\description{
The function computes the natural sensitivity correction factor (NCF) for
the assessment of single aliquot regeneration based palaeodoses. This allows
to produce \eqn{D_e} values that account for sensitivity changes that
may occur during the preheat and readout of the natural OSL.
}
\details{
This function is a wrapper around \link{analyse_SAR.CWOSL}: it
works in a similar way, but corrects the natural OSL signal according to
Singhvi et al. (2011) and by implementing the MatLab code reported in
Kaushal et al. (2022).
\subsection{The NCF procedure}{

The NCF protocol modifies the standard SAR protocol by adding a small dose
before the readout of the natural signal; the corresponding 110 deg. C TL
peak in quartz is monitored and compared to the TL peak recorded during the
typical cutheat measurement. The additional dose used for the NCF TL curves
is subtracted automatically. The published NCF sequence structure must be
followed, otherwise the correction will not work.

The natural sensitivity correction factor (NCF) is measured as the ratio of
the 110 deg. C TL peaks before and after the measurement of natural OSL,
accounting for an integration range around the peak, typically 15 deg. C or
the full width at half maximum (FWHM).
}

\subsection{TL peak finding}{

The TL peaks are found automatically by finding the position  with highest
counts below 160 deg. C (after some smoothing has been applied to the counts
data) for each of the two TL curves.

The peak search parameters can be controlled via the \code{method_control}
argument. If used, this must be a list with one or more of the following
named arguments (e.g. \code{method_control = list(search.threshold.degrees = 130)}):
\itemize{
\item \code{peak.temperature}: this can be used to set the temperature of the TL peak
manually; if set to \code{NULL} (default), the peak is found automatically
\item \code{search.threshold.degrees}: threshold in deg. C below which the peak
is searched (160 by default); this is ignored if \code{peak.temperature} is
not \code{NULL}
}

The identified TL curves and the chosen signal integral are shown
in the graphical output.
}
}
\note{
This function is a beta version.
}
\section{Function version}{
 0.1.1
}

\examples{
## Load example data
file <- system.file("extdata/NCF.binx", package = "Luminescence")
ncf <- read_BIN2R(file, fastForward = TRUE)

results <- analyse_SAR.NCF(
 object = ncf,
 signal_integral = 1:2,
 background_integral = 100:250,
 dose_rate_source = 0.1)

}
\references{
Kaushal, R.K., Chauhan, N., Singhvi, A.K., 2022. Luminescence dating of
quartz: A MATLAB-based program for computation of SAR paleodoses using
natural sensitivity correction (NCF). Ancient TL, 40 (2), 1-7.
\doi{10.26034/la.atl.2022.561}

Singhvi, A.K., Stokes, S.C., Chauhan, N., Nagar, Y.C., Jaiwal, M.K., 2011.
Changes in natural OSL sensitivity during single aliquot regeneration
procedure and their implications for equivalent dose determination.
Geochronometria, 38 (3), 231-241. \doi{10.2478/s13386-011-0028-3}
}
\seealso{
\link{analyse_SAR.CWOSL}, \link{fit_DoseResponseCurve}
}
\author{
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
Sebastian Kreutzer, LIAG Institute for Applied Geophysics (Germany)
}
\keyword{datagen}
\keyword{plot}
