% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_ThermochronometryData.R
\name{analyse_ThermochronometryData}
\alias{analyse_ThermochronometryData}
\title{Analyse Thermochronometry Data}
\usage{
analyse_ThermochronometryData(object, plot = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{object}{\link{character} or \link{RLum.Results} (\strong{required}): file path to XLSX file with ... ##TODO}

\item{plot}{\link{logical} (\emph{with default}): enable/disable plot output}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable terminal output}

\item{...}{further arguments passed to ...}
}
\value{
An \linkS4class{RLum.Results} object is returned: ##TODO
}
\description{
Analyse thermochronometry data based on MatLab code written
by XXX and XXX and XXX
}
\details{
##TODO
}
\section{Function version}{
 0.1.0
}

\examples{
# example code ##TOD

}
\seealso{
##TODO
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany), Svenja Riedesel, DTU Risø (Denmark)
}
\keyword{datagen}
