% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_RLum.R
\name{app_RLum}
\alias{app_RLum}
\title{Run Luminescence shiny apps (wrapper)}
\usage{
app_RLum(app = NULL, ...)
}
\arguments{
\item{app}{\link{character} (required): name of the application to start. See details for a list
of available apps.}

\item{...}{further arguments to pass to \code{\link[shiny]{runApp}}}
}
\description{
Wrapper for the function \link[RLumShiny:app_RLum]{RLumShiny::app_RLum} from the package
\link[RLumShiny:RLumShiny-package]{RLumShiny::RLumShiny-package}. For further details and examples please
see the manual of this package.
}
\section{Function version}{
 0.1.1
}

\author{
Christoph Burow, University of Cologne (Germany)
}
