% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_EfficiencyCorrection.R
\name{apply_EfficiencyCorrection}
\alias{apply_EfficiencyCorrection}
\title{Apply spectral efficiency correction to RLum.Data.Spectrum objects}
\usage{
apply_EfficiencyCorrection(object, spectral.efficiency)
}
\arguments{
\item{object}{\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} or \link[=RLum.Analysis-class]{RLum.Analysis} (\strong{required}):
object of class \code{RLum.Data.Spectrum}, \code{RLum.Analysis}or a \link{list} of such
objects. Other objects in the list are skipped.}

\item{spectral.efficiency}{\link{data.frame} or \link{character} (\strong{required}):
either a data frame with 2 columns containing wavelengths and relative
spectral response values (values between 0 and 1); or a character string
specifying the path to a CSV file (with a header row), which the function
will attempt to import via \link[utils:read.csv]{utils::read.csv}. When \code{object} is a \link{list},
the provided data will be used to correct all spectra.}
}
\value{
Returns same object as provided as input
}
\description{
The function applies spectral efficiency correction to
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects.
}
\details{
The efficiency correction is based on a spectral response dataset provided
by the user. Usually the data set for the quantum efficiency is of lower
resolution and values are interpolated for the required spectral resolution using
function \link[stats:approx]{stats::approx}.

If the energy calibration differs for both data set \code{NA} values are produces that
will be removed from the matrix.
}
\note{
Please note that the spectral efficiency data from the camera alone may not
sufficiently correct for spectral efficiency of the entire optical system
(e.g., spectrometer, camera ...).
}
\section{Function version}{
 0.3.0
}

\examples{

##(1) - use with your own data (uncomment for usage)
## spectral.efficiency <- read.csv("your data")
##
## your.spectrum <- apply_EfficiencyCorrection(your.spectrum, )

}
\seealso{
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}, \link[=RLum.Analysis-class]{RLum.Analysis}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Johannes Friedrich, University of Bayreuth (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., Friedrich, J., 2026. apply_EfficiencyCorrection(): Apply spectral efficiency correction to RLum.Data.Spectrum objects. Function version 0.3.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{manip}
