% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.Analysis-class.R,
%   R/RLum.Data.Curve-class.R, R/RLum.Data.Image-class.R,
%   R/RLum.Data.Spectrum-class.R, R/RLum.Results-class.R
\name{as}
\alias{as}
\title{as() - RLum-object coercion}
\usage{
\S4method{coerce}{list,RLum.Analysis}(from, to = "RLum.Analysis", strict = TRUE)

\S4method{coerce}{list,RLum.Data.Curve}(from, to = "RLum.Data.Curve", strict = TRUE)

\S4method{coerce}{data.frame,RLum.Data.Image}(from, to = "RLum.Data.Image", strict = TRUE)

\S4method{coerce}{data.frame,RLum.Data.Spectrum}(from, to = "RLum.Data.Spectrum", strict = TRUE)

\S4method{coerce}{list,RLum.Results}(from, to = "RLum.Results", strict = TRUE)
}
\arguments{
\item{from}{\link[=RLum-class]{RLum}, \link{list}, \link{data.frame}, \link{matrix} (\strong{required}):
object to be coerced from}

\item{to}{\link{character} (\strong{required}):
class name to be coerced to}
}
\description{
for \verb{[RLum.Analysis-class]}

for \verb{[RLum.Data.Curve-class]}

for \verb{[RLum.Data.Image-class]}

for \verb{[RLum.Data.Spectrum-class]}

for \verb{[RLum.Results-class]}
}
\details{
\strong{\link[=RLum.Analysis-class]{RLum.Analysis}}

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{list} \tab \code{list}\cr
}

Given that the \link{list} consists of \link[=RLum.Analysis-class]{RLum.Analysis} objects.

\strong{\link[=RLum.Data.Curve-class]{RLum.Data.Curve}}

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{list} \tab \code{list} \cr
\code{data.frame} \tab \code{data.frame}\cr
\code{matrix} \tab \code{matrix}
}

\strong{\link[=RLum.Data.Image-class]{RLum.Data.Image}}

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{data.frame} \tab \code{data.frame}\cr
\code{matrix} \tab \code{matrix}
}

\strong{\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}}

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{data.frame} \tab \code{data.frame}\cr
\code{matrix} \tab \code{matrix} \cr
\code{list} \tab \code{list}
}

\strong{\link[=RLum.Results-class]{RLum.Results}}

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{list} \tab \code{list}\cr
}

Given that the \link{list} consists of \link[=RLum.Results-class]{RLum.Results} objects.
}
\note{
Due to the complex structure of the \code{RLum} objects itself a coercing to standard
R data structures will be always loosely!
}
\seealso{
\link[methods:as]{methods::as}
}
