% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\name{BaseDataSet.FractionalGammaDose}
\alias{BaseDataSet.FractionalGammaDose}
\title{Base data set of fractional gamma-dose values}
\format{
A \code{\link{list}} with fractional gamma dose-rate values
sorted by article:

\tabular{ll}{

\code{Aitken1985}: \tab
Fractional gamma-dose values from table H.1
}
}
\source{
Fractional gamma dose values were carefully read from the tables given in the
references above.
}
\description{
Collection of (un-)published fractional gamma dose-rate values to scale the
gamma-dose rate considering layer-to-layer variations in soil radioactivity.
}
\section{Version}{
 0.1
}

\examples{

## Load data
data("BaseDataSet.FractionalGammaDose", envir = environment())

}
\references{
Aitken, M.J., 1985. Thermoluminescence Dating. Academic Press, London.
}
\keyword{datasets}
