% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\name{BaseDataSet.GrainSizeAttenuation}
\alias{BaseDataSet.GrainSizeAttenuation}
\title{Base dataset for grain size attenuation data by Guérin et al. (2012)}
\source{
Guérin, G., Mercier, N., Nathan, R., Adamiec, G., Lefrais, Y., 2012.
On the use of the infinite matrix assumption and associated concepts:
A critical review. Radiation Measurements, 47, 778-785.
}
\description{
Grain size correction data for beta-dose rates
published by Guérin et al. (2012).

#' @format

A \code{\link{data.frame}} seven columns and sixteen rows. Column headers
are \code{GrainSize}, \code{Q_K}, \code{FS_K}, \code{Q_Th}, \code{FS_Th}, \code{Q_U}, \code{FS_U}.
Grain sizes are quoted in µm (e.g., 20, 40, 60 etc.)
}
\section{Version}{
 0.1.0
}

\examples{

## load data
data("BaseDataSet.GrainSizeAttenuation", envir = environment())

}
\keyword{datasets}
