% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Data.Curve-class.R, R/RLum.Data.Spectrum-class.R
\name{bin_RLum.Data}
\alias{bin_RLum.Data}
\alias{bin_RLum.Data,RLum.Data.Curve-method}
\alias{bin_RLum.Data,RLum.Data.Spectrum-method}
\title{Channel binning for RLum.Data-class objects}
\usage{
bin_RLum.Data(object, ...)

\S4method{bin_RLum.Data}{RLum.Data.Curve}(object, bin_size = 2)

\S4method{bin_RLum.Data}{RLum.Data.Spectrum}(object, bin_size.col = 1, bin_size.row = 1)
}
\arguments{
\item{object}{\link[=RLum.Data-class]{RLum.Data} (\strong{required}):
S4 object of class \code{RLum.Data}.}

\item{...}{further arguments passed to the specific class method.}

\item{bin_size}{\link{integer} (\emph{with default}):
number of channels used for each bin, e.g. \code{bin_size = 2} means that
two channels are binned.}

\item{bin_size.col}{\link{integer} (\emph{with default}):
number of channels used for each bin, e.g. \code{bin_size.col = 2} means that
two channels are binned. Note: The function does not check the input, very large values
mean a full column binning (a single sum)}

\item{bin_size.row}{\link{integer} (\emph{with default}):
number of channels used for each bin, e.g. \code{bin_size.row = 2} means that
two channels are binned. Note: The function does not check the input, very large values
mean a full row binning (a single sum)}
}
\value{
An object of the same class as the input.
}
\description{
The function aggregates adjacent channels of \link[=RLum.Data-class]{RLum.Data}
objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{bin_RLum.Data(RLum.Data.Curve)}: Allows binning of RLum.Data.Curve data.

\item \code{bin_RLum.Data(RLum.Data.Spectrum)}: Allows binning of RLum.Data.Spectrum data. Count values and values on the
x-axis are summed up; for wavelength/energy values, the mean is calculated.

}}
\note{
Currently only \code{RLum.Data} objects of class \link[=RLum.Data.Curve-class]{RLum.Data.Curve}
and \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} are supported.
}
\section{Function version}{
 0.2.0
}

\examples{

## load example data
data(ExampleData.CW_OSL_Curve, envir = environment())

## create RLum.Data.Curve object from this example
curve <-
  set_RLum(
      class = "RLum.Data.Curve",
      recordType = "OSL",
      data = as.matrix(ExampleData.CW_OSL_Curve)
  )

## plot data without and with 2 and 4 channel binning
plot_RLum(curve)
plot_RLum(bin_RLum.Data(curve, bin_size = 2))
plot_RLum(bin_RLum.Data(curve, bin_size = 4))

}
\seealso{
\link[=RLum.Data.Curve-class]{RLum.Data.Curve}, \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{utilities}
