% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gSGC.R
\name{calc_gSGC}
\alias{calc_gSGC}
\title{Calculate De value based on the gSGC by Li et al., 2015}
\usage{
calc_gSGC(
  data,
  gSGC.type = c("0-250", "0-450"),
  gSGC.parameters = NULL,
  n.MC = 100,
  verbose = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{\link{data.frame} (\strong{required}):
input data  the following columns five columns in the given order:
\code{LnTn}, \code{LnTn.error}, \code{Lr1Tr1}, \code{Lr1Tr1.error}, \code{Dr1}. Column names are
not required.}

\item{gSGC.type}{\link{character} (\emph{with default}):
function parameters to use for the iteration procedure, either \code{"0-450"} or
\code{"0-250"}, as presented in Li et al., 2015 (Table 2). This is ignored if
\code{gSGC.parameters} is set.}

\item{gSGC.parameters}{\link{list} (\emph{optional}):
option to provide own function parameters used for fitting as named list.
Nomenclature follows Li et al., 2015, i.e. \code{list(A, A.error, D0, D0.error, c, c.error, Y0, Y0.error, range)}, where \code{range} is defines the interval
where the function is considered as valid, e.g. \code{range = c(0,250)}.\cr
If set, option \code{gSGC.type} will be ignored.}

\item{n.MC}{\link{integer} (\emph{with default}):
number of Monte Carlo simulation runs for error estimation, see details.}

\item{verbose}{\link{logical} (\emph{with default}):
enable/disable output to the terminal.}

\item{plot}{\link{logical} (\emph{with default}):
enable/disable the plot output.}

\item{...}{parameters will be passed to the plot output}
}
\value{
Returns an S4 object of type \link[=RLum.Results-class]{RLum.Results}.

\strong{\verb{@data}}\cr
\verb{$ De.value} (\link{data.frame}) \cr
\code{.. $ De}  \cr
\code{.. $ De.error} \cr
\code{.. $ Eta} \cr
\verb{$ De.MC} (\link{list}) contains the matrices from the error estimation.\cr
\verb{$ uniroot} (\link{list}) contains the \link{uniroot} outputs of the De estimations\cr

\strong{\verb{@info}}\cr
`$ call`` (\link{call}) the original function call
}
\description{
The function computes De value and De value error using the
global standardised growth curve (gSGC) assumption proposed by Li et al.,
2015 for OSL dating of sedimentary quartz.
}
\details{
The error of the De value is determined using a Monte Carlo simulation approach.
Solving of the equation is realised using \link{uniroot}.
Large values for \code{n.MC} will significantly increase the computation time.
}
\section{Function version}{
 0.1.3
}

\examples{
results <- calc_gSGC(data = data.frame(
LnTn =  2.361, LnTn.error = 0.087,
Lr1Tr1 = 2.744, Lr1Tr1.error = 0.091,
Dr1 = 34.4))

get_RLum(results, data.object = "De")

}
\references{
Li, B., Roberts, R.G., Jacobs, Z., Li, S.-H., 2015. Potential of establishing
a 'global standardised growth curve' (gSGC) for optical dating of quartz from sediments.
Quaternary Geochronology 27, 94-104. \doi{10.1016/j.quageo.2015.02.011}
}
\seealso{
\link[=RLum.Results-class]{RLum.Results}, \link{get_RLum}, \link{uniroot}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{datagen}
