% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_Crosstalk.R
\name{calc_MoransI_expt_no_cor}
\alias{calc_MoransI_expt_no_cor}
\title{Calculate Moran's I expected value in case of no auto correlation (H_0)}
\usage{
calc_MoransI_expt_no_cor(object = rep(1, times = 100), n = NULL)
}
\arguments{
\item{object}{\linkS4class{RLum.Results} or \link{numeric} (\emph{with default}) containing
the values of the grains of one or more discs. Should have length 100;
can contain \code{NA} values.}

\item{n}{\link{integer} (\emph{with default}) The number of grain observations as one integer.
Defaults to NULL. If provided, overrules parameter \code{object}. One full disc means "n = 100".}
}
\value{
Numerical value, a bit smaller than 0: the expected Moran's I for the null hypothesis
(= no spatial correlation), according to -1/(n-1) with n the number of used observations.
}
\description{
Perhaps a bit counter-intuitive, the expected value of Moran's I under
the null hypothesis of no spatial correlation is a value slightly smaller than zero. This function
calculates the expected value based on the number of observations or the length of the
observation vector (while taking out \code{NA} values). This can be useful for plotting.
Defaults, when no arguments are provided, to the expected value of one observation disc with
100 valid observations.
}
\details{
Note that the expected value only depends on the number of observed separate grain values.
}
\examples{
calc_MoransI_expt_no_cor()
calc_MoransI_expt_no_cor(n = 20)

}
\author{
Anna-Maartje de Boer, Luc Steinbuch, Wageningen University & Research, 2025
}
