% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_Crosstalk.R
\name{calc_MoransI_pseudo_p}
\alias{calc_MoransI_pseudo_p}
\title{Calculate Moran's I pseudo p}
\usage{
calc_MoransI_pseudo_p(
  object,
  n_moransI = calc_MoransI(object),
  n_perm = 999,
  df_neighbour = get_Neighbour(object = vn_values),
  bo_suppress_warnings = FALSE
)
}
\arguments{
\item{object}{\linkS4class{RLum.Results} or \link{numeric} (\strong{required}): containing a numerical vector of length 100,
representing one discs ("position") in a reader.}

\item{n_moransI}{\link{numeric} (\emph{with default}) The value of Moran's I to be tested against.
Defaults to \code{calc_MoransI(object)}.}

\item{n_perm}{\link{integer} (\emph{with default}) Number of random permutations tested to calculate the fraction
which is the pseudo p. Defaults to 999. Influences the calculation speed, which will have impact in case
of large scale simulation loops.}

\item{df_neighbour}{\link{data.frame} (\emph{with default}) Defaults to
\code{get_Neighbour(object)}.}

\item{bo_suppress_warnings}{\link{logical} (\emph{with default}) Should warnings be printed or suppressed?
Defaults to \code{FALSE}: warnings will be printed.}
}
\value{
Pseudo p-value
}
\description{
Calculate Moran's I pseudo p
}
\references{
de Boer, A-M., Steinbuch, L., Heuvelink, G.B.M., Wallinga, J., 2025.
A novel tool to assess crosstalk in single-grain luminescence detection.
Submitted.
}
\seealso{
https://geodacenter.github.io/workbook/5a_global_auto/lab5a.html#permutation-inference
}
\author{
Anna-Maartje de Boer, Luc Steinbuch, Wageningen University & Research, 2025
}
