% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_OSLLxTxRatio.R
\name{calc_OSLLxTxRatio}
\alias{calc_OSLLxTxRatio}
\title{Calculate \code{Lx/Tx} ratio for CW-OSL curves}
\usage{
calc_OSLLxTxRatio(
  Lx.data,
  Tx.data = NULL,
  signal_integral = NULL,
  background_integral = NULL,
  signal_integral_Tx = NULL,
  background_integral_Tx = NULL,
  integral_input = c("channel", "measurement"),
  background.count.distribution = c("non-poisson", "poisson"),
  use_previousBG = FALSE,
  sigmab = NULL,
  od_rates = NULL,
  sig0 = 0,
  digits = NULL,
  ...
)
}
\arguments{
\item{Lx.data}{\link[=RLum.Data.Curve-class]{RLum.Data.Curve}, \link{data.frame}, \link{list} (\strong{required}):
requires a CW-OSL shine down curve (x = time, y = counts). Data can also be
provided as a list.}

\item{Tx.data}{\link[=RLum.Data.Curve-class]{RLum.Data.Curve} or \link{data.frame} (\emph{optional}):
requires a CW-OSL shine down curve (x = time, y = counts). If no
input is given the \code{Tx.data} will be treated as \code{NA} and no \code{Lx/Tx} ratio
is calculated. When \code{Lx.data} is a list, it must be provided as a list of
the same length.}

\item{signal_integral}{\link{integer} (\strong{required}):
vector of channels for the signal integral. If set to \code{NA} (alternate mode),
no integrals are taken into account and their settings are ignored.}

\item{background_integral}{\link{integer} (\strong{required}):
vector of channels for the background integral. If set to \code{NA}, no
background integral is subtracted; in this case, \code{sigmab.LnLx} (unless
manually set) and the signal-to-noise ratio for \code{Ln/Lx} will be \code{NA}.}

\item{signal_integral_Tx}{\link{integer} (\emph{optional}):
vector of channels for the signal integral for the \code{Tx} curve.
If \code{NULL}, the \code{signal_integral} vector is used.}

\item{background_integral_Tx}{\link{integer} (\emph{optional}):
vector of channels for the background integral for the \code{Tx} curve.
If \code{NULL}, the \code{background_integral} vector is used. If set to \code{NA}, no
background integral for the \code{Tx} curve is subtracted; in this case,
\code{sigmab.TxTx} (unless manually set) and the signal-to-noise ratio for
\code{Tn/Tx} will be \code{NA}.}

\item{integral_input}{\link{character} (\emph{with default}):
input type for \code{signal_integral}, one of \code{"channel"} (default) or
\code{"measurement"}. If set to \code{"measurement"}, the best matching channels
corresponding to the given time/temperature range are selected.}

\item{background.count.distribution}{\link{character} (\emph{with default}):
sets the count distribution assumed for the error calculation.
Possible arguments are \code{"poisson"} or \code{"non-poisson"} (default). See
details for further information. It is ignored if \code{od_rates} is provided.}

\item{use_previousBG}{\link{logical} (\emph{with default}):
If set to \code{TRUE} the background of the \code{Lx}-signal is subtracted also
from the \code{Tx}-signal. Please note that in this case separate
signal integral limits for the \code{Tx}-signal are not allowed and will be reset.}

\item{sigmab}{\link{numeric} (\emph{optional}):
option to set a manual value for the overdispersion (for \code{LnTx} and \code{TnTx}),
used for the \code{Lx/Tx} error calculation. The value should be provided as
absolute squared count values, e.g. \code{sigmab = c(300,300)}. Despite its name
it is an additional variance (above Poisson variance), not a dispersion.
\strong{Note:} If only one value is provided this value is taken for both
(\code{LnTx} and \code{TnTx}) signals.}

\item{od_rates}{\link{numeric} (\emph{optional}):
a vector of three elements: the dark count rate (or background count rate),
the background count overdispersion, and the photon count overdispersion.
If set, an alternative way of computing the overdispersion will be used
(see details). It is ignored (with a warning) if \code{sigmab} is provided.}

\item{sig0}{\link{numeric} (\emph{with default}):
allow adding an extra component of error to the final \code{Lx/Tx} error value
(e.g., instrumental error, see details).}

\item{digits}{\link{integer} (\emph{with default}):
round numbers to the specified digits. If set to \code{NULL} no rounding occurs.}

\item{...}{currently not used.}
}
\value{
Returns an S4 object of type \link[=RLum.Results-class]{RLum.Results}.

Slot \code{data} contains a \link{list} with the following structure:

\strong{@data}

\if{html}{\out{<div class="sourceCode">}}\preformatted{$LxTx.table (data.frame)
.. $ LnLx
.. $ LnLx.BG
.. $ TnTx
.. $ TnTx.BG
.. $ Net_LnLx
.. $ Net_LnLx.Error
.. $ Net_TnTx
.. $ Net_TnTx.Error
.. $ SN_RATIO_LnLx,
.. $ SN_RATIO_TnTx,
.. $ LxTx
.. $ LxTx.Error
$calc.parameters (list)
.. $ sigmab.LnTx
.. $ sigmab.TnTx
.. $ k
.. $ od_rates
}\if{html}{\out{</div>}}

\strong{@info}

\if{html}{\out{<div class="sourceCode">}}\preformatted{$ call (original function call)
}\if{html}{\out{</div>}}
}
\description{
Calculate \code{Lx/Tx} ratios from a given set of CW-OSL curves assuming late light
background subtraction.
}
\details{
The function checks the integrity of the values chosen for the signal and
background integrals; the signal integral limits have to be lower than
the background integral limits. If a \link{vector} is given as input instead
of a \link{data.frame}, an artificial \link{data.frame} is produced.

The error calculation is done according to Galbraith (2002, 2014) or to
Bluszcz et a. (2015) depending on the presence of \code{od_rates} argument.

\strong{Please note:} In cases where the calculation results in \code{NaN} values (for
example due to zero-signal, and therefore a division of 0 by 0), these \code{NaN}
values are replaced by 0.

\strong{\code{background.count.distribution}}

This argument allows selecting the distribution assumption that is used for
the error calculation. According to Galbraith (2002, 2014) the background
counts may be overdispersed (i.e. not follow a Poisson distribution, which
is assumed for the photomultiplier counts). In that case (might be the
normal case) the overdispersion has to be accounted for by estimating
\eqn{\sigma_B^2} (i.e. the overdispersion value). Therefore the relative
standard error is calculated as:
\itemize{
\item \code{poisson}
\deqn{rse(\mu_{S}) \approx \sqrt{(Y_{0} + Y_{1}/k^2) / (Y_{0} - Y_{1}/k)} }
\item \code{non-poisson}
\deqn{rse(\mu_{S}) \approx \sqrt{(Y_{0} + Y_{1}/k^2 + \sigma_B^2(1+1/k)) / (Y_{0} - Y_{1}/k)} }
}

If \code{background_integral = NA}, then in both cases the relative standard
error simplifies to:

\deqn{rse(\mu_{S}) \approx \sqrt{Y_{0}} / Y_{0}}

\strong{\code{sigmab}}

The default value of \code{sigmab} (\eqn{\sigma_B^2}) is calculated assuming the background is
constant and \strong{would not} be applicable when the background varies,
e.g., as observed for the early light subtraction method.

\strong{Note:} When using the early background subtraction method in combination
with the 'non-poisson' distribution argument, the corresponding \code{Lx/Tx} error
may considerably increase due to a high \code{sigmab} value. Please check whether
this is valid for your data set; if necessary, consider providing a custom
value using the \code{sigmab} argument.

\strong{\code{od_rates}}

Setting the \code{od_rates} argument activates the error calculation according
to Bluszcz et al. (2015). The argument expects a 3-element vector:
\itemize{
\item \code{B_DC} is the PMT-specific dark count rate as pulses per second (not per channel)
\item \code{k_DC} is the PMT-specific dispersion excess factor for dark counts
\item \code{k_ph} is the PMT-specific dispersion excess factor for photon counts
}

The dark count rate and the dispersion excess factors have to be known in
advance, and can be obtained by direct experiments performed in laboratory
conditions, essentially the same as the routine measurement conditions
(for examples see: Bluszcz et al. (2015), Carter et al. (2018)).
The photon count overdispersion \eqn{k_{ph}} needs a photon source with a
constant photon emission rate and independent photon emissions (so the
number of photons emitted in a fixed time is a Poisson variable). Note that
\eqn{B_{DC}} must be non-negative, while \eqn{k_{DC}} and \eqn{k_{ph}} must
be positive (for readers with no pulse divider, they should be greater or
equal to 1).

Under the assumption of statistically independent background (\eqn{N_{DC}})
and photon counts (\eqn{N_{ph}}), the total number of counts (say, a certain
integral \eqn{N} of the OSL curve over time \eqn{t}) is:

\deqn{N = N_{ph} + N_{DC}}

and has the following variance:

\deqn{ s^2(N) = k_{ph}^2 N_{ph} + K_{DC}^2 N_{DC}
              = k_{ph}^2 (N - B_{DC} t) + k_{DC}^2 B_{DC} t
              = k_{ph}^2 N + (k_{DC}^2 - k_{ph}^2) B_{DC} t }

The \code{LnLx.Error} and \code{TnTx.Error} are computed as \eqn{\sqrt{s^2(N)}}.

\strong{\code{sig0}}

This argument allows adding an extra component of error to the final \code{Lx/Tx}
error value. The input will be treated as a factor that is multiplied by
the already calculated \code{LxTx} and the result is added according to:

\deqn{se(LxTx) = \sqrt{se(LxTx)^2 + (LxTx * sig0)^2}}

\strong{\code{SN_RATIO_LnLx} and \code{SN_RATIO_TnTx}}

For convenience, the function returns the signal-to-noise ratio (\code{SN_RATIO})
for the \code{LnLx} and the \code{TnTx} curves. This is simply the signal divided
by the background signal counts normalised to the \code{k} value (see below).
}
\note{
The results of this function have been cross-checked with the Analyst
(version 3.24b).

\strong{Caution:} If you are using early light subtraction (EBG), please either provide your
own \code{sigmab} value or use \code{background.count.distribution = "poisson"}.
}
\section{Function version}{
 0.9.8
}

\examples{

##load data
data(ExampleData.LxTxOSLData, envir = environment())

##calculate Lx/Tx ratio
results <- calc_OSLLxTxRatio(
 Lx.data = Lx.data,
 Tx.data = Tx.data,
 signal_integral = 1:2,
 background_integral = 85:100)

##get results object
get_RLum(results)

} 

\section{How to cite}{
Kreutzer, S., Colombo, M., Bluszcz, A., 2026. calc_OSLLxTxRatio(): Calculate Lx/Tx ratio for CW-OSL curves. Function version 0.9.8. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\references{
Duller, G., 2018. Analyst v4.57 - User Manual.
\verb{https://users.aber.ac.uk/ggd}\cr

Galbraith, R.F., 2002. A note on the variance of a background-corrected OSL
count. Ancient TL, 20 (2), 49-51. \doi{10.26034/la.atl.2002.348}

Galbraith, R.F., 2014. A further note on the variance of a
background-corrected OSL count. Ancient TL, 31 (2), 1-3. \doi{10.26034/la.atl.2014.477}

Bluszcz, A., Adamiec, G., Herr, A., 2015. Estimation of equivalent dose and
its uncertainty in the OSL SAR protocol when count numbers do not follow a
Poisson distribution. Radiation Measurements 81, 46-54.
\doi{10.1016/j.radmeas.2015.01.004}

Carter, J., Cresswell, A.J., Kinnaird, T.C., Carmichael, L.A., Murphy, S.,
Sanderson, D.C.W., 2018. Non-Poisson variations in photomultipliers and
implications for luminescence dating. Radiation Measurements 120, 267-273.
\doi{10.1016/j.radmeas.2018.05.010}
}
\seealso{
\link[=RLum.Data.Curve-class]{RLum.Data.Curve}, \link{fit_DoseResponseCurve},
\link{analyse_SAR.CWOSL}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany) \cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
Andrzej Bluszcz, Silesian University of Technology, Gliwice (Poland)\cr
, RLum Developer Team}
\keyword{datagen}
