% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_Dr_De.R
\name{combine_Dr_De}
\alias{combine_Dr_De}
\title{Combine Dose Rate and Equivalent Dose Distribution}
\usage{
combine_Dr_De(
  Dr,
  int_OD,
  De,
  s,
  Age_range,
  alpha = 0.05,
  method_control = list(),
  outlier_analysis_plot = TRUE,
  verbose = TRUE,
  plot = TRUE
)
}
\arguments{
\item{Dr}{\link{numeric} (\strong{required}): a dose rate sample}

\item{int_OD}{\link{numeric} (\strong{required}): the intrinsic overdispersion, typicall
the standard deviation characterizing the DRT distribution}

\item{De}{\link{numeric} (\strong{required}): a equivalent dose sample}

\item{s}{\link{numeric} (\strong{required}): a vector of measurement errors on the equivalent dose}

\item{Age_range}{\link{numeric} (\strong{required}): the age range to be investigated}

\item{alpha}{\link{numeric} (\emph{with default}): the required significance level used
for the outlier detection}

\item{method_control}{\link{list} (\emph{with default}): further parameters passed down
to the jags modelling process}

\item{outlier_analysis_plot}{\link{logical} (\emph{with default}): enables/disables the outlier analysis plot. Note: the outlier analysis will happen with or without plot output}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable terminal feedback}

\item{plot}{\link{logical} (\emph{with default}): enable/disable plot output}
}
\value{
##TODO
}
\description{
A Bayesian statistical analysis of OSL age requiring dose rate sample.
Estimation contains a preliminary step for detecting outliers in the equivalent
dose sample.
}
\details{
##TODO add table for method control and check method controll
}
\section{Function version}{
 0.1.0
}

\examples{
 n <- 1000
 sdt <- 0.3
 Dr <- stats::rlnorm (n, 0, sdt)
 int_OD <-  0.1
 k <- 50
 De <-  50*sample(Dr,k,replace=TRUE)
 s <- stats::rnorm(k, 10, 2)
 combine_Dr_De(Dr,int_OD,De,s,Age_range = c(0,100))
} 

\section{How to cite}{
Philippe, A., Galharret, J., Mercier, N., Kreutzer, S., 2021. combine_Dr_De(): Combine Dose Rate and Equivalent Dose Distribution. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.11.9000-68. https://CRAN.R-project.org/package=Luminescence
}

\references{
##TODO
}
\author{
Anne Philippe, Université de Nantes (France),
Jean-Michel Galharret, Université de Nantes (France),
Norbert Mercier, IRAMAT-CRP2A, Université Bordeaux Montaigne (France),
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team}
\keyword{datagen}
\keyword{distribution}
\keyword{dplot}
