% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.Data-class.R
\name{dim,RLum.Data-method}
\alias{dim,RLum.Data-method}
\title{Retrieve the dimension of an object}
\usage{
\S4method{dim}{RLum.Data}(x)
}
\arguments{
\item{x}{\link[=RLum.Data-class]{RLum.Data} (\strong{required}):
S4 object of class \code{RLum.Data}.}
}
\value{
The dimension of the object's \code{data} slot.
}
\description{
Retrieve the dimension of an object
}
\keyword{utilities}
