% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\name{ExampleData.CobbleData}
\alias{ExampleData.CobbleData}
\title{Example data for calc_CobbleDoseRate()}
\format{
A \code{\link{data.frame}}. Please see \link{calc_CobbleDoseRate} for detailed information
on the structure of the \link{data.frame}.
}
\description{
An example data set for the function \link{calc_CobbleDoseRate} containing layer
specific information for the cobble to be used in the function.
}
\section{Version}{
 0.1.0
}

\examples{
## Load data
data("ExampleData.CobbleData", envir = environment())

}
\keyword{datasets}
