% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\name{ExampleData.MortarData}
\alias{ExampleData.MortarData}
\alias{MortarData}
\title{Example equivalent dose data from mortar samples}
\format{
Two \code{\link{data.frame}}s containing De and De error
}
\source{
Arbitrary measurements.
}
\description{
Arbitrary data to test the function \code{calc_EED_Model}
}
\examples{
## load data
data(ExampleData.MortarData, envir = environment())

##plot data
plot(MortarData)

}
\references{
unpublished data
}
\keyword{datasets}
\keyword{internal}
