% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{All examples}
\description{
These examples are fully documented:

\itemize{
\item \link{ExampleData.CobbleData}
\item \link{ExampleData.DeValues}
\item \link{ExampleData.Fading}
\item \link{ExampleData.portableOSL}
\item \link{ExampleData.RLum.Data.Image}
\item \link{ExampleData.ScaleGammaDose}
\item \link{ExampleData.SurfaceExposure}
\item \link{ExampleData.TR_OSL}
}

The following ones are also available, but are visible only via the
internal documentation, as they correspond to unpublished or synthetic data:

\itemize{
\item \link{ExampleData.Al2O3C}
\item \link{ExampleData.BINfileData}
\item \link{ExampleData.CW_OSL_Curve}
\item \link{ExampleData.FittingLM}
\item \link{ExampleData.LxTxData}
\item \link{ExampleData.LxTxOSLData}
\item \link{ExampleData.MortarData}
\item \link{ExampleData.RF70Curves}
\item \link{ExampleData.RLum.Analysis}
\item \link{ExampleData.XSYG}
}
}
\keyword{datasets}
