% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_Crosstalk.R
\name{get_Neighbour}
\alias{get_Neighbour}
\title{Get neighbour positions}
\usage{
get_Neighbour(object = NULL, bo_restrict_to_8x8 = FALSE)
}
\arguments{
\item{object}{\linkS4class{RLum.Results} or \link{numeric} (\emph{with default}): numerical
vector containing the values of one or more discs, or an \code{RLum.Results}
object containing such values. Can contain \code{NA} values.
If \code{NULL} (default), a complete 10x10 disc is assumed.}

\item{bo_restrict_to_8x8}{\link{logical} (\emph{with default}) Defaults to \code{FALSE}.
If \code{TRUE}, only on-disc borders to grain locations who do not border the
disc are considered; thus only the inner 8x8 grain locations rather than
the full 10x10.}
}
\value{
Dataframe with columns \code{location}, \code{neighbour} and \code{weight}, respectively.
The integers in the first two columns indicate the index of the vector with observations; the
relative weight is standard set to one. Note that the concept of "location" versus
"neighbour" is symmetric.
}
\description{
Determines which locations on one or more
10x10 discs are neighbouring horizontally and/or vertically ("rook"), in
order to calculate Moran's I.
For the general case, we assume symmetry, and all adjacency get
an equal relative weight of one.
In case there are holes in the image, indicated by \code{NA} (no value observed),
the related border or borders are removed.
}
\details{
Will mostly be used as an helper function. If called without any arguments,
will return the neighbouring positions dataframe for one disc (POS) with on every
grain location an observation.
}
\examples{
head(get_Neighbour())


}
\author{
Anna-Maartje de Boer, Luc Steinbuch, Wageningen University & Research, 2025
}
\keyword{crosstalk}
