% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rightAnswer.R
\name{get_rightAnswer}
\alias{get_rightAnswer}
\title{Function to get the right answer}
\usage{
get_rightAnswer(...)
}
\arguments{
\item{...}{you can pass an infinite number of further arguments}
}
\value{
Returns the right answer
}
\description{
This function returns just the right answer
}
\section{Function version}{
 0.1.0
}

\examples{

## you really want to know?
get_rightAnswer()

}
\author{
inspired by R.G.
, RLum Developer Team} 

\section{How to cite}{
NA, NA, , , 2026. get_rightAnswer(): Function to get the right answer. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

