% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R
\name{get_Risoe.BINfileData}
\alias{get_Risoe.BINfileData}
\title{General accessor function for Risoe.BINfileData objects}
\usage{
get_Risoe.BINfileData(object, ...)
}
\arguments{
\item{object}{\linkS4class{Risoe.BINfileData} (\strong{required}):
S4 object of class \code{Risoe.BINfileData}}

\item{...}{further arguments passed to the specific class method}
}
\value{
An object of the same type as the input object provided.
}
\description{
The function provides a generalised access point for specific
\linkS4class{Risoe.BINfileData} objects. Depending on the input object, the
corresponding function will be selected.
Allowed arguments can be found in the documentations of the corresponding
\linkS4class{Risoe.BINfileData} class.
}
\section{Function version}{
 0.1.0
}

\seealso{
\linkS4class{Risoe.BINfileData}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{internal}
\keyword{utilities}
