\name{get_RLum.Analysis}
\alias{get_RLum.Analysis}
\title{get RLum Analysis}

\usage{get_RLum.Analysis(object, record.id, recordType, curveType, RLum.type, 
    get.index, keep.object = FALSE)}
\arguments{
  \item{object}{
}
  \item{record.id}{
}
  \item{recordType}{
}
  \item{curveType}{
}
  \item{RLum.type}{
}
  \item{get.index}{
}
  \item{keep.object}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





