\name{get_RLum.Data.Curve}
\alias{get_RLum.Data.Curve}
\title{get RLum Data Curve}

\usage{get_RLum.Data.Curve(object, info.object)}
\arguments{
  \item{object}{
}
  \item{info.object}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





