\name{get_RLum.Data.Image}
\alias{get_RLum.Data.Image}
\title{get RLum Data Image}

\usage{get_RLum.Data.Image(object, info.object)}
\arguments{
  \item{object}{
}
  \item{info.object}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





