\name{get_RLum.Results}
\alias{get_RLum.Results}
\title{get RLum Results}

\usage{get_RLum.Results(object, data.object)}
\arguments{
  \item{object}{
}
  \item{data.object}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





