\name{get_structure.RLum.Analysis}
\alias{get_structure.RLum.Analysis}
\title{get structure RLum Analysis}

\usage{get_structure.RLum.Analysis(object)}
\arguments{
  \item{object}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





