% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data-class.R, R/RLum.Data.Curve-class.R,
%   R/RLum.Results-class.R
\name{hist}
\alias{hist}
\alias{hist,RLum.Analysis-method}
\alias{hist,RLum.Data-method}
\alias{hist,RLum.Data.Curve-method}
\alias{hist,RLum.Results-method}
\title{Simple histograms of RLum-class objects}
\usage{
\S4method{hist}{RLum.Analysis}(x, ...)

\S4method{hist}{RLum.Data}(x, ...)

\S4method{hist}{RLum.Data.Curve}(x, ...)

\S4method{hist}{RLum.Results}(x, ...)
}
\arguments{
\item{x}{\link[=RLum-class]{RLum} (\strong{required}):
S4 object of class \code{RLum}.}

\item{...}{further arguments passed to \link[graphics:hist]{graphics::hist} (or to the specific
class methods).}
}
\value{
\itemize{
\item \link[=RLum.Analysis-class]{RLum.Analysis}: a list of histograms, one for
each record's data.
}

\itemize{
\item \link[=RLum.Data-class]{RLum.Data}: the histogram of the object's \code{data} slot.
}

\itemize{
\item \link[=RLum.Data.Curve-class]{RLum.Data.Curve}: the histogram of the 2nd column
of the data matrix.
}

\itemize{
\item \link[=RLum.Results-class]{RLum.Results}: the histogram of the data via
\link{plot_Histogram}.
}
}
\description{
Simple histograms of RLum-class objects
}
\section{Function version}{
 0.1.0
}

\seealso{
\link{plot_Histogram}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
}
\keyword{utilities}
