% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data.Curve-class.R,
%   R/RLum.Data.Image-class.R, R/RLum.Data.Spectrum-class.R,
%   R/RLum.Results-class.R, R/Risoe.BINfileData-class.R
\name{length}
\alias{length}
\alias{length,RLum.Analysis-method}
\alias{length,RLum.Data.Curve-method}
\alias{length,RLum.Data.Image-method}
\alias{length,RLum.Data.Spectrum-method}
\alias{length,RLum.Results-method}
\alias{length,Risoe.BINfileData-method}
\title{Length of RLum-class objects}
\usage{
\S4method{length}{RLum.Analysis}(x)

\S4method{length}{RLum.Data.Curve}(x)

\S4method{length}{RLum.Data.Image}(x)

\S4method{length}{RLum.Data.Spectrum}(x)

\S4method{length}{RLum.Results}(x)

\S4method{length}{Risoe.BINfileData}(x)
}
\arguments{
\item{x}{\link[=RLum-class]{RLum} (\strong{required}):
S4 object of class \code{RLum}.}
}
\value{
\itemize{
\item \link[=RLum.Analysis-class]{RLum.Analysis}: the number of records stored in the
object.
}

\itemize{
\item \link[=RLum.Data.Curve-class]{RLum.Data.Curve}: the maximum time/temperature value
of the curve (corresponding to the stimulation length).
}

\itemize{
\item \link[=RLum.Data.Image-class]{RLum.Data.Image}: the number of frames stored in the
object.
}

\itemize{
\item \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}: the maximum time/temperature
value of the spectrum.
}

\itemize{
\item \link[=RLum.Results-class]{RLum.Results}: the number of data elements stored in
the object.
}

\itemize{
\item \link[=Risoe.BINfileData-class]{Risoe.BINfileData}: the number of curves stored in
the object.
}
}
\description{
Length of RLum-class objects
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
}
\keyword{utilities}
