\name{length_RLum.Analysis}
\alias{length_RLum.Analysis}
\title{length RLum Analysis}

\usage{length_RLum.Analysis(object)}
\arguments{
  \item{object}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





