% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data.Curve-class.R, R/RLum.Results-class.R
\name{length_RLum}
\alias{length_RLum}
\alias{length_RLum,RLum.Analysis-method}
\alias{length_RLum,RLum.Data.Curve-method}
\alias{length_RLum,RLum.Results-method}
\title{Length retrieval function for RLum-class objects}
\usage{
length_RLum(object)

\S4method{length_RLum}{RLum.Analysis}(object)

\S4method{length_RLum}{RLum.Data.Curve}(object)

\S4method{length_RLum}{RLum.Results}(object)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
S4 object of class \code{RLum}}
}
\value{
An \link{integer} indicating the length of the object.
}
\description{
The function provides a generalised access point for specific
\link[=RLum-class]{RLum} objects. Depending on the input object, the corresponding
function will be selected.
}
\section{Methods (by class)}{
\itemize{
\item \code{length_RLum(RLum.Analysis)}: Returns the number of records stored in the object.

\item \code{length_RLum(RLum.Data.Curve)}: Returns the length of the curve object, which is the maximum of the value
time/temperature of the curve (corresponding to the stimulation length).

\item \code{length_RLum(RLum.Results)}: Returns the number of stored data elements.

}}
\section{Function version}{
 0.1.0
}

\seealso{
\link[=RLum.Data.Curve-class]{RLum.Data.Curve},
\link[=RLum.Data.Image-class]{RLum.Data.Image},
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum},
\link[=RLum.Analysis-class]{RLum.Analysis},
\link[=RLum.Results-class]{RLum.Results}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2026. length_RLum(): Length retrieval function for RLum-class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{utilities}
