\name{Luminescence-package}
\alias{Luminescence-package}
\alias{Luminescence}
\docType{package}
\title{
\packageTitle{Luminescence}
}
\description{
\packageDescription{Luminescence}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Luminescence}
\packageIndices{Luminescence}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Luminescence}

Maintainer: \packageMaintainer{Luminescence}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
