% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data.Curve-class.R
\name{melt_RLum}
\alias{melt_RLum}
\alias{melt_RLum,list-method}
\alias{melt_RLum,RLum.Analysis-method}
\alias{melt_RLum,RLum.Data.Curve-method}
\title{Melt RLum-class objects into a flat data.frame}
\usage{
melt_RLum(object, ...)

\S4method{melt_RLum}{list}(object, ...)

\S4method{melt_RLum}{RLum.Analysis}(object)

\S4method{melt_RLum}{RLum.Data.Curve}(object)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
S4 object of class \code{RLum}.}

\item{...}{further arguments passed to the specific class method.}
}
\value{
A flat \link{data.frame}.
}
\description{
The function melts \link[=RLum-class]{RLum} objects into a flat data frame
with columns \code{X}, \code{Y}, \code{TYPE} and \code{UID}, which is the "long" data format
expected by other packages such as \code{ggplot2}.
}
\section{Methods (by class)}{
\itemize{
\item \code{melt_RLum(list)}: Returns a \link{data.frame} with the melted data of all \link[=RLum-class]{RLum}
objects in the input list; elements for which melting fails are silently
removed.

\item \code{melt_RLum(RLum.Analysis)}: Melts \link[=RLum.Analysis-class]{RLum.Analysis} objects into a flat data.frame with columns
\code{X}, \code{Y}, \code{TYPE}, \code{UID}, to be used in combination with other packages
such as \code{ggplot2}.

\item \code{melt_RLum(RLum.Data.Curve)}: Melts \link[=RLum.Data.Curve-class]{RLum.Data.Curve} objects into a flat data.frame with columns
\code{X}, \code{Y}, \code{TYPE}, \code{UID}, to be used in combination with other packages
such as \code{ggplot2}.

}}
\section{Function version}{
 0.1.0
}

\examples{

## load example data
data(ExampleData.XSYG, envir = environment())

melt_RLum(OSL.SARMeasurement[[2]][[1]])

}
\seealso{
\link[=RLum.Data.Curve-class]{RLum.Data.Curve}, \link[=RLum.Analysis-class]{RLum.Analysis}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{utilities}
