% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data-class.R, R/Risoe.BINfileData-class.R
\name{add_metadata<-}
\alias{add_metadata<-}
\alias{rename_metadata<-}
\alias{replace_metadata<-}
\alias{add_metadata<-,RLum.Analysis-method}
\alias{rename_metadata<-,RLum.Analysis-method}
\alias{replace_metadata<-,RLum.Analysis-method}
\alias{add_metadata<-,RLum.Data-method}
\alias{rename_metadata<-,RLum.Data-method}
\alias{replace_metadata<-,RLum.Data-method}
\alias{add_metadata<-,Risoe.BINfileData-method}
\alias{rename_metadata<-,Risoe.BINfileData-method}
\alias{replace_metadata<-,Risoe.BINfileData-method}
\title{Safe manipulation of object metadata}
\usage{
add_metadata(object, ...) <- value

rename_metadata(object, ...) <- value

replace_metadata(object, ...) <- value

\S4method{add_metadata}{RLum.Analysis}(object, info_element) <- value

\S4method{rename_metadata}{RLum.Analysis}(object, info_element) <- value

\S4method{replace_metadata}{RLum.Analysis}(object, info_element, subset = NULL) <- value

\S4method{add_metadata}{RLum.Data}(object, info_element) <- value

\S4method{rename_metadata}{RLum.Data}(object, info_element) <- value

\S4method{replace_metadata}{RLum.Data}(object, info_element, subset = NULL, verbose = TRUE) <- value

\S4method{add_metadata}{Risoe.BINfileData}(object, info_element) <- value

\S4method{rename_metadata}{Risoe.BINfileData}(object, info_element) <- value

\S4method{replace_metadata}{Risoe.BINfileData}(object, info_element, subset = NULL) <- value
}
\arguments{
\item{object}{\link[=RLum.Analysis-class]{RLum.Analysis},
\link[=Risoe.BINfileData-class]{Risoe.BINfileData} (\strong{required}):
object of class \code{RLum.Analysis} or \code{Risoe.BINfileData} to manipulate.}

\item{...}{further arguments passed to the specific class method.}

\item{value}{(\strong{required}):
value to be assigned to the selected metadata entry. A \code{NULL} value is
acceptable only for \code{replace_metadata}, in which case the elements named
in \code{info_element} will be removed.}

\item{info_element}{\link{character} (\strong{required}):
name of the metadata entry to manipulate.}

\item{subset}{\link{expression} (\emph{optional}):
logical expression to limit the substitution only to the selected subset
of elements.}

\item{verbose}{\link{logical} (\emph{with default}):
enable/disable output to the terminal.}
}
\description{
Generic functions for manipulation of metadata in
\link[=Risoe.BINfileData-class]{Risoe.BINfileData}, \link[=RLum.Analysis-class]{RLum.Analysis}
and \link[=RLum.Data-class]{RLum.Data} objects.
}
\section{Functions}{
\itemize{
\item \code{add_metadata(RLum.Analysis) <- value}: Adds metadata to \link[=RLum.Analysis-class]{RLum.Analysis} objects.

\item \code{rename_metadata(RLum.Analysis) <- value}: Renames a metadata entry of \link[=RLum.Analysis-class]{RLum.Analysis} objects.

\item \code{replace_metadata(RLum.Analysis) <- value}: Replaces or removes metadata of \link[=RLum.Analysis-class]{RLum.Analysis} objects.

\item \code{add_metadata(RLum.Data) <- value}: Add metadata entries to \link[=RLum.Data-class]{RLum.Data} objects.

\item \code{rename_metadata(RLum.Data) <- value}: Rename a metadata entry of \link[=RLum.Data-class]{RLum.Data} objects.

\item \code{replace_metadata(RLum.Data) <- value}: Replaces or removes metadata of \link[=RLum.Data-class]{RLum.Data} objects.

\item \code{add_metadata(Risoe.BINfileData) <- value}: Adds metadata to \link[=Risoe.BINfileData-class]{Risoe.BINfileData} objects.

\item \code{rename_metadata(Risoe.BINfileData) <- value}: Renames a metadata entry of \link[=Risoe.BINfileData-class]{Risoe.BINfileData} objects.

\item \code{replace_metadata(Risoe.BINfileData) <- value}: Replaces or removes metadata of \link[=Risoe.BINfileData-class]{Risoe.BINfileData} objects.

}}
\section{Function version}{
 0.1.0
}

\examples{

## load example data
data(ExampleData.BINfileData, envir = environment())

## show data
CWOSL.SAR.Data

## add a new field
add_metadata(CWOSL.SAR.Data,
             info_element = "INSTITUTE") <- "Heidelberg University"

## rename a field
rename_metadata(CWOSL.SAR.Data,
                info_element = "INSTITUTE") <- "INSTITUTION"

## replace all LTYPE to RSL
## but only for the first position
replace_metadata(
 object = CWOSL.SAR.Data,
 info_element = "LTYPE",
 subset = (POSITION == 1)) <- "RSL"

## replacing a field with NULL allows to remove that field
replace_metadata(CWOSL.SAR.Data,
                 info_element = "PREVIOUS") <- NULL

## show the modified data
CWOSL.SAR.Data

}
\seealso{
\link[=RLum.Data-class]{RLum.Data}, \link[=RLum.Analysis-class]{RLum.Analysis},
\link[=Risoe.BINfileData-class]{Risoe.BINfileData}
}
\author{
Marco Colombo, Institute of Geography, Heidelberg University (Germany)
}
\keyword{utilities}
