% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data-class.R, R/RLum.Results-class.R,
%   R/Risoe.BINfileData-class.R
\name{names}
\alias{names}
\alias{names,RLum.Analysis-method}
\alias{names,RLum.Data-method}
\alias{names,RLum.Results-method}
\alias{names,Risoe.BINfileData-method}
\title{Names of RLum-class and Risoe-class objects}
\usage{
\S4method{names}{RLum.Analysis}(x)

\S4method{names}{RLum.Data}(x)

\S4method{names}{RLum.Results}(x)

\S4method{names}{Risoe.BINfileData}(x)
}
\arguments{
\item{x}{\link[=RLum-class]{RLum} or \link[=Risoe.BINfileData-class]{Risoe.BINfileData} (\strong{required}):
S4 object.}
}
\value{
\itemize{
\item \link[=RLum.Analysis-class]{RLum.Analysis}: the names of the objects stored in
its records.
}

\itemize{
\item \link[=RLum.Data-class]{RLum.Data}: the names of the object's \code{info} slot.
}

\itemize{
\item \link[=RLum.Results-class]{RLum.Results}: the names of the object's \code{data} slot.
}

\itemize{
\item \link[=Risoe.BINfileData-class]{Risoe.BINfileData}: the \code{LTYPE} field of the curves
stored in the object
}
}
\description{
Names of RLum-class and Risoe-class objects
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
}
\keyword{utilities}
