% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Crosstalk.R
\name{plot_Disc}
\alias{plot_Disc}
\title{Plot a disc with its values}
\usage{
plot_Disc(
  object,
  bo_show_coordinates = FALSE,
  bo_show_location_ids = FALSE,
  bo_show_legend = FALSE,
  bo_show_neighbours = FALSE,
  df_neighbour = get_Neighbour(object = object),
  bo_show_positioning_holes = TRUE,
  str_transform = "sqrt",
  main = "",
  col = "darkolivegreen",
  pch = 17,
  ...
)
}
\arguments{
\item{object}{The values to show, a numerical vector of size 100. ## NOG AANPASSEN}

\item{bo_show_coordinates}{\link{logical} (\emph{with default}): Single Boolean, show coordinates (1..10) in x and in y direction. Defaults to FALSE}

\item{bo_show_location_ids}{\link{logical} (\emph{with default}): Single Boolean, show id with every grain location  (1..100). Defaults to FALSE}

\item{bo_show_legend}{\link{logical} (\emph{with default}): Single Boolean, defaults to FALSE. Show rudimentary legend.}

\item{bo_show_neighbours}{\link{logical}  (\emph{with default}): Single Boolean. Show which neighbour connections are taken into account if
calculating Moran's I. This makes sense when there are NA observations, or when a non-standard neighbour setting is defined.}

\item{df_neighbour}{\link{dataframe} (\emph{with default}): Dataframe as returned by \code{get_Neighbour()}. Is only relevant if
\code{bo_show_neighbours} is TRUE. Defaults to \code{get_Neighbour(object)}.}

\item{str_transform}{\link{string} (\emph{with default}): The observed value of each individual grain is
reflected in the size of a triangle (or other dot-like element). To account for large value differences,
the transformation from value to triangle size can be "lin" (linear), "log" (logarithmic) and "sqrt"
(square root). Defaults to "sqrt", so that the surface is linear to the value. Note that
the log and sqrt transformations can come with an addition to avoid negative values. When the legend
is shown, the actual lower, middle and upper values are printed.}

\item{...}{other arguments to be given to the base R plot function, such as "main", "col" and "pch" (with
default values "", "darkolivegreen" and 17, respectively). Many other base plot arguments are ignored.}

\item{bo_show_locationing_holes}{\link{logical} (\emph{with default}): Single Boolean, show the 3 positioning holes for orientation.
defaults to TRUE}
}
\value{
none
}
\description{
Shows a schematic representation of the physical appearance of one disc (on position in the reader)
and illustrates the measured or calculated values per grain location.
}
\details{
Depending of the available plotting space, some optional elements might have not enough room
to be displayed. As this function is wrapped around the base plot function, one can also choose to add elements
manually.
}
\examples{

 plot_Disc(1:100)

}
\author{
Anna-Maartje de Boer, Luc Steinbuch, Wageningen University & Research, 2025
}
