% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLum.Data.Curve.R
\name{plot_RLum,RLum.Data.Curve-method}
\alias{plot_RLum,RLum.Data.Curve-method}
\title{Plot function for an \code{RLum.Data.Curve} S4 class object}
\usage{
\S4method{plot_RLum}{RLum.Data.Curve}(
  object,
  norm = FALSE,
  smooth = FALSE,
  auto_scale = FALSE,
  interactive = FALSE,
  par.local = TRUE,
  ...
)
}
\arguments{
\item{object}{\link[=RLum.Data.Curve-class]{RLum.Data.Curve} (\strong{required}):
S4 object of class \link[=RLum.Data.Curve-class]{RLum.Data.Curve}.}

\item{norm}{\link{logical} \link{character} (\emph{with default}):
whether curve normalisation should occur (\code{FALSE} by default), or any of
the normalisation methods provided by \link{normalise_RLum}.}

\item{smooth}{\link{logical} (\emph{with default}):
provides automatic curve smoothing based on the internal function \code{.smoothing}.}

\item{auto_scale}{\link{logical} (\emph{with default}):
if activated, auto scales \code{xlim} or \code{ylim} to the extent of the other. If
both are set, no auto-scaling occurs.}

\item{interactive}{\link{logical} (\emph{with default}):
enables/disables interactive plotting mode using \link[plotly:plot_ly]{plotly::plot_ly}.}

\item{par.local}{\link{logical} (\emph{with default}):
whether local graphical parameters should be used for plotting. If \code{TRUE}
(default), the plot is shown in one column and one row; otherwise, global
parameters set via \code{par()} are inherited.}

\item{...}{further arguments and graphical parameters that will be passed
to \link[graphics:plot.default]{graphics::plot.default} and \link[graphics:par]{graphics::par}.}
}
\value{
Returns a plot.
}
\description{
The function provides a standardised plot output for curve data of an
\link[=RLum.Data.Curve-class]{RLum.Data.Curve} S4-class object. Only single curve
data can be plotted with this function.
}
\note{
Not all arguments of \link{plot} will be passed!
}
\section{Function version}{
 0.5.0
}

\examples{
## load example data
data(ExampleData.CW_OSL_Curve, envir = environment())

## transform data.frame to RLum.Data.Curve object and plot it
temp <- as(ExampleData.CW_OSL_Curve, "RLum.Data.Curve")
plot_RLum(temp)

}
\seealso{
\link{plot}, \link{plot_RLum}, \link{normalise_RLum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{aplot}
