% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLum.Data.Image.R
\name{plot_RLum,RLum.Data.Image-method}
\alias{plot_RLum,RLum.Data.Image-method}
\title{Plot function for an \code{RLum.Data.Image} S4 class object}
\usage{
\S4method{plot_RLum}{RLum.Data.Image}(
  object,
  frames = NULL,
  plot.type = c("plot.raster", "contour"),
  par.local = TRUE,
  ...
)
}
\arguments{
\item{object}{\link[=RLum.Data.Image-class]{RLum.Data.Image} (\strong{required}):
S4 object of class \link[=RLum.Data.Image-class]{RLum.Data.Image}.}

\item{frames}{\link{numeric} (\emph{optional}):
indices of frames to plot. If \code{NULL} (default) all frames are plotted.}

\item{plot.type}{\link{character} (\emph{with default}):
one of \code{plot.raster} (default) or \code{contour}.}

\item{par.local}{\link{logical} (\emph{with default}):
whether local graphical parameters should be used for plotting. If \code{TRUE}
(default), the plot is shown in one column and one row; otherwise, global
parameters set via \code{par()} are inherited.}

\item{...}{further arguments and graphical parameters that will be passed
to the specific plot functions. Standard supported parameters are \code{xlim},
\code{ylim}, \code{zlim}, \code{xlab}, \code{ylab}, \code{main}, \code{mtext}, \code{legend} (\code{TRUE} or \code{FALSE}),
\code{col}, \code{cex}, \code{axes} (\code{TRUE} or \code{FALSE}), \code{zlim_image} (adjust the z-scale
over different images), \code{stretch}, \code{digits}, scientific (\code{TRUE} or \code{FALSE}).}
}
\value{
Returns a plot.
}
\description{
The function provides very basic plot functionality for image data of an
\link[=RLum.Data.Image-class]{RLum.Data.Image} object. For more sophisticated
plotting, it is recommended to use more powerful packages for image processing.
}
\note{
The axes limits (\code{xlim}, \code{ylim}, \code{zlim}) work directly on the object,
so that regardless of the chosen limits, the image parameters can be adjusted
for best visibility. However, in particular for z-scale limits, this is not
always desirable; use \code{zlim_image} to maintain a particular value range over
a series of images.
}
\section{Supported plot types}{


\strong{\code{plot.type = "plot.raster"}}

Uses the standard plot function of R \link[graphics:image]{graphics::image}. If wanted, the image
is enhanced using the argument \code{stretch}. Possible values are \code{hist}, \code{lin}, and
\code{NULL}. The latter does nothing. The argument \code{useRaster = TRUE} is used by default, but
can be set to \code{FALSE}.

\strong{\code{plot.type = "contour"}}

This uses the function \link[graphics:contour]{graphics::contour}
}

\section{Function version}{
 0.3.0
}

\examples{
## load example data
data(ExampleData.RLum.Data.Image, envir = environment())

## plot data
plot_RLum(ExampleData.RLum.Data.Image)

}
\seealso{
\link[=RLum.Data.Image-class]{RLum.Data.Image}, \link{plot}, \link{plot_RLum},
\link[graphics:image]{graphics::image}, \link[graphics:contour]{graphics::contour}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{aplot}
