% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLum.Results.R
\name{plot_RLum,RLum.Results-method}
\alias{plot_RLum,RLum.Results-method}
\title{Plot function for an \code{RLum.Results} S4 class object}
\usage{
\S4method{plot_RLum}{RLum.Results}(object, single = TRUE, ...)
}
\arguments{
\item{object}{\link[=RLum.Results-class]{RLum.Results} (\strong{required}):
S4 object of class \link[=RLum.Results-class]{RLum.Results}.}

\item{single}{\link{logical} (\emph{with default}):
single plot output (\code{TRUE/FALSE}) to allow for plotting the results in as
few plot windows as possible.}

\item{...}{further arguments and graphical parameters will be passed to
the \code{plot} function.}
}
\value{
Returns multiple plots.
}
\description{
The function provides a standardised plot output for data of an \code{RLum.Results}
S4 class object, redirecting to specialised functions if necessary.

The function may produce multiple plots, in which case a file output is
recommended (e.g., \link{pdf}).
}
\note{
Not all arguments available for \link{plot} will be passed!
Only plotting of \code{RLum.Results} objects are supported.
}
\section{Function version}{
 0.3.0
}

\examples{
## load example data
data(ExampleData.DeValues, envir = environment())

## apply the un-logged minimum age model
mam <- calc_MinDose(data = ExampleData.DeValues$CA1, sigmab = 0.2,
                    log = FALSE, plot = FALSE)
plot_RLum(mam)

## estimate the number of grains on an aliquot
grains <- calc_AliquotSize(grain.size = c(100, 150), sample.diameter = 1,
                           plot = FALSE, MC.iter = 100)
plot_RLum(grains)

}
\seealso{
\link{plot}, \link{plot_RLum}
}
\author{
Christoph Burow, University of Cologne (Germany) \cr
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{aplot}
