% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLum.Analysis.R
\name{plot_RLum.Analysis}
\alias{plot_RLum.Analysis}
\title{Plot function for an \code{RLum.Analysis} S4 class object}
\usage{
plot_RLum.Analysis(
  object,
  nrows = NULL,
  ncols = NULL,
  abline = NULL,
  combine = FALSE,
  records_max = NULL,
  curve.transformation = "None",
  plot_singlePanels = FALSE,
  ...,
  subset = NULL
)
}
\arguments{
\item{object}{\link[=RLum.Analysis-class]{RLum.Analysis} (\strong{required}):
S4 object of class \link[=RLum.Analysis-class]{RLum.Analysis}.}

\item{nrows}{\link{integer} (\emph{optional}):
number of rows in the plot output. If set to \code{NULL} the function tries to
find a reasonable value.}

\item{ncols}{\link{integer} (\emph{optional}):
number of columns in the plot output. If set to \code{NULL} the function tries to
find a reasonable value.}

\item{abline}{\link{list} (\emph{optional}):
list of straight lines to add to the plot via the \link{abline} function,
e.g., \code{list(v = c(10, 100))} adds two vertical lines add 10 and 100 to all
plots; in contrast \code{list(v = 10, v = 100)} adds a vertical at 10 to
the first plot and a vertical line at 100 to the second.}

\item{combine}{\link{logical} (\emph{with default}):
allows to combine all \link[=RLum.Data.Curve-class]{RLum.Data.Curve} objects in one single plot.}

\item{records_max}{\link{integer} (\emph{optional}):
limits number of records shown when \code{combine = TRUE} is used. The first and
last curves are always shown, the other curves shown are distributed evenly;
this may result in fewer curves plotted than specified. This argument must
be at least 2 to have an effect.}

\item{curve.transformation}{\link{character} (\emph{with default}):
allows transforming CW-OSL and CW-IRSL curves to pseudo-LM curves via
transformation functions. Allowed values are: \code{CW2pLM}, \code{CW2pLMi},
\code{CW2pHMi} and \code{CW2pPMi}, see details. If set to \code{None} (default), no
transformation is applied.}

\item{plot_singlePanels}{\link{logical} (\emph{with default}):
global par settings are considered, normally this should end in one plot per page.}

\item{...}{further arguments and graphical parameters will be passed to
the \code{plot} function.
Supported arguments: \code{main}, \code{mtext}, \code{log}, \code{lwd}, \code{lty} \code{type}, \code{pch},
\code{col}, \code{cex}, \code{norm} (see \link{plot_RLum.Data.Curve}), \code{sub},
\code{xlim}, \code{ylim}, \code{xlab}, \code{ylab}; for \code{combine = TRUE} also: \code{legend},
\code{legend.text}, \code{legend.pos} (typical values plus 'outside'), \code{legend.col},
\code{smooth}.

All arguments can be provided as \code{vector} or \code{list} to gain in full control
of all plot settings.}

\item{subset}{\link{list} (\emph{optional}):
named list of elements to plot, to be passed directly to \link{get_RLum}
(e.g., \code{subset = list(curveType = "measured")}).}
}
\value{
Returns multiple plots.
}
\description{
The function provides a standardised plot output for curve data of an
\link[=RLum.Analysis-class]{RLum.Analysis} object. As the function produces
multiple plots, file output is recommended (e.g., \link{pdf}).
}
\note{
Not all arguments available for \link{plot} will be passed and they partly do not behave in the
way you might expect them to work. This function was designed to serve as an overview
plot, if you want to have more control, extract the objects and plot them individually.
}
\section{The \code{curve.transformation} argument}{


This argument allows transforming continuous wave (CW) curves to pseudo
(linear) modulated curves. For the transformation, the functions of the
package are used. Currently, it is not possible to pass further arguments to
the transformation functions. The argument works only for \code{ltype}
\code{OSL} and \code{IRSL}.

\strong{Note:} The curve transformation within this function works roughly,
i.e. every IRSL or OSL curve is transformed, without considering whether it
is measured with the PMT or not. However, for a fast look it might be
helpful.
}

\section{Function version}{
 0.3.17
}

\examples{
## load example data
data(ExampleData.BINfileData, envir = environment())

## convert values for position 1
temp <- Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos = 1)

## combine TL curves in one plot
plot_RLum.Analysis(
  temp,
  subset = list(recordType = "TL"),
  combine = TRUE,
  norm = TRUE,
  smooth = TRUE,
  abline = list(v = 110)
 )

}
\seealso{
\link{plot}, \link{plot_RLum}, \link{plot_RLum.Data.Curve}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{aplot}
