% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLum.Data.Curve.R
\name{plot_RLum.Data.Curve}
\alias{plot_RLum.Data.Curve}
\title{Plot function for an RLum.Data.Curve S4 class object}
\usage{
plot_RLum.Data.Curve(
  object,
  norm = FALSE,
  smooth = FALSE,
  auto_scale = FALSE,
  interactive = FALSE,
  par.local = TRUE,
  ...
)
}
\arguments{
\item{object}{\link[=RLum.Data.Curve-class]{RLum.Data.Curve} (\strong{required}):
S4 object of class \link[=RLum.Data.Curve-class]{RLum.Data.Curve}.}

\item{norm}{\link{logical} \link{character} (\emph{with default}):
whether curve normalisation should occur (\code{FALSE} by default), or any of
the normalisation methods provided by \link{normalise_RLum}.}

\item{smooth}{\link{logical} (\emph{with default}):
provides automatic curve smoothing based on the internal function \code{.smoothing}}

\item{auto_scale}{\link{logical} (\emph{with default}): if activated, auto scales \code{xlim} or \code{ylim}
to the extent of the other. If both are set, the auto-scaling is skipped.}

\item{interactive}{\link{logical} (\emph{with default}): enables/disables interactive
plotting mode using \link[plotly:plot_ly]{plotly::plot_ly}}

\item{par.local}{\link{logical} (\emph{with default}):
whether local graphical parameters should be used for plotting. If \code{TRUE}
(default), the plot is shown in one column and one row; otherwise, global
parameters set via \code{par()} are inherited.}

\item{...}{further arguments and graphical parameters that will be passed
to \link[graphics:plot.default]{graphics::plot.default} and \link[graphics:par]{graphics::par}}
}
\value{
Returns a plot.
}
\description{
The function provides a standardised plot output for curve data of an
\link[=RLum.Data.Curve-class]{RLum.Data.Curve} S4-class object.
}
\details{
Only single curve data can be plotted with this function. Arguments
according to \link{plot}.
}
\note{
Not all arguments of \link{plot} will be passed!
}
\section{Function version}{
 0.4.1
}

\examples{

##plot curve data

#load Example data
data(ExampleData.CW_OSL_Curve, envir = environment())

#transform data.frame to RLum.Data.Curve object
temp <- as(ExampleData.CW_OSL_Curve, "RLum.Data.Curve")

#plot RLum.Data.Curve object
plot_RLum.Data.Curve(temp)

}
\seealso{
\link{plot}, \link{plot_RLum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{aplot}
