% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R
\name{plot_RLum}
\alias{plot_RLum}
\alias{plot_RLum,list-method}
\alias{plot_RLum,matrix-method}
\title{Plotting of RLum-class objects}
\usage{
plot_RLum(object, ...)

\S4method{plot_RLum}{list}(object, ...)

\S4method{plot_RLum}{matrix}(object, ...)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
object of class \link[=RLum-class]{RLum} or a list of such objects. If a
list is provided, the function tries to plot every object in the list
according to its \code{RLum} class, after removing non-RLum objects.}

\item{...}{further arguments and graphical parameters to pass to the
specific plot functions. The only arguments that are supported directly are
\code{main} (plot title) and \code{mtext} (plot subtitle), which can be provided as a
list and the arguments in the list will be dispatched to the plots if
\code{object} is of type \code{list} as well.}
}
\value{
Produces a plot depending on the input object.
}
\description{
The function plots \link[=RLum-class]{RLum} objects by selecting the
function that corresponds to the class of the input object; each of these
functions comes with its own set of arguments, which are documented along
with the respective function:
\itemize{
\item For \link[=RLum.Analysis-class]{RLum.Analysis} objects:
\link{plot_RLum,RLum.Analysis-method}
\item For \link[=RLum.Data.Curve-class]{RLum.Data.Curve} objects:
\link{plot_RLum,RLum.Data.Curve-method}
\item For \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects:
\link{plot_RLum,RLum.Data.Spectrum-method}
\item For \link[=RLum.Data.Image-class]{RLum.Data.Image} objects:
\link{plot_RLum,RLum.Data.Image-method}
\item For \link[=RLum.Results-class]{RLum.Results} objects:
\link{plot_RLum,RLum.Results-method}
}
}
\section{Functions}{
\itemize{
\item \code{plot_RLum(list)}: Plot method for a list of \link[=RLum-class]{RLum} objects: after
flattening sublists and removing non-RLum objects, each object is plotted
according to its class. \code{main} and \code{mtext} can be provided as a list to
specify different titles/subtitles for each object.

\item \code{plot_RLum(matrix)}: A \code{matrix} containing count values of the spectrum. Note that row and column
names are set automatically if not provided.

}}
\section{Function version}{
 0.7.0
}

\examples{
## load example data
data(ExampleData.CW_OSL_Curve, envir = environment())

## transform data.frame to RLum.Data.Curve object and plot it
temp <- as(ExampleData.CW_OSL_Curve, "RLum.Data.Curve")
plot_RLum(temp)

}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
}
\keyword{dplot}
