% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_VolcanoPlot.R
\name{plot_VolcanoPlot}
\alias{plot_VolcanoPlot}
\title{Create a volcano plot}
\usage{
plot_VolcanoPlot(data, boxplot = TRUE, rug = TRUE, summary = NULL,
  summary.pos = "sub", na.rm = FALSE, ...)
}
\arguments{
\item{data}{\code{\link{numeric}} or \code{\linkS4class{RLum.Results}}
object (required): input data for plotting. Alternatively a \code{\link{data.frame}} or
a \code{\link{matrix}} can be provided, but only the first column will be considered by the
function}

\item{boxplot}{\code{\link{logical}} (with default): enable or disable boxplot}

\item{rug}{\code{\link{logical}} (with default): enable or disable rug}

\item{summary}{\code{\link{character}} (optional): add statistic measures of
centrality and dispersion to the plot. Can be one or more of several
keywords. See details for available keywords.}

\item{summary.pos}{\code{\link{numeric}} or \code{\link{character}} (with
default): optional position keywords (cf., \code{\link{legend}})
for the statistical summary. Alternatively, the keyword \code{"sub"} may be
specified to place the summary below the plot header. However, this latter
option in only possible if \code{mtext} is not used.}

\item{na.rm}{\code{\link{logical}} (with default): exclude NA values
from the data set prior to any further operations.}

\item{\dots}{further arguments and graphical parameters passed to
\code{\link{plot.default}}, \code{\link[stats]{density}} and \code{\link{boxplot}}. See details for
further information}
}
\description{
Draws a volcano plot in combination with a boxplot in its middle. The shape of the volcano
is constructed using a mirrored density curve. This plot is especially designed for cases
where the individual errors are zero or to small to be visualised.
}
\details{
The function is passing several arguments to the function \code{\link{plot}},
\code{\link[stats]{density}}, \code{\link[graphics]{boxplot}}: Supported arguments are: \code{xlim}, \code{main}, \code{xlab},
\code{ylab}, \code{col.volcano}, \code{col.boxplot}, \code{mtext}, \code{cex}, \code{mtext}
}
\note{
-
}
\section{Function version}{
 0.1.0 (2015-09-13 18:55:08)
}
\examples{
## read example data set
data(ExampleData.DeValues, envir = environment())
ExampleData.DeValues <- Second2Gray(ExampleData.DeValues$BT998, c(0.0438,0.0019))

## create plot straightforward
plot_VolcanoPlot(data = ExampleData.DeValues)
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\seealso{
\code{\link[stats]{density}}, \code{\link{plot}}, \code{\link{boxplot}}, \code{\link{rug}}
}

