% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSL2Risoe.BINfileData.R
\name{PSL2Risoe.BINfileData}
\alias{PSL2Risoe.BINfileData}
\title{Convert portable OSL data to an Risoe.BINfileData object}
\usage{
PSL2Risoe.BINfileData(object, ...)
}
\arguments{
\item{object}{[RLum.Analysis-class] (\bold{required}):
\code{RLum.Analysis} object produced by [read_PSL2R]}

\item{...}{currently not used.}
}
\value{
Returns an S4 [Risoe.BINfileData-class] object that can
be used to write a BIN file using [write_R2BIN].
}
\description{
Converts an \code{RLum.Analysis} object produced by the function \code{read_PSL2R()} to
an \code{Risoe.BINfileData} object \bold{(BETA)}.
}
\details{
This function converts an [RLum.Analysis-class] object that was produced
by the [read_PSL2R] function to an [Risoe.BINfileData-class].
The \code{Risoe.BINfileData} can be used to write a Risoe BIN file via
[write_R2BIN].
}
\section{Function version}{
 0.0.1
}

\examples{

# (1) load and plot example data set
data("ExampleData.portableOSL", envir = environment())
plot_RLum(ExampleData.portableOSL)

# (2) merge all RLum.Analysis objects into one
merged <- merge_RLum(ExampleData.portableOSL)
merged

# (3) convert to RisoeBINfile object
bin <- PSL2Risoe.BINfileData(merged)
bin

# (4) write Risoe BIN file
\dontrun{
write_R2BIN(bin, "~/portableOSL.binx")
}

}
\seealso{
[RLum.Analysis-class], [RLum.Data.Curve-class],
[Risoe.BINfileData-class]
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\keyword{IO}
