% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Daybreak2R.R
\name{read_Daybreak2R}
\alias{read_Daybreak2R}
\title{Import measurement data produced by a Daybreak TL/OSL reader into R}
\usage{
read_Daybreak2R(file, raw = FALSE, verbose = TRUE, txtProgressBar = TRUE, ...)
}
\arguments{
\item{file}{\link{character} or \link{list} (\strong{required}):
name of one or multiple \code{.DAT} or \code{.txt} files (URLs are supported); it can
be the path to a directory, in which case the function tries to detect and
import all files with \code{.txt} extension.}

\item{raw}{\link{logical} (\emph{with default}):
if the input is a DAT-file (binary) a \link[data.table:data.table]{data.table::data.table} instead of
the \link[=RLum.Analysis-class]{RLum.Analysis} object can be returned for debugging purposes.}

\item{verbose}{\link{logical} (\emph{with default}):
enable/disable output to the terminal.}

\item{txtProgressBar}{\link{logical} (\emph{with default}):
enable/disable \link{txtProgressBar}.}

\item{...}{not in use, for compatibility reasons only}
}
\value{
A list of \link[=RLum.Analysis-class]{RLum.Analysis} objects (each per position)
is returned.
}
\description{
Import a TXT-file (ASCII file) or a DAT-file (binary file) produced by a
Daybreak reader into R. The import of the DAT-files is limited to the file
format described for the software TLAPLLIC v.3.2 used for Daybreak models
1100 and 1150, and for FLConsole (any release) for Daybreak models 11xx and
2200.
}
\note{
\strong{\verb{[BETA VERSION]}}
This function still needs to be tested properly. In particular
the function has underwent only very rough tests using a few files.
}
\section{Function version}{
 0.4.1
}

\examples{

\dontrun{
file <- system.file("extdata/Daybreak_TestFile.txt", package = "Luminescence")
temp <- read_Daybreak2R(file)
}

}
\seealso{
\link[=RLum.Analysis-class]{RLum.Analysis}, \link[=RLum.Data.Curve-class]{RLum.Data.Curve}, \link[data.table:data.table]{data.table::data.table}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Andrzej Bluszcz, Silesian University of Technology, Gliwice (Poland)\cr
Antoine Zink, C2RMF, Palais du Louvre, Paris (France)

The ASCII-file import is based on a suggestion by William Amidon and
Andrew Louis Gorin.
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., Bluszcz, A., Zink, A., 2026. read_Daybreak2R(): Import measurement data produced by a Daybreak TL/OSL reader into R. Function version 0.4.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{IO}
