% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HeliosOSL2R.R
\name{read_HeliosOSL2R}
\alias{read_HeliosOSL2R}
\title{Import Luminescence Data from Helios Luminescence Reader}
\usage{
read_HeliosOSL2R(file, verbose = TRUE, ...)
}
\arguments{
\item{file}{\link{character}, \link{list} (\strong{required}):
name of one or multiple \code{.osl} files (URLs are supported); it can be the
path to a directory, in which case the function tries to detect and import
all \code{.osl} files found in the directory.}

\item{verbose}{\link{logical} (\emph{with default}):
enable/disable output to the terminal.}

\item{...}{not in use, for compatibility reasons only}
}
\value{
A \link[=RLum.Analysis-class]{RLum.Analysis} object. Results are returned as a
list when multiple files are processed or \code{file} is a list.
}
\description{
Import of files with \code{.osl} extension produced by the zero rad Helios
luminescence reader and conversion to \link[=RLum.Analysis-class]{RLum.Analysis}
objects.
}
\note{
Thanks to Krzysztof Maternicki for providing example data.
}
\section{Function version}{
 0.1.1
}

\examples{
file <- system.file("extdata/HeliosOSL_Example.osl", package = "Luminescence")
read_HeliosOSL2R(file)

}
\seealso{
\link[=RLum.Data.Curve-class]{RLum.Data.Curve}, \link[=RLum.Analysis-class]{RLum.Analysis}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2026. read_HeliosOSL2R(): Import Luminescence Data from Helios Luminescence Reader. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{IO}
