% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TIFF2R.R
\name{read_TIFF2R}
\alias{read_TIFF2R}
\title{Import TIFF Image Data into R}
\usage{
read_TIFF2R(file, merge2stack = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{file}{\link{character} (\strong{required}):
name of the TIFF file to read (URLs are supported).}

\item{merge2stack}{\link{logical} (\emph{with default}): if \code{file} is a \link{list} it merges
the individual images into one image stack. Please note that the smallest image
dimension determines pixel dimension of the output stack.}

\item{verbose}{\link{logical} (\emph{with default}):
enable/disable output to the terminal.}

\item{...}{not in use, for compatibility reasons only}
}
\value{
Returns an \link[=RLum.Data.Image-class]{RLum.Data.Image} object. Results are
returned as a list when multiple files are processed or \code{file} is a list.
}
\description{
Simple wrapper around \link[tiff:readTIFF]{tiff::readTIFF} to import TIFF images and TIFF image
stacks to be further processed within the 'Luminescence' package.
}
\section{Function version}{
 0.2.1
}

\examples{

## use system file
file <- system.file("extdata", "TIFFfile.tif", package = "Luminescence")

## import image
image <- read_TIFF2R(file)

}
\seealso{
\link[tiff:readTIFF]{tiff::readTIFF}, \link[=RLum.Data.Image-class]{RLum.Data.Image}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2026. read_TIFF2R(): Import TIFF Image Data into R. Function version 0.2.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{IO}
