% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{replace_metadata<-}
\alias{replace_metadata<-}
\title{Safe replacement of object metadata}
\usage{
replace_metadata(object, ...) <- value
}
\arguments{
\item{object}{(\strong{required}) object to manipulate}

\item{...}{further arguments passed to the function}

\item{value}{the value assigned}
}
\description{
Generic function for replacement of object metadata.
}
\author{
Marco Colombo, Institute of Geography, Heidelberg University (Germany)
}
\keyword{utilities}
