% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.Data-class.R
\docType{class}
\name{RLum.Data-class}
\alias{RLum.Data-class}
\title{Class \code{"RLum.Data"}}
\description{
Generalized virtual data class for luminescence data.
}
\section{Objects from the Class}{

A virtual class: no objects can be created from it.
}

\section{Class version}{
 0.2.1
}

\examples{

showClass("RLum.Data")

}
\seealso{
\linkS4class{RLum}, \linkS4class{RLum.Data.Curve},
\linkS4class{RLum.Data.Spectrum}, \linkS4class{RLum.Data.Image}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{classes}
\keyword{internal}
