% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R
\name{set_Risoe.BINfileData}
\alias{set_Risoe.BINfileData}
\title{General setter function for Risoe.BINfileData objects}
\usage{
set_Risoe.BINfileData(
  METADATA = data.frame(),
  DATA = list(),
  .RESERVED = list()
)
}
\arguments{
\item{METADATA}{\link{data.frame} (\emph{with default}):
metadata information for each curve.}

\item{DATA}{\link{list} (\emph{with default}):
numeric vectors with the count data, one per row of \code{METADATA}.}

\item{.RESERVED}{\link{list} (\emph{with default}):
list of undocumented raw values for internal use only.}
}
\value{
A \link[=Risoe.BINfileData-class]{Risoe.BINfileData} object.
}
\description{
Constructs an empty \link[=Risoe.BINfileData-class]{Risoe.BINfileData} object, which can
then be filled with the supplied metadata and data. Objects of this class are
normally produced as output of \link{read_BIN2R}; this construction
method is intended for internal usage only.
}
\section{Function version}{
 0.1.0
}

\examples{

## create an empty Risoe.BINfileData object
set_Risoe.BINfileData()

}
\seealso{
\link[=Risoe.BINfileData-class]{Risoe.BINfileData}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{utilities}
