\name{set_RLum.Analysis}
\alias{set_RLum.Analysis}
\title{set RLum Analysis}

\usage{set_RLum.Analysis(records, protocol)}
\arguments{
  \item{records}{
}
  \item{protocol}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





