\name{set_RLum.Data.Curve}
\alias{set_RLum.Data.Curve}
\title{set RLum Data Curve}

\usage{set_RLum.Data.Curve(recordType, curveType, data, info)}
\arguments{
  \item{recordType}{
}
  \item{curveType}{
}
  \item{data}{
}
  \item{info}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





