\name{set_RLum.Results}
\alias{set_RLum.Results}
\title{set RLum Results}

\usage{set_RLum.Results(originator, data)}
\arguments{
  \item{originator}{
}
  \item{data}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





