% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data.Curve-class.R,
%   R/RLum.Data.Image-class.R, R/RLum.Data.Spectrum-class.R,
%   R/RLum.Results-class.R, R/Risoe.BINfileData-class.R
\name{show}
\alias{show}
\alias{show,RLum.Analysis-method}
\alias{show,RLum.Data.Curve-method}
\alias{show,RLum.Data.Image-method}
\alias{show,RLum.Data.Spectrum-method}
\alias{show,RLum.Results-method}
\alias{show,Risoe.BINfileData-method}
\title{Show the structure of RLum-class and Risoe.BINfileData-class objects}
\usage{
\S4method{show}{RLum.Analysis}(object)

\S4method{show}{RLum.Data.Curve}(object)

\S4method{show}{RLum.Data.Image}(object)

\S4method{show}{RLum.Data.Spectrum}(object)

\S4method{show}{RLum.Results}(object)

\S4method{show}{Risoe.BINfileData}(object)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum}, \link[=Risoe.BINfileData-class]{Risoe.BINfileData} (\strong{required}):
object of class \code{RLum} or \code{Risoe.BINfileData}.}
}
\description{
The function shows the structure of \link[=RLum-class]{RLum} and
\link[=Risoe.BINfileData-class]{Risoe.BINfileData} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{show(RLum.Analysis)}: Show the structure of \code{RLum.Analysis} objects.

\item \code{show(RLum.Data.Curve)}: Show the structure of \code{RLum.Data.Curve} objects.

\item \code{show(RLum.Data.Image)}: Show the structure of \code{RLum.Data.Image} objects.

\item \code{show(RLum.Data.Spectrum)}: Show the structure of \code{RLum.Data.Spectrum} objects.

\item \code{show(RLum.Results)}: Show the structure of \code{RLum.Results} objects.

\item \code{show(Risoe.BINfileData)}: Show the structure of \code{Risoe.BINfileData} objects.

}}
