% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sTeve}
\alias{sTeve}
\title{sTeve - sophisticated tool for efficient data validation and evaluation}
\usage{
sTeve(n_frames = 10, t_animation = 2, n.tree = 7, type = NULL)
}
\arguments{
\item{n_frames}{\link{integer} (\emph{with default}):
n frames}

\item{t_animation}{\link{integer} (\emph{with default}):
t animation}

\item{n.tree}{\link{integer} (\emph{with default}):
how many trees do you want to cut?}

\item{type}{\link{integer} (\emph{optional}):
Make a decision: 1, 2 or 3}
}
\value{
Validates your data.
}
\description{
This function provides a sophisticated routine for comprehensive
luminescence dating data analysis.
}
\details{
This amazing sophisticated function validates your data seriously.
}
\note{
This function should not be taken too seriously.
}
\examples{

##no example available

}
\seealso{
\link{plot_KDE}
}
\author{
R Luminescence Team, 2012-2046
, RLum Developer Team} 

\section{How to cite}{
NA, NA, , , 2026. sTeve(): sTeve - sophisticated tool for efficient data validation and evaluation. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{manip}
