% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R
\name{structure_RLum}
\alias{structure_RLum}
\alias{structure_RLum,list-method}
\alias{structure_RLum,RLum.Analysis-method}
\title{General structure function for RLum-class objects}
\usage{
structure_RLum(object, ...)

\S4method{structure_RLum}{list}(object, ...)

\S4method{structure_RLum}{RLum.Analysis}(object, fullExtent = FALSE)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
S4 object of class \code{RLum}.}

\item{...}{further arguments passed to the specific class method.}

\item{fullExtent}{\link{logical} (\emph{with default}):
extends the returned \code{data.frame} to its full extent, i.e. all info elements
are part of the return as well. The default value is \code{FALSE} as the data
frame might become rather big.}
}
\value{
A \link{data.frame} with the structure of the object, or a \link{list} of such
\link{data.frame}s if a \link{list} was provided.
}
\description{
The function summarises the structure of \link[=RLum-class]{RLum} objects
in a \link{data.frame} with one row per record and one column per slot and
\code{info} element. This is a convenient way of comparing the records of an
object, or of several objects, without having to inspect each of them
individually.
}
\section{Methods (by class)}{
\itemize{
\item \code{structure_RLum(list)}: Returns a \link{list} containing the structure of each \link[=RLum-class]{RLum}
object in the input list; elements that are not of class
\link[=RLum-class]{RLum} are returned unchanged.

\item \code{structure_RLum(RLum.Analysis)}: Returns the structure of an \link[=RLum.Analysis-class]{RLum.Analysis} object.

}}
\section{Function version}{
 0.2.0
}

\examples{

## load example data
data(ExampleData.XSYG, envir = environment())

## show structure
structure_RLum(OSL.SARMeasurement$Sequence.Object)

}
\seealso{
\link[=RLum.Data.Curve-class]{RLum.Data.Curve}, \link[=RLum.Data.Image-class]{RLum.Data.Image},
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}, \link[=RLum.Analysis-class]{RLum.Analysis},
\link[=RLum.Results-class]{RLum.Results}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{utilities}
