% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_SingleGrainData.R
\name{subset_SingleGrainData}
\alias{subset_SingleGrainData}
\title{Simple Subsetting of Single Grain Data from Risø BIN/BINX files}
\usage{
subset_SingleGrainData(object, selection)
}
\arguments{
\item{object}{\link[=Risoe.BINfileData-class]{Risoe.BINfileData} (\strong{required}):
input object with the data to subset.}

\item{selection}{\link{data.frame}, \link{matrix} (\strong{required}):
selection table with two columns for position and grain, respectively.}
}
\value{
A subset \link[=Risoe.BINfileData-class]{Risoe.BINfileData} object
}
\description{
Most measured single grains do not exhibit light and it usually makes sense
to subset single grain datasets using a table of position and grain pairs.
}
\section{Function version}{
 0.1.1
}

\examples{

## load example data
data(ExampleData.BINfileData, envir = environment())

## set POSITION/GRAIN pair dataset
selection <- data.frame(POSITION = c(1,5,7), GRAIN = c(0,0,0))

##subset
subset_SingleGrainData(object = CWOSL.SAR.Data, selection = selection)

}
\seealso{
\link[=Risoe.BINfileData-class]{Risoe.BINfileData}, \link{read_BIN2R},
\link{verify_SingleGrainData}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2026. subset_SingleGrainData(): Simple Subsetting of Single Grain Data from Risø BIN/BINX files. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{datagen}
\keyword{manip}
