% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data-class.R, R/RLum.Results-class.R
\name{summary}
\alias{summary}
\alias{summary,RLum.Analysis-method}
\alias{summary,RLum.Data-method}
\alias{summary,RLum.Results-method}
\title{Summary of RLum-class objects}
\usage{
\S4method{summary}{RLum.Analysis}(object, ...)

\S4method{summary}{RLum.Data}(object, ...)

\S4method{summary}{RLum.Results}(object, ...)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
S4 object of class \code{RLum}.}

\item{...}{further arguments passed to \link[base:summary]{base::summary}.}
}
\value{
\itemize{
\item \link[=RLum.Analysis-class]{RLum.Analysis}: a list with the summary of each record.
}

\itemize{
\item \link[=RLum.Data-class]{RLum.Data}: the summary of the object's \code{data} slot.
}

\itemize{
\item \link[=RLum.Results-class]{RLum.Results}: the data stored in the object (via
\link{get_RLum}).
}
}
\description{
Summary of RLum-class objects
}
\section{Function version}{
 0.1.0
}

\author{
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
}
\keyword{utilities}
