% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_Data.R
\name{tune_Data}
\alias{tune_Data}
\title{Tune data for experimental purpose}
\usage{
tune_Data(data, decrease.error = 0, increase.data = 0)
}
\arguments{
\item{data}{\link{data.frame} (\strong{required}):
input values, structure: values (\code{data[, 1]}) and error (\code{data[, 2]})
are required.}

\item{decrease.error}{\link{numeric}:
factor by which the error is decreased, ranges between 0 and 1.}

\item{increase.data}{\link{numeric}:
factor by which the error is decreased, ranges between 0 and 1000.}
}
\value{
Returns a 2-column \link{data.frame} with tuned values.
}
\description{
The error can be reduced and sample size increased for specific purpose.
}
\note{
You should not use this function to improve your poor data set!
}
\section{Function version}{
 0.5.1
}

\examples{

## load example data set
data(ExampleData.DeValues, envir = environment())
x <- ExampleData.DeValues$CA1

## plot original data
plot_AbanicoPlot(data = x,
                 summary = c("n", "mean"))

## decrease error by 10 \%
plot_AbanicoPlot(data = tune_Data(x, decrease.error = 0.1),
                 summary = c("n", "mean"))

## increase sample size by 200 \%
#plot_AbanicoPlot(data = tune_Data(x, increase.data = 2),
#                summary = c("n", "mean"))

}
\author{
Michael Dietze, GFZ Potsdam (Germany)
, RLum Developer Team} 

\section{How to cite}{
Dietze, M., 2026. tune_Data(): Tune data for experimental purpose. Function version 0.5.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{manip}
