% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data-class.R, R/RLum.Results-class.R,
%   R/Risoe.BINfileData-class.R
\name{view}
\alias{view}
\alias{view,RLum.Analysis-method}
\alias{view,RLum.Data-method}
\alias{view,RLum.Results-method}
\alias{view,Risoe.BINfileData-method}
\title{Convenience data visualisation function}
\usage{
view(object, ...)

\S4method{view}{RLum.Analysis}(object, ...)

\S4method{view}{RLum.Data}(object, ...)

\S4method{view}{RLum.Results}(object, element = 1, ...)

\S4method{view}{Risoe.BINfileData}(object, ...)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
the object to view.}

\item{...}{further arguments passed to the specific class method.}

\item{element}{\link{integer} (\emph{with default}):
index of the element to display.}
}
\value{
The data is shown in the viewer.
}
\description{
The function invokes the \link[utils:View]{utils::View} function tailored to objects in the
package. If started from RStudio, it uses the RStudio viewer.
}
\section{Methods (by class)}{
\itemize{
\item \code{view(RLum.Analysis)}: View method for \link[=RLum.Analysis-class]{RLum.Analysis} objects.

\item \code{view(RLum.Data)}: View method for \link[=RLum.Data-class]{RLum.Data} objects.

\item \code{view(RLum.Results)}: View method for \link[=RLum.Results-class]{RLum.Results} objects.

\item \code{view(Risoe.BINfileData)}: View method for \link[=Risoe.BINfileData-class]{Risoe.BINfileData} objects.

}}
\section{Function version}{
 0.1.0
}

\seealso{
\code{\link[utils:View]{utils::View()}}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{utilities}
