% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_Risoe.BINfileData.R
\name{view_Risoe.BINfileData}
\alias{view_Risoe.BINfileData}
\title{Data view function}
\usage{
view_Risoe.BINfileData(object, ...)
}
\arguments{
\item{object}{\linkS4class{Risoe.BINfileData} (\strong{required}):}

\item{...}{further arguments}
}
\value{
NULL and a view window
}
\description{
Data view function that sits on \link[utils:View]{utils::View}
}
\section{Function version}{
 0.1.0
}

\seealso{
\linkS4class{Risoe.BINfileData}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{internal}
\keyword{utilities}
