% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_RLum.Data.R
\name{view_RLum.Data}
\alias{view_RLum.Data}
\title{Data view function}
\usage{
view_RLum.Data(object, ...)
}
\arguments{
\item{object}{\linkS4class{RLum.Data} (\strong{required}):}

\item{...}{further arguments}
}
\value{
NULL and a view window
}
\description{
Data view function that sits on \link[utils:View]{utils::View}
}
\section{Function version}{
 0.1.0
}

\seealso{
\linkS4class{RLum.Data}
}
\author{
Marco Colombo, Institute of Geography, Heidelberg University (Germany)
}
\keyword{internal}
\keyword{utilities}
