% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_RLum.Results.R
\name{view_RLum.Results}
\alias{view_RLum.Results}
\title{Data view function}
\usage{
view_RLum.Results(object, element = 1, ...)
}
\arguments{
\item{object}{\linkS4class{RLum.Results} (\strong{required}):}

\item{element}{\link{integer} (\emph{with default}): index of the element to display}

\item{...}{further arguments}
}
\value{
NULL and a view window
}
\description{
Data view function that sits on \link[utils:View]{utils::View}
}
\section{Function version}{
 0.1.0
}

\seealso{
\linkS4class{RLum.Results}
}
\author{
Marco Colombo, Institute of Geography, Heidelberg University (Germany)
}
\keyword{internal}
\keyword{utilities}
