% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_R2TIFF.R
\name{write_R2TIFF}
\alias{write_R2TIFF}
\title{Export RLum.Data.Image and RLum.Data.Spectrum objects to TIFF Images}
\usage{
write_R2TIFF(object, file = tempfile(), norm = 65535, ...)
}
\arguments{
\item{object}{\link[=RLum.Data.Image-class]{RLum.Data.Image} or \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} object (\strong{required}):
input object, can be a \link{list} of such objects.}

\item{file}{\link{character} (\strong{required}):
name of the output file.}

\item{norm}{\link{numeric} (\emph{with default}):
normalisation value. Usually, in imaging applications the pixel values are
integer count values, but values in TIFF files must be in the 0-1 range.
Normalising to the to the highest 16-bit integer values - 1 ensures that
the numerical values are retained in the exported image. If \code{1} nothing is
normalised.}

\item{...}{further arguments to be passed to \link[tiff:writeTIFF]{tiff::writeTIFF}.}
}
\value{
A TIFF file
}
\description{
Simple wrapper around \link[tiff:writeTIFF]{tiff::writeTIFF} to export suitable \link[=RLum-class]{RLum}
objects to TIFF images. Per default 16-bit TIFF files are exported.
}
\section{Function version}{
 0.1.2
}

\examples{
data(ExampleData.RLum.Data.Image, envir = environment())
write_R2TIFF(ExampleData.RLum.Data.Image, file = tempfile(fileext = ".tiff"))

}
\seealso{
\link[tiff:writeTIFF]{tiff::writeTIFF}, \link[=RLum.Data.Image-class]{RLum.Data.Image},
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2026. write_R2TIFF(): Export RLum.Data.Image and RLum.Data.Spectrum objects to TIFF Images. Function version 0.1.2. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{IO}
