% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_Al2O3C_Measurement.R
\name{analyse_Al2O3C_Measurement}
\alias{analyse_Al2O3C_Measurement}
\title{Al2O3:C Passive Dosimeter Measurement Analysis}
\usage{
analyse_Al2O3C_Measurement(object, signal_integral = NULL,
  dose_points = c(0, 4), recordType = c("OSL (UVVIS)", "TL (UVVIS)"),
  irradiation_time_correction = NULL, cross_talk_correction = NULL,
  travel_dosimeter = NULL, verbose = TRUE, plot = TRUE, ...)
}
\arguments{
\item{object}{\linkS4class{RLum.Analysis} \strong{(required)}:
measurement input}

\item{signal_integral}{\link{numeric} (\emph{optional}):
signal integral, used for the signal and the background.
If nothing is provided the full range is used}

\item{dose_points}{\link{numeric} (\emph{with default}):
vector with dose points, if dose points are repeated, only the general
pattern needs to be provided. Default values follow the suggestions
made by Kreutzer et al., XXXX}

\item{recordType}{\link{character} (\emph{with default}): input curve selection, which is passed to
function \link{get_RLum}. To deactivate the automatic selection set the argument to \code{NULL}}

\item{irradiation_time_correction}{\link{numeric} or \linkS4class{RLum.Results} (\emph{optional}):
information on the used irradiation time correction obained by another experiements.
I a \code{numeric} is provided it has to be of length two: mean, standard error}

\item{cross_talk_correction}{\link{numeric} or \linkS4class{RLum.Results} (\emph{optional}):
information on the used irradiation time correction obained by another experiements.
If a \code{numeric} vector is provided it has to be of length three:
mean, 2.5 \% quantile, 97.5 \% quantile.}

\item{travel_dosimeter}{\link{numeric} (\emph{optional}): specify the position of the travel dosimeter
(so far measured a the same time). The dose of travel dosimeter will be subtracted from all
other values.}

\item{verbose}{\link{logical} (\emph{with default}):
enable/disable verbose mode}

\item{plot}{\link{logical} (\emph{with default}):
enable/disable plot output}

\item{...}{further arguments that can be passed to the plot output}
}
\value{
Function returns results numerically and graphically:

-----------------------------------\cr
\code{[ NUMERICAL OUTPUT ]}\cr
-----------------------------------\cr

\strong{\code{RLum.Results}}-object

\strong{slot:} \strong{\code{@data}}

\tabular{lll}{
\strong{Element} \tab \strong{Type} \tab \strong{Description}\cr
\code{$data} \tab \code{data.frame} \tab the estimated equivalent dose \cr
\code{$data_talbe} \tab \code{data.frame} \tab full dose and signal table \cr
}

\strong{slot:} \strong{\code{@info}}

The original function call

------------------------\cr
\code{[ PLOT OUTPUT ]}\cr
------------------------\cr
\itemize{
\item OSL and TL curves, combined on two plots.
}
}
\description{
The function provides the analysis routines for measurements on a
FI lexsyg SMART reader using Al2O3:C pellets according to Kreutzer et al., 2017
}
\details{
##ADD INFORMATION ON HOW IT WORKS WITH THE TRAVEL DOSIMETERS
}
\section{Function version}{
 0.1.1 (2017-07-18 21:46:37)
}

\examples{

## nothing so far TODO ... add tests with example

} 

\section{How to cite}{
Kreutzer, S. (2017). analyse_Al2O3C_Measurement(): Al2O3:C Passive Dosimeter Measurement Analysis. Function version 0.1.1. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence
}

\references{
TODO
}
\seealso{
\link{analyse_Al2O3C_ITC}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\keyword{datagen}
