% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_Al2O3C_Measurement.R
\name{analyse_Al2O3C_Measurement}
\alias{analyse_Al2O3C_Measurement}
\title{Al2O3:C Passive Dosimeter Measurement Analysis}
\usage{
analyse_Al2O3C_Measurement(object, signal_integral = NULL,
  dose_points = c(0, 4), irradiation_time_correction = NULL,
  cross_talk_correction = NULL, verbose = TRUE, plot = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{RLum.Analysis}} \bold{(required)}: measurement input}

\item{signal_integral}{\code{\link{numeric}} (optional): signal integral, used for the signal
and the background. If nothing is provided the full range is used}

\item{dose_points}{\code{\link{numeric}} (with default): vector with dose points, if dose points
are repeated, only the general pattern needs to be provided. Default values follow the suggestions
made by Kreutzer et al., 2017}

\item{irradiation_time_correction}{\code{\link{numeric}} or \code{\linkS4class{RLum.Results}} (optional):
information on the used irradiation time correction obained by another experiements. I a \code{numeric}
is provided it has to be of length two: mean, standard error}

\item{cross_talk_correction}{\code{\link{numeric}} or \code{\linkS4class{RLum.Results}} (optional):
information on the used irradiation time correction obained by another experiements. If a \code{numeric}
vector is provided it has to be of length three: mean, 2.5 \% quantile, 97.5 \% quantile.}

\item{verbose}{\code{\link{logical}} (with default): enable/disable verbose mode}

\item{plot}{\code{\link{logical}} (with default): enable/disable plot output}

\item{...}{further arguments that can be passed to the plot output}
}
\value{
Function returns results numerically and graphically:\cr

-----------------------------------\cr
[ NUMERICAL OUTPUT ]\cr
-----------------------------------\cr
\bold{\code{RLum.Reuslts}}-object\cr

\bold{slot:} \bold{\code{@data}}\cr
\tabular{lll}{
\bold{Element} \tab \bold{Type} \tab \bold{Description}\cr
 \code{$data} \tab \code{data.frame} \tab the estimated equivalent dose \cr
 \code{$data_talbe} \tab \code{data.frame} \tab full dose and signal table \cr
}

\bold{slot:} \bold{\code{@info}}\cr

The original function call\cr

------------------------\cr
[ PLOT OUTPUT ]\cr
------------------------\cr

\itemize{
 \item OSL and TL curves, combined on two plots.

}
}
\description{
The function provides the analysis routines for measurements on a FI lexsyg SMART reader using
Al2O3:C pellets according to Kreutzer et al., 2017
}
\section{Function version}{
 0.1.0
}

\examples{

##nothing so far TODO ... add tests with example

}
\references{
TODO
}
\seealso{
\code{\link{analyse_Al2O3C_ITC}}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
}
\keyword{datagen}
