% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_baSAR.R
\name{analyse_baSAR}
\alias{analyse_baSAR}
\title{Bayesian models (baSAR) applied on luminescence data}
\usage{
analyse_baSAR(object, XLS_file = NULL, aliquot_range = NULL,
  source_doserate = c(1, 0), signal.integral, signal.integral.Tx = NULL,
  background.integral, background.integral.Tx = NULL, sigmab = 0,
  sig0 = 0.025, distribution = "cauchy", fit.method = "EXP",
  fit.force_through_origin = TRUE, fit.includingRepeatedRegPoints = TRUE,
  plot = TRUE, plot_reduced = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{Risoe.BINfileData}} or \code{\linkS4class{RLum.Results}} or
\code{{character}} or \code{\link{list}} (\bold{required}):
input object used for the Bayesian analysis. If a \code{character} is provided the function
assumes a file connection and tries to import a BIN-file using the provided path. If a \code{list} is
provided the list can only contain either \code{Risoe.BINfileData} objects or \code{character}s
providing a file connection. Mixing of both types is not allowed. If an \code{\linkS4class{RLum.Results}}
is provided the function direclty starts with the Bayesian Analysis (see details)}

\item{XLS_file}{\code{\link{character}} (optional): XLS_file with data for the analysis TODO}

\item{aliquot_range}{\code{\link{numeric}} (optional): allows to limit the range of the aliquots
used for the analysis. This argument has only an effect of the argument \code{XLS_file} is used as
well}

\item{source_doserate}{\code{\link{numeric}} (with default): source dose rate of beta-source used
for the measuremnt and its uncertainty in Gy/s}

\item{signal.integral}{\code{\link{vector}} (optional): vector with the
limits for the signal integral used for the calculation, e.g., \code{signal.integral = c(1:5)}
Ignored if \code{object} is an \code{\linkS4class{RLum.Results}} object.}

\item{signal.integral.Tx}{\code{\link{vector}} (optional): vector with the
limits for the signal integral for the Tx curve. If nothing is provided the
value from \code{signal.integral} is used and it is ignored
if \code{object} is an \code{\linkS4class{RLum.Results}} object.}

\item{background.integral}{\code{\link{vector}} (\bold{required}): vector with the
bounds for the background integral.
Ignored if \code{object} is an \code{\linkS4class{RLum.Results}} object}

\item{background.integral.Tx}{\code{\link{vector}} (optional): vector with the
limits for the background integral for the Tx curve. If nothing is provided the
value from \code{background.integral} is used.
Ignored if \code{object} is an \code{\linkS4class{RLum.Results}} object}

\item{sigmab}{\code{\link{numeric}} (with default): option to set a manual value for
the overdispersion (for LnTx and TnTx), used for the Lx/Tx error
calculation. The value should be provided as absolute squared count values, cf. \code{\link{calc_OSLLxTxRatio}}}

\item{sig0}{\code{\link{numeric}} (with default): allow adding an extra component of error
to the final Lx/Tx error value (e.g., instrumental errror, see details is \code{\link{calc_OSLLxTxRatio}})}

\item{distribution}{\code{\link{character}} (with default): type of distribution that is used for
the Bayesian calculation. Allowed inputs are \code{cauchy},\code{normal} and \code{log_normal}}

\item{fit.method}{\code{\link{character}} (with default): fit method used for fitting the growth
curve using the function \code{\link{plot_GrowthCurve}}. Here supported methods: \code{EXP},
\code{EXP+LIN} and \code{LIN}}

\item{fit.force_through_origin}{\code{\link{logical}} (with default): force fitting through origin}

\item{fit.includingRepeatedRegPoints}{\code{\link{logical}} (with default):
includes the recycling point (assumed to be measured during the last cycle)}

\item{plot}{\code{\link{logical}} (with default): enables or disables plot output}

\item{plot_reduced}{\code{\link{logical}} (with default): enables or disables the advanced plot output}

\item{verbose}{\code{\link{logical}} (with default): enables or disables verbose mode}

\item{...}{parameters that can be passed to the function \code{\link{calc_OSLLxTxRatio}} (almost full support)
\code{\link[readxl]{read_excel}} (full support), \code{\link{read_BIN2R}} (\code{n.records},
\code{position}, \code{duplicated.rm}), see details.}
}
\description{
This function allows the application of Bayesian models (baSAR) on luminescecence data
}
\details{
\bold{Additional arguments support via the \code{...} argument}\cr

\tabular{llll}{
\bold{Supported argument} \tab \bold{Corresponding function} \tab \bold{Default} \tab \bold{Short description}\cr
\code{sheet} \tab \code{\link[readxl]{read_excel}} \tab \code{1} \tab select XLS-sheet for import\cr
\code{col_names} \tab \code{\link[readxl]{read_excel}} \tab \code{TRUE} \tab first column in XLS-file is header\cr
\code{col_types} \tab \code{\link[readxl]{read_excel}} \tab \code{NULL} \tab limit import to specific columns\cr
\code{skip} \tab \code{\link[readxl]{read_excel}} \tab \code{0} \tab number of rows to be skipped during import\cr
\code{n.records} \tab \code{\link{read_BIN2R}} \tab \code{NULL} \tab limit records during BIN-file import\cr
\code{duplicated.rm} \tab \code{\link{read_BIN2R}} \tab \code{TRUE} \tab remove duplicated records in the BIN-file\cr
\code{position} \tab \code{\link{read_BIN2R}} \tab \code{NULL} \tab limit import to a specific position\cr
\code{background.count.distribution} \tab \code{\link{calc_OSLLxTxRatio}} \tab \code{"non-poisson"} \tab set assumed count distribution\cr
\code{fit.weights} \tab \code{\link{plot_GrowthCurve}} \tab \code{TRUE} \tab enable/disable fit weights\cr
\code{fit.bounds} \tab \code{\link{plot_GrowthCurve}} \tab \code{TRUE} \tab enable/disable fit bounds\cr
\code{NumberIterations.MC} \tab \code{\link{plot_GrowthCurve}} \tab \code{100} \tab number of MC runs for error calculation\cr
\code{output.plot} \tab \code{\link{plot_GrowthCurve}} \tab \code{TRUE} \tab enables/disables dose repsonse curve plot\cr
\code{output.plotExtended} \tab \code{\link{plot_GrowthCurve}} \tab \code{TRUE} \tab enables/disables extended dose repsonse curve plot\cr
}
}
\note{
The current version of the function works with standard Risoe BIN-files only!
}
\section{Function version}{
 0.1.0 (2016-05-18 19:58:26)
}
\examples{

#nothing so far

}
\author{
Norbert Mercier, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France), Sebastian Kreutzer,
IRAMAT-CRP2A, Universite Bordeaux Montaigne (France) \cr

The underlying Bayesian model based on a contribution by Combes et al., 2015.
\cr R Luminescence Package Team}
\references{
Combes, B., Philippe, A., Lanos, P., Mercier, N., Tribolo, C., Guerin, G., Guibert, P., Lahaye, C., 2015.
A Bayesian central equivalent dose model for optically stimulated luminescence dating.
Quaternary Geochronology 28, 62-70. doi:10.1016/j.quageo.2015.04.001
}
\seealso{
\code{\link{read_BIN2R}}, \code{\link{calc_OSLLxTxRatio}}, \code{\link{plot_GrowthCurve}},
\code{\link[readxl]{read_excel}}, \code{\link{verify_SingleGrainData}},
\code{\link[rjags]{jags.model}}, \code{\link[rjags]{coda.samples}}
}
\keyword{datagen}

