% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_baSAR.R
\name{analyse_baSAR}
\alias{analyse_baSAR}
\title{Bayesian models (baSAR) applied on luminescence data}
\usage{
analyse_baSAR(object, source_doserate = 1, signal.integration.limits = c(6,
  16, 26, 60, 6, 16, 26, 60), distribution = "normal", fit.method = "EXP",
  XLS_file = NULL, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{\code{\link{character}} (\bold{required}): BIN-file, either a single file or a list
of file names}

\item{source_doserate}{\code{\link{numeric}} (with default): source dose rate of beta-source used
for the measuremnt in Gy/s}

\item{signal.integration.limits}{\code{\link{numeric}} (with default): signal integration limits
TODO: ask Norbert what does it mean here!}

\item{distribution}{\code{\link{character}} (with default): type of distribution that is used for
the Bayesian calculation. Allowed inputs are \code{normal}, \code{cauchy} and \code{log_normal}}

\item{fit.method}{\code{\link{character}} (with default): fit method used for fitting the growth
curve using the function \code{\link{plot_GrowthCurve}}. Here supported methods: \code{EXP},
\code{EXP+LIN} and \code{LIN}}

\item{XLS_file}{\code{\link{character}} (with default): XLS_file with data for the analysis}

\item{plot}{\code{\link{logical}} (with default): enables or disables plot output}

\item{verbose}{\code{\link{logical}} (with default): enables or disables verbose mode}
}
\description{
This function allows the application of Bayesian models (baSAR) on luminescecence data
}
\section{Function version}{
 0.1.0 (2016-04-20 11:47:27)
}
\examples{

#nothing so far

}
\author{
Norbert Mercier, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France), Sebastian Kreutzer,
IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\references{
nothing so far
}
\seealso{
\code{\link{read_BIN2R}}, \code{\link[rjags]{jags.model}}, \code{\link[rjags]{coda.samples}}
}
\keyword{datagen}

