% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_baSAR.R
\name{analyse_baSAR}
\alias{analyse_baSAR}
\title{Bayesian models (baSAR) applied on luminescence data}
\usage{
analyse_baSAR(object, source_doserate = 1, Lx.data, Tx.data, signal.integral,
  signal.integral.Tx = NULL, background.integral,
  background.integral.Tx = NULL,
  background.count.distribution = "non-poisson", sigmab, sig0 = 0,
  digits = NULL, distribution = "normal", fit.method = "EXP",
  XLS_file = NULL, plot = TRUE, plot_reduced = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{\code{\linkS4class{Risoe.BINfileData}} or \code{\link{character}} or \code{\link{list}}
(\bold{required}): input object used for the Bayesian analysis. If a \code{character} is provided the function
assums a file connection an tries to import a BIN-file using the provided path. If a \code{list} is
provided the list can only contain either \code{Risoe.BINfileData} objects or \code{character}s
providing a file connection. Mixing of both types is not allowed.}

\item{source_doserate}{\code{\link{numeric}} (with default): source dose rate of beta-source used
for the measuremnt in Gy/s}

\item{Lx.data}{\code{\linkS4class{RLum.Data.Curve}} or \link{data.frame}
(\bold{required}): requires a CW-OSL shine down curve (x = time, y = counts)}

\item{Tx.data}{\code{\linkS4class{RLum.Data.Curve}} or \link{data.frame}
(optional): requires a CW-OSL shine down curve (x = time, y = counts). If no
input is given the Tx.data will be treated as \code{NA} and no Lx/Tx ratio
is calculated.}

\item{signal.integral}{\code{\link{vector}} (\bold{required}): vector with the
limits for the signal integral.}

\item{signal.integral.Tx}{\code{\link{vector}} (optional): vector with the
limits for the signal integral for the Tx curve. If nothing is provided the
value from \code{signal.integral} is used.}

\item{background.integral}{\code{\link{vector}} (\bold{required}): vector with the
bounds for the background integral.}

\item{background.integral.Tx}{\code{\link{vector}} (optional): vector with the
limits for the background integral for the Tx curve. If nothing is provided the
value from \code{background.integral} is used.}

\item{background.count.distribution}{\code{\link{character}} (with default): sets
the count distribution assumed for the error calculation. Possible arguments
\code{poisson} or \code{non-poisson}. See details for further information}

\item{sigmab}{\code{\link{numeric}} (optional): option to set a manual value for
the overdispersion (for LnTx and TnTx), used for the Lx/Tx error
calculation. The value should be provided as absolute squared count values,
e.g. \code{sigmab = c(300,300)}. Note: If only one value is provided this
value is taken for both (LnTx and TnTx) signals.}

\item{sig0}{\code{\link{numeric}} (with default): allow adding an extra component of error
to the final Lx/Tx error value (e.g., instrumental errror, see details).}

\item{digits}{\code{\link{integer}} (with default): round numbers to the specified digits. If
digits is set to \code{NULL} nothing is rounded.}

\item{distribution}{\code{\link{character}} (with default): type of distribution that is used for
the Bayesian calculation. Allowed inputs are \code{normal}, \code{cauchy} and \code{log_normal}}

\item{fit.method}{\code{\link{character}} (with default): fit method used for fitting the growth
curve using the function \code{\link{plot_GrowthCurve}}. Here supported methods: \code{EXP},
\code{EXP+LIN} and \code{LIN}}

\item{XLS_file}{\code{\link{character}} (with default): XLS_file with data for the analysis}

\item{plot}{\code{\link{logical}} (with default): enables or disables plot output}

\item{plot_reduced}{\code{\link{logical}} (with default): enables or disables the advanced plot output}

\item{verbose}{\code{\link{logical}} (with default): enables or disables verbose mode}
}
\description{
This function allows the application of Bayesian models (baSAR) on luminescecence data
}
\note{
The current version of the function works with standard Risoe BIN-files only!
}
\section{Function version}{
 0.1.0 (2016-05-04 10:18:28)
}
\examples{

#nothing so far

}
\author{
Norbert Mercier, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France), Sebastian Kreutzer,
IRAMAT-CRP2A, Universite Bordeaux Montaigne (France) \cr

The underlying Bayesian model based on a contribution by Combes et al., 2015.
\cr R Luminescence Package Team}
\references{
Combes, B., Philippe, A., Lanos, P., Mercier, N., Tribolo, C., Guerin, G., Guibert, P., Lahaye, C., 2015.
A Bayesian central equivalent dose model for optically stimulated luminescence dating.
Quaternary Geochronology 28, 62-70. doi:10.1016/j.quageo.2015.04.001
}
\seealso{
\code{\link{read_BIN2R}}, \code{\link{plot_GrowthCurve}},
\code{\link[rjags]{jags.model}}, \code{\link[rjags]{coda.samples}}
}
\keyword{datagen}

