% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_baSAR.R
\name{analyse_baSAR}
\alias{analyse_baSAR}
\title{Bayesian models (baSAR) applied on luminescence data}
\usage{
analyse_baSAR(object, XLS_file = NULL, aliquot_range = NULL,
  source_doserate = NULL, signal.integral, signal.integral.Tx = NULL,
  background.integral, background.integral.Tx = NULL, sigmab = 0,
  sig0 = 0.025, distribution = "cauchy", n.MCMC = 1e+05,
  fit.method = "EXP", fit.force_through_origin = TRUE,
  fit.includingRepeatedRegPoints = TRUE, method_control = list(),
  plot = TRUE, plot_reduced = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{Risoe.BINfileData}} or \code{\linkS4class{RLum.Results}} or
\code{\link{character}} or \code{\link{list}} (\bold{required}):
input object used for the Bayesian analysis. If a \code{character} is provided the function
assumes a file connection and tries to import a BIN-file using the provided path. If a \code{list} is
provided the list can only contain either \code{Risoe.BINfileData} objects or \code{character}s
providing a file connection. Mixing of both types is not allowed. If an \code{\linkS4class{RLum.Results}}
is provided the function direclty starts with the Bayesian Analysis (see details)}

\item{XLS_file}{\code{\link{character}} (optional): XLS_file with data for the analysis. This file must contain 3 columns: the name of the file, the disc position and the grain position (the last being 0 for multi-grain measurements)}

\item{aliquot_range}{\code{\link{numeric}} (optional): allows to limit the range of the aliquots
used for the analysis. This argument has only an effect if the argument \code{XLS_file} is used as
well}

\item{source_doserate}{\code{\link{numeric}} (optional): source dose rate of beta-source used
for the measuremnt and its uncertainty in Gy/s, e.g., \code{source_doserate = c(0.12, 0.04)}.
If nothing is provided the results are returned in the same domain as the input values.
Paramater can be provided as \code{list}, for the case that more than one BIN-file is provided, e.g.,
\code{source_doserate = list(c(0.04, 0.004), c(0.05, 0.004))}.}

\item{signal.integral}{\code{\link{vector}} (\bold{required}): vector with the
limits for the signal integral used for the calculation, e.g., \code{signal.integral = c(1:5)}
Ignored if \code{object} is an \code{\linkS4class{RLum.Results}} object.
The parameter can be provided as \code{list}, \code{source_doserate}.}

\item{signal.integral.Tx}{\code{\link{vector}} (optional): vector with the
limits for the signal integral for the Tx curve. If nothing is provided the
value from \code{signal.integral} is used and it is ignored
if \code{object} is an \code{\linkS4class{RLum.Results}} object.
The parameter can be provided as \code{list}, \code{source_doserate}.}

\item{background.integral}{\code{\link{vector}} (\bold{required}): vector with the
bounds for the background integral.
Ignored if \code{object} is an \code{\linkS4class{RLum.Results}} object.
The parameter can be provided as \code{list}, \code{source_doserate}.}

\item{background.integral.Tx}{\code{\link{vector}} (optional): vector with the
limits for the background integral for the Tx curve. If nothing is provided the
value from \code{background.integral} is used.
Ignored if \code{object} is an \code{\linkS4class{RLum.Results}} object.
The parameter can be provided as \code{list}, \code{source_doserate}.}

\item{sigmab}{\code{\link{numeric}} (with default): option to set a manual value for
the overdispersion (for LnTx and TnTx), used for the Lx/Tx error
calculation. The value should be provided as absolute squared count values, cf. \code{\link{calc_OSLLxTxRatio}}.
The parameter can be provided as \code{list}, \code{source_doserate}.}

\item{sig0}{\code{\link{numeric}} (with default): allow adding an extra component of error
to the final Lx/Tx error value (e.g., instrumental errror, see details is \code{\link{calc_OSLLxTxRatio}}).
The parameter can be provided as \code{list}, \code{source_doserate}.}

\item{distribution}{\code{\link{character}} (with default): type of distribution that is used during
Bayesian calculations for determining the Central dose and overdispersion values.
Allowed inputs are \code{"cauchy"}, \code{"normal"} and \code{"log_normal"}.}

\item{n.MCMC}{\code{\link{integer}} (with default): number of iterations for the Markov chain Monte Carlo (MCMC)
simulations}

\item{fit.method}{\code{\link{character}} (with default): fit method used for fitting the growth
curve using the function \code{\link{plot_GrowthCurve}}. Here supported methods: \code{EXP},
\code{EXP+LIN} and \code{LIN}}

\item{fit.force_through_origin}{\code{\link{logical}} (with default): force fitting through origin}

\item{fit.includingRepeatedRegPoints}{\code{\link{logical}} (with default):
includes the recycling point (assumed to be measured during the last cycle)}

\item{method_control}{\code{\link{list}} (optional): named list of control parameters that can be directly
passed to the Bayesian analysis, e.g., \code{method_control = list(lower_De = 0.01)}.
See details for further information}

\item{plot}{\code{\link{logical}} (with default): enables or disables plot output}

\item{plot_reduced}{\code{\link{logical}} (with default): enables or disables the advanced plot output}

\item{verbose}{\code{\link{logical}} (with default): enables or disables verbose mode}

\item{...}{parameters that can be passed to the function \code{\link{calc_OSLLxTxRatio}} (almost full support)
\code{\link[readxl]{read_excel}} (full support), \code{\link{read_BIN2R}} (\code{n.records},
\code{position}, \code{duplicated.rm}), see details.}
}
\description{
This function allows the application of Bayesian models on luminescence data, measured
with the single-aliquot regenerative-dose (SAR, Murray and Wintle, 2000) protocol. In particular,
it follows the idea proposed by Combes et al., 2015 of using an hierarchical model for estimating
a central equivalent dose from a set of luminescence measurements. This function is (I) the adaption
of this approach for the R environement and (II) an extension and a technical refinement of the
published code. \cr
}
\details{
Internally the function consists of two parts: (I) The Bayesian core for the bayesian calculations
and applying the hierchical model and (II) a data pre-processing part. The Bayesian core can be run
independently, if the input data are sufficient (see below). The data pre-processing part was
implemented to simplify the analysis for the user as all needed data pre-processing is done
by the function, i.e. in theory it is enough to provide a BIN/BINX-file with the SAR measurement
data. For the Bayesian analysis for each aliquot the following information are needed from the SAR analysis.
LxTx, the LxTx error and the dose values for all regeneration points.

\bold{Allowed input data}\cr

Various inputs are allowed for this function. Unfortunately this makes the function handling rather
complex, but at the same time very powerful. Available scenarios:\cr

\bold{(1) - \code{object} is BIN-file or link to a BIN-file}

Finally it does not matter how the information of the BIN/BINX file are provided. The function
supports (a) either a path to a file or directory or a \code{list} of file names or paths or (b)
a \code{\linkS4class{Risoe.BINfileData}} object or a list of these objects. The latter one can
be produced by using the function \code{\link{read_BIN2R}}, but this function is called automatically
if only a filename and/or a path is provided. In both cases it will become the data that can be
used for the analysis.

\code{XLS_file = NULL}\cr

If no XLS file (or data frame with the same format) is provided the functions runs an automatic process that
consists of the following steps:

\itemize{
 \item Select all valid aliquots using the function \code{\link{verify_SingleGrainData}}
 \item Calculate Lx/Tx values using the function \code{\link{calc_OSLLxTxRatio}}
 \item Calculate De values using the function \code{\link{plot_GrowthCurve}}
}

These proceded data are subsequently used in for the Bayesian analysis

\code{XLS_file != NULL}\cr

If an XLS-file is provided or a \code{data.frame} providing similar information the pre-processing
steps consists of the following steps:

\itemize{
 \item Calculate Lx/Tx values using the function \code{\link{calc_OSLLxTxRatio}}
 \item Calculate De values using the function \code{\link{plot_GrowthCurve}}
}

Means, the XLS file should contain a selection of the BIN-file names and the aliquots selected
for the further analysis. This allows a manual selection of input data, as the automatic selection
by \code{\link{verify_SingleGrainData}} might be not totally suffcient.\cr

\bold{(2) - \code{object} \code{RLum.Results object}}

If an \code{\linkS4class{RLum.Results}} object is provided as input and(!) this object was
previously created by the function \code{analyse_baSAR()} itself, the pre-processing part
is skipped and the function starts directly the Bayesian analysis. This option is very powerfull
as it allows to change parameters for the Bayesian analysis without the need to repeat
the data pre-processing.\cr

\bold{\code{method_control}}\cr

This are arguments that can be passed directly to the Bayesian calculation core, supported arguments
are:

\tabular{lll}{
\bold{Parameter} \tab \bold{Type} \tab \bold{Descritpion}\cr
\code{lower_De} \tab \code{\link{numeric}} \tab sets the lower bound for the expected De range\cr
\code{upper_De} \tab \code{\link{numeric}} \tab sets the upper bound for the expected De range\cr
}

\bold{Additional arguments support via the \code{...} argument}\cr

This list summarizes the additional arguments that can be passed to the internally used
funtions.

\tabular{llll}{
\bold{Supported argument} \tab \bold{Corresponding function} \tab \bold{Default} \tab \bold{Short description}\cr
\code{threshold} \tab \code{\link{verify_SingleGrainData}} \tab \code{30} \tab change rejection threshold for curve selection \cr
\code{sheet} \tab \code{\link[readxl]{read_excel}} \tab \code{1} \tab select XLS-sheet for import\cr
\code{col_names} \tab \code{\link[readxl]{read_excel}} \tab \code{TRUE} \tab first row in XLS-file is header\cr
\code{col_types} \tab \code{\link[readxl]{read_excel}} \tab \code{NULL} \tab limit import to specific columns\cr
\code{skip} \tab \code{\link[readxl]{read_excel}} \tab \code{0} \tab number of rows to be skipped during import\cr
\code{n.records} \tab \code{\link{read_BIN2R}} \tab \code{NULL} \tab limit records during BIN-file import\cr
\code{duplicated.rm} \tab \code{\link{read_BIN2R}} \tab \code{TRUE} \tab remove duplicated records in the BIN-file\cr
\code{pattern} \tab \code{\link{read_BIN2R}} \tab \code{TRUE} \tab select BIN-file by name pattern\cr
\code{position} \tab \code{\link{read_BIN2R}} \tab \code{NULL} \tab limit import to a specific position\cr
\code{background.count.distribution} \tab \code{\link{calc_OSLLxTxRatio}} \tab \code{"non-poisson"} \tab set assumed count distribution\cr
\code{fit.weights} \tab \code{\link{plot_GrowthCurve}} \tab \code{TRUE} \tab enable/disable fit weights\cr
\code{fit.bounds} \tab \code{\link{plot_GrowthCurve}} \tab \code{TRUE} \tab enable/disable fit bounds\cr
\code{NumberIterations.MC} \tab \code{\link{plot_GrowthCurve}} \tab \code{100} \tab number of MC runs for error calculation\cr
\code{output.plot} \tab \code{\link{plot_GrowthCurve}} \tab \code{TRUE} \tab enables/disables dose repsonse curve plot\cr
\code{output.plotExtended} \tab \code{\link{plot_GrowthCurve}} \tab \code{TRUE} \tab enables/disables extended dose repsonse curve plot\cr
}
}
\note{
\bold{If you provide more than one BIN-file}, it is \bold{strongly} recommanded to provide
a \code{list} with the same number of elements for the following parameters:\cr
\code{source_doserate}, \code{signal.integral}, \code{signal.integral.Tx}, \code{background.integral},
\code{background.integral.Tx}, \code{sigmab}, \code{sig0}.\cr

Example for two BIN-files: \code{source_doserate = list(c(0.04, 0.006), c(0.05, 0.006))}\cr

\bold{This function has beta status!} and is limited to work with
standard Risoe BIN-files only!
}
\section{Function version}{
 0.1.0 (2016-05-30 11:29:10)
}
\examples{

##(1) load package test data set
data(ExampleData.BINfileData, envir = environment())

##(2) selecting relevant curves, and limit dataset
CWOSL.SAR.Data <- subset(
 CWOSL.SAR.Data,
 subset = POSITION == c(1:3) & LTYPE == "OSL")

\dontrun{
##(3) run analysis
##please not that the here selected parameters are
##choosen for performance, not for reliability
analyse_baSAR(
 object = CWOSL.SAR.Data,
 signal.integral = c(1:2),
 background.integral = c(80:100),
 fit.method = "LIN",
 plot = FALSE,
 n.MCMC = 200
)


##XLS_file template
##copy and paste this the code below in the terminal
##you can further use the function write.csv() to export the example

XLS_file <-
structure(
list(
 BIN_FILE = NA_character_,
 DISC = NA_real_,
 GRAIN = NA_real_),
   .Names = c("BIN_FILE", "DISC", "GRAIN"),
   class = "data.frame",
   row.names = 1L
)

}

}
\author{
Norbert Mercier, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France), Sebastian Kreutzer,
IRAMAT-CRP2A, Universite Bordeaux Montaigne (France) \cr

The underlying Bayesian model based on a contribution by Combes et al., 2015.
\cr R Luminescence Package Team}
\references{
Combes, B., Philippe, A., Lanos, P., Mercier, N., Tribolo, C., Guerin, G., Guibert, P., Lahaye, C., 2015.
A Bayesian central equivalent dose model for optically stimulated luminescence dating.
Quaternary Geochronology 28, 62-70. doi:10.1016/j.quageo.2015.04.001

\bold{Further reading}

Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B., Vehtari, A., Rubin, D.B., 2013.
Bayesian Data Analysis, Third Edition. CRC Press.

Murray, A.S., Wintle, A.G., 2000. Luminescence dating of quartz using an improved single-aliquot
regenerative-dose protocol. Radiation Measurements 32, 57-73. doi:10.1016/S1350-4487(99)00253-X
}
\seealso{
\code{\link{read_BIN2R}}, \code{\link{calc_OSLLxTxRatio}}, \code{\link{plot_GrowthCurve}},
\code{\link[readxl]{read_excel}}, \code{\link{verify_SingleGrainData}},
\code{\link[rjags]{jags.model}}, \code{\link[rjags]{coda.samples}}
}
\keyword{datagen}

