% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_FadingMeasurement.R
\name{analyse_FadingMeasurement}
\alias{analyse_FadingMeasurement}
\title{Analyse fading measurements and returns the fading rate per decade (g-value)}
\usage{
analyse_FadingMeasurement(object, structure = c("Lx", "Tx"), signal.integral,
  background.integral, n.MC = 100, verbose = TRUE, plot = TRUE,
  plot.single = FALSE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{RLum.Analysis}} (\bold{required}): input object with the
measurement data. Alternatively a \code{\link{list}} containing \code{\linkS4class{RLum.Analysis}}
objects can be provided}

\item{structure}{\code{\link{character}} (with default): sets the structure of the measurement
data. Allowed are \code{'Lx'} or \code{c('Lx','Tx')}. Other input is ignored}

\item{signal.integral}{\code{\link{vector}} (\bold{required}): vector with the
limits for the signal integral}

\item{background.integral}{\code{\link{vector}} (\bold{required}): vector with the
bounds for the background integral}

\item{n.MC}{\code{\link{integer}} (with default): number for Monte Carlo runs for the error
estimation}

\item{verbose}{\code{\link{logical}} (with default): enables/disables verbose mode}

\item{plot}{\code{\link{logical}} (with default): enables/disables plot output}

\item{plot.single}{\code{\link{logical}} (with default): enables/disables single plot
mode, i.e. one plot window per plot}

\item{\dots}{(optional) further arguments that can be passed to internally used functions (see details)}
}
\value{
An RLum.Results objects is returned ... TODO
}
\description{
The function analysis fading measurements and returns a fading rate including an error estimation.
The function is not limited to standard fading measurements, as can be seen, e.g., Huntely and
Lamothe 2001.
}
\note{
\bold{This function has BETA status and should not be used for publication work!}
}
\section{Function version}{
 0.1.0 (2016-07-05 16:39:52)
}
\examples{

##nothing so far

}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\references{
-
}
\seealso{
\code{\link{calc_OSLLxTxRatio}}, \code{\link{read_BIN2R}}, \code{\link{read_XSYG2R}},
\code{\link{extract_IrradiationTimes}}
}
\keyword{datagen}

