% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyse_IRSAR.RF.R
\name{analyse_IRSAR.RF}
\alias{analyse_IRSAR.RF}
\title{Analyse IRSAR RF measurements}
\usage{
analyse_IRSAR.RF(object, sequence.structure = c("NATURAL", "REGENERATED"),
  RF_nat.lim, RF_reg.lim, method = "FIT", test_parameter, fit.trace = FALSE,
  n.MC = 10, slide.show_density = FALSE, txtProgressBar = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{RLum.Analysis}} (\bold{required}): input
object containing data for protocol analysis. Generally the function expects two curves.
(1) RF_nat, (2) RF_reg}

\item{sequence.structure}{\code{\link{vector}} \link{character} (with
default): specifies the general sequence structure. Allowed steps are
\code{NATURAL}, \code{REGENERATED}. In addition any other character is
allowed in the sequence structure; such curves will be ignored during the analysis.}

\item{RF_nat.lim}{\code{\link{vector}} (with default): set minimum and maximum
channel range for natural signal fitting and sliding. If only one value is provided this
will be treated as minium value and the maximum limit will be added automatically.}

\item{RF_reg.lim}{\code{\link{vector}} (with default): set minimum and maximum
channel range for regenerated signal fitting and sliding. If only one value is provided this
will be treated as minium value and the maximum limit will be added automatically.}

\item{method}{\code{\link{character}} (with default): setting method applied
for the data analysis. Possible options are \code{"FIT"} or \code{"SLIDE"}.}

\item{test_parameter}{\code{\link{list} (with default)}: set test parameter
Supported parameters are: \code{curves_ratio}, \code{residuals_slope} (only for
\code{method = "SLIDE"}), \code{curves_bounds}, \code{dynamic_ratio},
\code{lambda}, \code{beta} and \code{delta.phi}. All input: \code{\link{numeric}}
values, \code{NA} and \code{NULL} (s. Details)

(see Details for further information)}

\item{fit.trace}{\code{\link{logical}} (with default): trace fitting (for
debugging use)}

\item{n.MC}{\code{\link{numeric}} (with default): set number of Monte
Carlo runs for start parameter estimation (\code{method = "FIT"}) or
error estimation (\code{method = "SLIDE"}). Note: Large values will
significantly increase the computation time}

\item{slide.show_density}{\code{\link{logical}} (with default): enable or
disable KDE for MC run results. If the distribution is too narrow nothing is shown}

\item{txtProgressBar}{\code{\link{logical}} (with default): enables \code{TRUE} or
disables \code{FALSE} the progression bar during MC runs}

\item{plot}{\code{\link{logical}} (with default): plot output (\code{TRUE}
or \code{FALSE})}

\item{\dots}{further arguments that will be passed to the plot output.
Currently supported arguments are \code{main}, \code{xlab}, \code{ylab},
\code{xlim}, \code{ylim}, \code{log}, \code{legend.pos} (passes argument to x,y in
\code{\link[graphics]{legend}})}
}
\value{
A plot (optional) and an \code{\linkS4class{RLum.Results}} object is
returned. The slot data contains the following elements: \cr


$ De.values: \code{\link{data.frame}} table with De and corresponding values\cr
..$ DE : \code{numeric}: the obtained equivalent dose\cr
..$ DE.ERROR : \code{numeric}: (only method = "SLIDE") standard deviation obtained from MC runs \cr
..$ DE.LOWER : \code{numeric}: 2.5\% quantile for De values obtabined by MC runs \cr
..$ DE.UPPER : \code{numeric}: 97.5\% quantile for De values obtabined by MC runs  \cr
..$ DE.STATUS  : \code{character}: test parameter status\cr
..$ RF_NAT.LIM  : \code{charcter}: used RF_nat curve limits \cr
..$ RF_REG.LIM : \code{character}: used RF_reg curve limits\cr
..$ POSITION : \code{integer}: (optional) position of the curves\cr
..$ DATE : \code{character}: (optional) measurement date\cr
..$ SEQUENCE_NAME : \code{character}: (optional) sequence name)\cr
..$ UID : \code{character}: unique data set ID \cr
$ test_parameter : \code{\link{data.frame}} table test parameters \cr
$ fit : {\code{\link{nls}} \code{nlsModel} object} \cr
$ slide : \code{\link{list}} data from the sliding process\cr
$ call : \code{\link[methods]{language-class}}: the orignal function call \cr

The output (\code{De.values}) should be accessed using the
function \code{\link{get_RLum}}
}
\description{
Function to analyse IRSAR RF measurements on K-feldspar samples, performed
using the protocol according to Erfurt et al. (2003)
}
\details{
The function performs an IRSAR analysis described for K-feldspar samples by
Erfurt et al. (2003) assuming a negligible sensitivity change of the RF
signal.\cr \bold{General Sequence Structure} (according to Erfurt et al.
(2003)) \enumerate{ \item Measuring IR-RF intensity of the natural dose for
a few seconds (\eqn{D_{natural}}) \item Bleach the samples under solar
conditions for at least 30 min without changing the geometry \item Waiting
for at least one hour \item Regeneration of the IR-RF signal to at least the
natural level \item Fitting data with a stretched exponential function \item
Calculate the the palaeodose \eqn{D_{e}} using the parameters from the
fitting } \bold{Function Used For The Fitting} (according to Erfurt et al.
(2003))\cr \deqn{\phi(D) = \phi_{0}-\Delta\phi(1-exp(-\lambda*D))^\beta}
with \eqn{\phi(D)} the dose dependent IR-RF flux, \eqn{\phi_{0}} the inital
IR-RF flux, \eqn{\Delta\phi} the dose dependent change of the IR-RF flux,
\eqn{\lambda} the exponential parameter, \eqn{D} the dose and \eqn{\beta}
the dispersive factor.\cr\cr To obtain the palaeodose \eqn{D_{e}} the
function is changed to:\cr \deqn{D_{e} = ln(-(\phi(D) -
\phi_{0})/(-\lambda*\phi)^{1/\beta}+1)/-\lambda}\cr The fitting is done
using the \code{port} algorithm of the \code{\link{nls}} function.\cr

Two methods are supported to obtain the De:\cr

\bold{\code{method = "FIT"}}\cr

The principle is described above and follows the orignal suggestions by
Erfurt et al., 2003. For the fitting the mean count value of the RF_nat curve is used.

\bold{\code{method = "SLIDE"}}\cr

For this method the natural curve is slided along the x-axis until
congruence with the regenerated curve is reached. Instead of fitting this
allows to work with the original data without the need of any physical
model. This approach was introduced for RF curves by Buylaert et al., 2012
and Lapp et al., 2012.

Here the sliding is done by searching for the minimum of the squared residuals.

//WILL BE ADDED TODO

\bold{Error estimation}\cr

For \bold{\code{method = "FIT"}} the asymmetric error range is taken from
the standard deviation of the natural signal.\cr

For \bold{\code{method = "SLIDE"}} an

//WILL BE ADDED TODO

\bold{Test parameters}\cr

The argument \code{test_parameters} allows to evaluate ..

//WILL BE ADDED TODO
}
\note{
This function assumes that there is no sensitivity change during the
measurements (natural vs. regenerated signal), which is in contrast to the
findings from Buylaert et al. (2012).\cr
}
\section{Function version}{
 0.4.0 (2015-08-07 20:20:50)
}
\examples{
##load data
data(ExampleData.RLum.Analysis, envir = environment())

##perform analysis
temp <- analyse_IRSAR.RF(object = IRSAR.RF.Data)
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\references{
Buylaert, J.P., Jain, M., Murray, A.S., Thomsen, K.J., Lapp, T.,
2012. IR-RF dating of sand-sized K-feldspar extracts: A test of accuracy.
Radiation Measurements 44 (5-6), 560-565. doi: 10.1016/j.radmeas.2012.06.021

Erfurt, G., Krbetschek, M.R., 2003. IRSAR - A single-aliquot
regenerative-dose dating protocol applied to the infrared radiofluorescence
(IR-RF) of coarse- grain K-feldspar. Ancient TL 21, 35-42.

Erfurt, G., 2003. Infrared luminescence of Pb+ centres in potassium-rich
feldspars. physica status solidi (a) 200, 429-438.

Erfurt, G., Krbetschek, M.R., 2003. Studies on the physics of the infrared
radioluminescence of potassium feldspar and on the methodology of its
application to sediment dating. Radiation Measurements 37, 505-510.

Erfurt, G., Krbetschek, M.R., Bortolot, V.J., Preusser, F., 2003. A fully
automated multi-spectral radioluminescence reading system for geochronometry
and dosimetry. Nuclear Instruments and Methods in Physics Research Section
B: Beam Interactions with Materials and Atoms 207, 487-499.

Lapp, T., Jain, M., Thomsen, K.J., Murray, A.S., Buylaert, J.P., 2012. New
luminescence measurement facilities in retrospective dosimetry. Radiation
Measurements 47, 803-808. doi:10.1016/j.radmeas.2012.02.006

Trautmann, T., 2000. A study of radioluminescence kinetics of natural
feldspar dosimeters: experiments and simulations. Journal of Physics D:
Applied Physics 33, 2304-2310.

Trautmann, T., Krbetschek, M.R., Dietrich, A., Stolz, W., 1998.
Investigations of feldspar radioluminescence: potential for a new dating
technique. Radiation Measurements 29, 421-425.

Trautmann, T., Krbetschek, M.R., Dietrich, A., Stolz, W., 1999. Feldspar
radioluminescence: a new dating method and its physical background. Journal
of Luminescence 85, 45-58.

Trautmann, T., Krbetschek, M.R., Stolz, W., 2000. A systematic study of the
radioluminescence properties of single feldspar grains. Radiation
Measurements 32, 685-690.
}
\seealso{
\code{\linkS4class{RLum.Analysis}},
\code{\linkS4class{RLum.Results}}, \code{\link{get_RLum}},
\code{\link{nls}}
}
\keyword{datagen}

