% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_SAR.TL.R
\name{analyse_SAR.TL}
\alias{analyse_SAR.TL}
\title{Analyse SAR TL measurements}
\usage{
analyse_SAR.TL(
  object,
  object.background,
  signal_integral,
  integral_input = "channel",
  sequence.structure = c("PREHEAT", "SIGNAL", "BACKGROUND"),
  rejection.criteria = list(recycling.ratio = 10, recuperation.rate = 10),
  dose.points = NULL,
  log = "",
  ...
)
}
\arguments{
\item{object}{\link[=RLum.Analysis-class]{RLum.Analysis} or a \link{list} of such objects (\strong{required}) :
input object containing data for analysis}

\item{object.background}{currently not used}

\item{signal_integral}{\link{integer} (\strong{required}):
vector of inputs (as defined by \code{integral_input}) for the signal integral.}

\item{integral_input}{\link{character} (\emph{with default}):
input type for \code{signal_integral}, one of \code{"channel"} (default) or
\code{"measurement"}. If set to \code{"measurement"}, the best matching channels
corresponding to the given temperature range are selected.}

\item{sequence.structure}{\link{vector} \link{character} (\emph{with default}):
specifies the general sequence structure. Three steps are allowed
(\code{"PREHEAT"}, \code{"SIGNAL"}, \code{"BACKGROUND"}), in addition a
parameter \code{"EXCLUDE"}. This allows excluding TL curves which are not
relevant for the protocol analysis.  (\strong{Note:} No TL are removed by default)}

\item{rejection.criteria}{\link{list} (\emph{with default}):
list containing rejection criteria in percentage for the calculation.}

\item{dose.points}{\link{numeric} (\emph{optional}):
option set dose points manually}

\item{log}{\link{character} (\emph{with default}):
a character string which contains \code{"x"} if the x-axis is to be logarithmic,
\code{"y"} if the y axis is to be logarithmic and \code{"xy"} or \code{"yx"} if both axes
are to be logarithmic. See
\link{plot.default}).}

\item{...}{further arguments that will be passed to the function \link{fit_DoseResponseCurve}}
}
\value{
A plot (\emph{optional}) and an \link[=RLum.Results-class]{RLum.Results} object is
returned containing the following elements:

\item{De.values}{\link{data.frame} containing De-values and further parameters}
\item{LnLxTnTx.values}{\link{data.frame} of all calculated \code{Lx/Tx} values including signal, background counts and the dose points.}
\item{rejection.criteria}{\link{data.frame} with values that might by used as rejection criteria. \code{NA} is produced if no R0 dose point exists.}

The output should be accessed using the function \link{get_RLum}.
}
\description{
The function performs a SAR TL analysis on a
\link[=RLum.Analysis-class]{RLum.Analysis} object including growth curve fitting.
}
\details{
This function performs a SAR TL analysis on a set of curves. The SAR
procedure in general is given by Murray and Wintle (2000). For the
calculation of the \code{Lx/Tx} value the function \link{calc_TLLxTxRatio} is
used.

\strong{Provided rejection criteria}

\verb{[recycling.ratio]}: calculated for every repeated regeneration dose point.

\verb{[recuperation.rate]}: recuperation rate calculated by
comparing the \code{Lx/Tx} values of the zero regeneration point with the \code{Ln/Tn}
value (the \code{Lx/Tx} ratio of the natural signal).  For methodological
background see Aitken and Smith (1988)
}
\note{
\strong{THIS IS A BETA VERSION}

No TL curves will be removed from the input object without further warning.
}
\section{Function version}{
 0.3.5
}

\examples{

##load data
data(ExampleData.BINfileData, envir = environment())

##transform the values from the first position in a RLum.Analysis object
object <- Risoe.BINfileData2RLum.Analysis(TL.SAR.Data, pos=3)

##perform analysis
analyse_SAR.TL(
 object = object,
 signal_integral = 210:220,
 fit.method = "SSE OR LIN",
 sequence.structure = c("SIGNAL", "BACKGROUND"))

} 

\section{How to cite}{
Kreutzer, S., 2026. analyse_SAR.TL(): Analyse SAR TL measurements. Function version 0.3.5. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\references{
Aitken, M.J. and Smith, B.W., 1988. Optical dating: recuperation
after bleaching.  Quaternary Science Reviews 7, 387-393.

Murray, A.S. and Wintle, A.G., 2000. Luminescence dating of quartz using an
improved single-aliquot regenerative-dose protocol. Radiation Measurements
32, 57-73.
}
\seealso{
\link{calc_TLLxTxRatio}, \link{fit_DoseResponseCurve}, \link[=RLum.Analysis-class]{RLum.Analysis},
\link[=RLum.Results-class]{RLum.Results}, \link{get_RLum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team}
\keyword{datagen}
\keyword{plot}
