% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_ThermochronometryData.R
\name{analyse_ThermochronometryData}
\alias{analyse_ThermochronometryData}
\title{Analyse Thermochronometry Data}
\usage{
analyse_ThermochronometryData(
  object,
  ITL_model = c("GOK", "BTS"),
  plot = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{\link{character} (\strong{required}): path to a CSV file ##TODO;
alternatively a \link{vector} of paths}

\item{ITL_model}{\link{character} (\emph{with default}): type of model to fit,
either \code{"GOK"} (default) or \code{"BTS"}}

\item{plot}{\link{logical} (\emph{with default}): enable/disable the plot output.}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable output to the
terminal.}

\item{...}{further parameters passed to \link{fit_IsothermalHolding}}
}
\value{
An \linkS4class{RLum.Results} object is returned: ##TODO
}
\description{
Analyse thermochronometry data based on MatLab code written
by XXX and XXX and XXX
}
\details{
##TODO
}
\section{Function version}{
 0.1.0
}

\examples{
# example code ##TODO

}
\seealso{
\link{analyse_FadingMeasurement}, \link{fit_IsothermalHolding}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany),
Svenja Riedesel, DTU Risø (Denmark)
}
\keyword{datagen}
