% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_RLum.R
\name{app_RLum}
\alias{app_RLum}
\title{Run Luminescence shiny apps (wrapper)}
\usage{
app_RLum(app = NULL, ...)
}
\arguments{
\item{app}{\code{\link{character}} (required): name of the application to start. See details for a list
of available apps.}

\item{...}{further arguments to pass to \code{\link[shiny]{runApp}}}
}
\description{
Wrapper for the function \code{\link[RLumShiny]{app_RLum}} from the package
\link[RLumShiny]{RLumShiny-package}. For further details and examples please
see the manual of this package.
}
\section{Function version}{
 0.1.1 (2017-01-24 21:10:47)
}

\author{
Christoph Burow, University of Cologne (Germany)
\cr R Luminescence Package Team} 

\section{How to cite}{
Burow, C. (2017). app_RLum(): Run Luminescence shiny apps (wrapper). Function version 0.1.1. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.7.3. https://CRAN.R-project.org/package=Luminescence
}

