% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.Analysis-class.R,
%   R/RLum.Data.Curve-class.R, R/RLum.Data.Image-class.R,
%   R/RLum.Data.Spectrum-class.R, R/RLum.Results-class.R
\name{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\title{as() - RLum-object coercion}
\arguments{
\item{from}{\linkS4class{RLum} or \link{list}, \link{data.frame}, \link{matrix}
(\strong{required}): object to be coerced from}

\item{to}{\link{character} (\strong{required}): class name to be coerced to}
}
\description{
for \code{[RLum.Analysis]}

for \code{[RLum.Data.Curve]}

for \code{[RLum.Data.Image]}

for \code{[RLum.Data.Spectrum]}

for \code{[RLum.Results]}
}
\details{
\strong{\link{RLum.Analysis}}\cr

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{list} \tab \code{list}\cr
}

Given that the \link{list} consits of \linkS4class{RLum.Analysis} objects.

\strong{\link{RLum.Data.Curve}}\cr

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{list} \tab \code{list} \cr
\code{data.frame} \tab \code{data.frame}\cr
\code{matrix} \tab \code{matrix}

}

\strong{\link{RLum.Data.Image}}\cr

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{data.frame} \tab \code{data.frame}\cr
\code{matrix} \tab \code{matrix}

}

\strong{\link{RLum.Data.Spectrum}}\cr

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{data.frame} \tab \code{data.frame}\cr
\code{matrix} \tab \code{matrix}

}

\strong{\link{RLum.Results}}\cr

\tabular{ll}{
\strong{from} \tab \strong{to}\cr
\code{list} \tab \code{list}\cr
}

Given that the \link{list} consits of \linkS4class{RLum.Results} objects.
}
\note{
Due to the complex structure of the \code{RLum} objects itself a coercing to standard
R data structures will be always loosely!
}
\seealso{
\link[methods:as]{methods::as}
}
