% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.Analysis-class.R,
%   R/RLum.Data.Curve-class.R, R/RLum.Data.Image-class.R,
%   R/RLum.Data.Spectrum-class.R, R/RLum.Results-class.R
\name{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\alias{as}
\title{as() - RLum-object coercion}
\arguments{
\item{from}{[RLum-class] or [list], [data.frame], [matrix]
(\bold{required}): object to be coerced from}

\item{to}{[character] (\bold{required}): class name to be coerced to}
}
\description{
for \code{[RLum.Analysis]}

for \code{[RLum.Data.Curve]}

for \code{[RLum.Data.Image]}

for \code{[RLum.Data.Spectrum]}

for \code{[RLum.Results]}
}
\details{
\bold{[RLum.Analysis]}\cr

\tabular{ll}{
 \bold{from} \tab \bold{to}\cr
  \code{list} \tab \code{list}\cr
}

Given that the [list] consits of [RLum.Analysis-class] objects.

\bold{[RLum.Data.Curve]}\cr

\tabular{ll}{
 \bold{from} \tab \bold{to}\cr
  \code{list} \tab \code{list} \cr
  \code{data.frame} \tab \code{data.frame}\cr
  \code{matrix} \tab \code{matrix}

}

\bold{[RLum.Data.Image]}\cr

\tabular{ll}{
 \bold{from} \tab \bold{to}\cr
  \code{data.frame} \tab \code{data.frame}\cr
  \code{matrix} \tab \code{matrix}

}

\bold{[RLum.Data.Spectrum]}\cr

\tabular{ll}{
 \bold{from} \tab \bold{to}\cr
  \code{data.frame} \tab \code{data.frame}\cr
  \code{matrix} \tab \code{matrix}

}

\bold{[RLum.Results]}\cr

\tabular{ll}{
 \bold{from} \tab \bold{to}\cr
  \code{list} \tab \code{list}\cr
}

Given that the [list] consits of [RLum.Results-class] objects.
}
\note{
Due to the complex structure of the \code{RLum} objects itself a coercing to standard
R data structures will be always loosely!
}
\seealso{
[methods::as]
}
