% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_RLum.Data.R
\name{bin_RLum.Data}
\alias{bin_RLum.Data}
\title{General accessor function for RLum.Data S4 class objects}
\usage{
bin_RLum.Data(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{RLum.Data}} (\bold{required}): S4 object of
class \code{RLum.Data}}

\item{...}{further arguments passed to the specifc class method}
}
\value{
Return is the same as input
}
\description{
Function calls object-specific bin functions for RLum.Data S4 class objects.
}
\details{
The function provides a generalised access point for specific
\code{\linkS4class{RLum.Data}} objects.\cr Depending on the input object, the
corresponding function will be selected. Allowed arguments can be found
in the documentations of the corresponding \code{\linkS4class{RLum.Data}} class.
}
\section{Function version}{
 0.1.0 (2016-04-24 17:21:35)
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
\cr R Luminescence Package Team}
\seealso{
\code{\linkS4class{RLum.Data}}
}
\keyword{utilities}

