% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EED_Model.R
\name{calc_EED_Model}
\alias{calc_EED_Model}
\title{Modelling Exponential Exposure Distribution}
\usage{
calc_EED_Model(data, D0 = 120L, expected_dose = 12L, kappa = NULL,
  sigma_distr = NULL, n.simul = 5000L, n.minSimExp = 50L,
  sample_name = "", method_control = list(), verbose = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{data}{\link{data.frame} (\strong{required}): input data consisting of two columns, the De and the
SE(De). Values are expected in Gy}

\item{D0}{\link{integer} (\emph{with default}): D0 value, defining the characteristation behaviour
of the the quartz. Value are expected in Gy}

\item{expected_dose}{\link{integer} (): TODO}

\item{kappa}{\link{numeric} (\emph{optional}): positive dimensionless exposure parameter
characterising the bleaching state of the grains. Low values (< 10) indicate
poor bleaching}

\item{sigma_distr}{\link{numeric} (\emph{optional}): positive dose rate parameter, representing the
dose variability to which the grains were exposed ##TODO perhaps it should be renamed}

\item{n.simul}{\link{integer} (\emph{with default}): number of simulations}

\item{n.minSimExp}{\link{integer} (\emph{with default}): number of MC runs for calculating the uncertainty
contribution from the sampling}

\item{sample_name}{\link{character} (\emph{with default}): name of the sample}

\item{method_control}{\link{list} (\emph{with default}): additional deep control parameters, parameters
need to be provided as names list, see details}

\item{verbose}{\link{logical} (with default): enable/disable verbose mode}

\item{plot}{\link{logical} (with default): enable/disable plot output}

\item{...}{further parameters that can be passed to better control the plot output. Support arguments
are \code{xlab}, \code{xlim}.}
}
\description{
Modelling incomplete and heterogeneous bleaching of mobile grains partially
exposed to the light, an implementation of the EED model proposed by Guibert et al. (2019)
}
\details{
The function is an implementation and enhancement of the scripts used for
the work by Guibert et al. (2019).

\strong{Method control parameters}

\tabular{llll}{
\strong{ARGUMENT} \tab \strong{FUNCTION} \tab \strong{DEFAULT} \tab \strong{DESCRIPTION}\cr
\code{lower}  \tab  \link[DEoptim:DEoptim]{DEoptim::DEoptim} \tab \code{c(0,0)} \tab set lower bounds for kappa and sigma auto
parameter finding \cr
\code{upper} \tab  \link[DEoptim:DEoptim]{DEoptim::DEoptim} \tab \code{c(1000,2)} \tab set upper bounds for kapp and sigma
auto parameter finding \cr
\code{itermax} \tab \link[DEoptim:DEoptim.control]{DEoptim::DEoptim.control} \tab \code{500} \tab maximum number for iterations to
find the parameters \cr
\code{VTR} \tab \link[DEoptim:DEoptim.control]{DEoptim::DEoptim.control} \tab 1e-05 \tab stop value to reached for the
optimisation \cr
\code{trace} \tab \link[DEoptim:DEoptim.control]{DEoptim::DEoptim.control} \tab \code{FALSE} \tab enable/disable travce mode \cr
\code{trave_plot} \tab - \tab \code{FALSE} \tab enable/disable additional trace plot output \cr

}
}
\section{Function version}{
 0.1.0
}

\examples{

##TODO

} 

\section{How to cite}{
Guibert, P., Kreutzer, S., 2019. calc_EED_Model(): Modelling Exponential Exposure Distribution. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., 2019. Luminescence: Comprehensive Luminescence Dating Data AnalysisR package version 0.9.1.9000-58. https://CRAN.R-project.org/package=Luminescence
}

\references{
Guibert, P., Christophe, C., Urbanova, P., Guérin, G., Blain, S., 2017.
Modeling incomplete and heterogeneous bleaching of mobile grains partially exposed to the
light - Towards a new tool for single grain OSL dating of poorly bleached mortars.
Radiation Measurements 107, 48–57. \doi{10.1016/j.radmeas.2017.10.003}
}
\seealso{
\link[DEoptim:DEoptim]{DEoptim::DEoptim}, \linkS4class{RLum.Results}, \link{calc_MinDose}, \link{calc_FuchsLang2001}, \link{calc_IEU},
\link{calc_FiniteMixture}
}
\author{
Pierre Guibert, IRAMAT-CRP2A, UMR 5060, Université Bordeaux Montaigne (France),
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, Université Bordeaux Montaigne (France)
, RLum Developer Team}
\keyword{datagen}
