% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_FastRatio.R
\name{calc_FastRatio}
\alias{calc_FastRatio}
\title{Calculate the Fast Ratio for CW-OSL curves}
\usage{
calc_FastRatio(object, stimulation.power = 30.6, wavelength = 470,
  sigmaF = 2.6e-17, sigmaM = 4.28e-18, Ch_L1 = 1, x = 1, x2 = 0.1,
  dead.channels = c(0, 0), fitCW.sigma = FALSE, fitCW.curve = FALSE,
  plot = TRUE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{RLum.Analysis}}, 
\code{\linkS4class{RLum.Data.Curve}} or \code{\link{data.frame}} 
(\bold{required}):}

\item{stimulation.power}{\code{\link{numeric}}:}

\item{wavelength}{\code{\link{numeric}}:}

\item{sigmaF}{\code{\link{numeric}}:}

\item{sigmaM}{\code{\link{numeric}}:}

\item{Ch_L1}{\code{\link{numeric}}:}

\item{x}{\code{\link{numeric}}:}

\item{x2}{\code{\link{numeric}}:}

\item{dead.channels}{\code{\link{numeric}}: Vector of length 2 in the form of
\code{c(x, y)}.}

\item{fitCW.sigma}{\code{\link{logical}}: fit CW-OSL curve using \code{\link{fit_CWCurve}}
to calculate \code{sigmaF} and \code{sigmaM} and apply the fast ratio on the
non-linear least-square fit (experimental).}

\item{fitCW.curve}{\code{\link{logical}}: fit CW-OSL curve using \code{\link{fit_CWCurve}}
and derive the counts of L2 and L3 from the fitted OSL curve (experimental).}

\item{plot}{\code{\link{numeric}}: plot output (\code{TRUE}/\code{FALSE})}

\item{...}{available options: \code{verbose} (\code{\link{logical}}). Further
arguments passed to \code{\link{fit_CWCurve}}.}
}
\value{
Returns a plot (optional) and an S4 object of type \code{\linkS4class{RLum.Results}}. 
The slot \code{data} contains a \code{\link{list}} with the following elements:\cr
.. $ summary \cr
.. $ data \cr
.. $ fit \cr
.. $ args \cr
.. $ call \cr
}
\description{
This function calculates the fast ratio of CW-OSL curves after
Durcan & Duller (2001).
}
\details{
Some more details on the function.
}
\section{Function version}{
 0.1.0 (2016-03-13 20:00:10)
}
\examples{
# load example CW-OSL curve
data("ExampleData.CW_OSL_Curve")

# calculate the fast ratio w/o further adjustments
res <- calc_FastRatio(ExampleData.CW_OSL_Curve)

# show the summary table
get_RLum(res)

}
\author{
Georgina King, University of Cologne (Germany) \cr
Julie A. Durcan, Institute of Geography and Earth Sciences, Aberystwyth University (United Kingdom) \cr
Christoph Burow, University of Cologne (Germany) \cr
\cr R Luminescence Package Team}
\references{
Durcan, J.A. & Duller, G.A.T., 2011. The fast ratio: A rapid measure for testing
the dominance of the fast component in the initial OSL signal from quartz.
Radiation Measurements 46, 1065-1072. \cr
}

