% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_Crosstalk.R
\name{calc_MoransI}
\alias{calc_MoransI}
\title{Calculate Moran's I}
\usage{
calc_MoransI(
  object,
  df_neighbour = get_Neighbour(object = object),
  bo_return_inbetween_numbers = FALSE
)
}
\arguments{
\item{object}{\linkS4class{RLum.Results} or \link{numeric} (\strong{required}) containing
the values of the grains of one. Should have length 100; can contain \code{NA}
values.}

\item{df_neighbour}{\link{data.frame} (\emph{with default}), indicating which
neighbours to consider, and their respective weights. Defaults to
\code{get_Neighbour(object)}.}

\item{bo_return_inbetween_numbers}{\link{logical} (\emph{with default}). Defaults
to \code{FALSE}. If \code{TRUE}, the function returns a list with several
intermediate calculation results.}
}
\value{
By default one numerical value, roughly between -1 and 1, where
close to zero means no spatial correlation, and value close to 1 a positive
spatial correlation given the pattern we interested in (by default all rook neighbours).
A value closer to -1 has no meaning within the context of luminescence crosstalk.
If \code{bo_return_inbetween_numbers} is set to \code{TRUE}, a list with several numbers
used for calculation is returned instead of the single outcome.
}
\description{
Calculate Moran's I
}
\examples{

## Test a fictional sample with spatial correlation
calc_MoransI(object = c(1:100))

## Test some fictional samples without spatial correlation;
## note the randomness with each repetition
calc_MoransI(object = rnorm(n = 100))
calc_MoransI(object = rnorm(n = 100))
calc_MoransI(object = rnorm(n = 100))

}
\author{
Anna-Maartje de Boer, Luc Steinbuch, Wageningen University & Research, 2025
}
