% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Activity2Concentration.R
\name{convert_Activity2Concentration}
\alias{convert_Activity2Concentration}
\title{Convert Nuclide Activities to Concentrations and Vice Versa}
\usage{
convert_Activity2Concentration(data, input_unit = "Bq/kg", verbose = TRUE)
}
\arguments{
\item{data}{\code{\link{data.frame}} \bold{(required)}: provide dose rate data (activity or concentration)
in three columns. The first column indicates the nuclides, the 2nd column measured value and
in the 3rd column its error value. Allowed nuclide data are \code{'U-238'}, \code{'Th-232'} and \code{'K-40'}.
See examples for an example.}

\item{input_unit}{\code{\link{character}} (with default): specify unit of input data given in the
dose rate data frame, choose between \code{'Bq/kg'} and \code{'ppm/\%'} the default is \code{'Bq/kg'}}

\item{verbose}{\code{\link{logical}} (with default): enable or disable verbose mode}
}
\description{
Convert Nuclide Activities to Concentrations and Vice Versa
}
\section{Function version}{
 0.1.0 (2017-05-30 18:19:19)
}

\examples{

##construct data.frame
data <- data.frame(
 NUCLIDES = c("U-238", "Th-232", "K-40"),
 VALUE = c(40,80,100),
 VALUE_ERROR = c(4,8,10),
 stringsAsFactors = FALSE)

##perform analysis
convert_Activity2Concentration(data)

}

\section{How to cite}{
Fuchs, M.C. (2017). convert_Activity2Concentration(): Convert Nuclide Activities to Concentrations and Vice Versa. Function version 0.1.0. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence
}

\references{
-
}
\author{
Margret C. Fuchs, Helmholtz-Institut Freiberg for Resource Technology (Germany)
\cr R Luminescence Package Team}
\keyword{IO}
