% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_PSL2CSV.R
\name{convert_PSL2CSV}
\alias{convert_PSL2CSV}
\title{Export PSL-file(s) to CSV-files}
\usage{
convert_PSL2CSV(file, ...)
}
\arguments{
\item{file}{\code{\link{character}} (\bold{required}): name of the PSL-file to be converted to CSV-files}

\item{\dots}{further arguments that will be passed to the function \code{\link{read_PSL2R}} and \code{\link{write_RLum2CSV}}}
}
\value{
The function returns either a CSV-file (or many of them) or for the option \code{export = FALSE}
a list comprising objects of type \code{link{data.frame}} and \code{\link{matrix}}
}
\description{
This function is a wrapper function around the functions \code{\link{read_PSL2R}} and
\code{\link{write_RLum2CSV}} and it imports an PSL-file (SUERC portable OSL reader file format)
and directly exports its content to CSV-files.
If nothing is set for the argument \code{path} (\code{\link{write_RLum2CSV}}) the input folder will
become the output folder.
}
\section{Function version}{
 0.1.0 (2017-01-24 21:10:47)
}

\examples{

\dontrun{
##select your BIN-file
file <- file.choose()

##convert
convert_PSL2CSV(file)

}

}
\seealso{
\code{\linkS4class{RLum.Analysis}}, \code{\linkS4class{RLum.Data}}, \code{\linkS4class{RLum.Results}},
\code{\link[utils]{write.table}}, \code{\link{write_RLum2CSV}}, \code{\link{read_PSL2R}}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team} 

\section{How to cite}{
Kreutzer, S. (2017). convert_PSL2CSV(): Export PSL-file(s) to CSV-files. Function version 0.1.0. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.7.3. https://CRAN.R-project.org/package=Luminescence
}

\keyword{IO}
