% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_SG2MG.R
\name{convert_SG2MG}
\alias{convert_SG2MG}
\title{Converts Single-Grain Data to Multiple-Grain Data}
\usage{
convert_SG2MG(object, write_file = FALSE, ...)
}
\arguments{
\item{object}{\linkS4class{Risoe.BINfileData}, \link{character} (\strong{required}):
\linkS4class{Risoe.BINfileData} object or BIN/BINX-file name.}

\item{write_file}{\link{logical} (\emph{with default}):
whether the output should be written to a file (only considered if \code{object}
is of type character). The multiple grain file will be written to a file
of named after the original one with \code{-SG} appended to it.}

\item{...}{further arguments passed down to \link{read_BIN2R} and
\link{write_BIN2R} if \code{object} is a path to a file.}
}
\value{
A \linkS4class{Risoe.BINfileData} object. If \code{write_file = TRUE} an
the input was a file path, a file is also written out to origin folder.
}
\description{
Conversion of single-grain data to multiple-grain data by adding signals
from grains belonging to one disc (unique pairs of position, set and run).
}
\section{Function version}{
 0.1.1
}

\examples{
## simple run
## (please not that the example is not using SG data)
data(ExampleData.BINfileData, envir = environment())
convert_SG2MG(CWOSL.SAR.Data)

}
\seealso{
\linkS4class{Risoe.BINfileData}, \link{read_BIN2R},
\link{write_R2BIN}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Norbert Mercier, IRAMAT-CRP2A, UMR 5060, CNRS-Université Bordeaux Montaigne (France)
}
\keyword{IO}
