% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Wavelength2Energy.R
\name{convert_Wavelength2Energy}
\alias{convert_Wavelength2Energy}
\title{Emission Spectra Conversion from Wavelength to Energy Scales}
\usage{
convert_Wavelength2Energy(object)
}
\arguments{
\item{object}{\linkS4class{RLum.Data.Spectrum}, \link{data.frame}, \link{matrix} (\strong{required}): input object to be converted.
If the input is not an \linkS4class{RLum.Data.Spectrum}, the first column is always treated as the wavelength
column. The function supports a list of allowed input objects.}
}
\value{
The same object class as provided as input is returned.
}
\description{
The function provides a convienent and fast way to convert a wavelength
to an energy scale. The function works on \linkS4class{RLum.Data.Spectrum}, \link{data.frame} and \link{matrix} and
a \link{list} of such objects. The function was written to smooth the workflow while using emission
spectra data. This is in particular useful if you want to further treat your data and apply, e.g.,
a signal deconvolution.
}
\details{
The intensity of the spectrum is re-calcualted using the
following approach (cf. Appendix 4 in Blasse and Grabmeier, 1994):

\deqn{\phi_{E} = \phi_{\lambda} * \lambda^2 / (hc)}

with
\eqn{\phi_{E}} the intensity per interval of energy \eqn{E} (1/eV),
\eqn{\phi_{\lambda}} the intensity per interval of wavelength \eqn{\lambda}
(1/nm) and
\eqn{h} (eV * s) the Planck constant and \eqn{c} (nm/s) the velocity of light.

For transforming the wavelength axis (x-values) the equation as follow is used

\deqn{E = hc/\lambda}
}
\note{
This conversion works solely for emission spectra.
}
\section{Function version}{
 0.1.0 (2019-01-22 20:23:23)
}
 

\section{How to cite}{
Kreutzer, S. (2019). convert_Wavelength2Energy(): Emission Spectra Conversion from Wavelength to Energy Scales. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2019). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0.21. https://CRAN.R-project.org/package=Luminescence
}

\references{
Blasse, G., Grabmaier, B.C., 1994. Luminescent Materials. Springer.

Mooney, J., Kambhampati, P., 2013. Get the Basics Right: Jacobian Conversion of Wavelength and
Energy Scales for Quantitative Analysis of Emission Spectra. J. Phys. Chem. Lett. 4, 3316–3318.
\doi{10.1021/jz401508t}
}
\seealso{
\linkS4class{RLum.Data.Spectrum}, \link{plot_RLum}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\keyword{IO}
