% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\docType{data}
\name{ExampleData.CW_OSL_Curve}
\alias{ExampleData.CW_OSL_Curve}
\alias{CW_Curve.BosWallinga2012}
\title{Example CW-OSL curve data for the package Luminescence}
\format{
Data frame with 1000 observations on the following 2 variables:
\itemize{
\item \code{list("x")}: a numeric vector, time
\item \code{list("y")}: a numeric vector, counts
}
}
\source{
\strong{ExampleData.CW_OSL_Curve}

\tabular{ll}{
Lab: \tab Luminescence Laboratory Bayreuth\cr
Lab-Code: \tab BT607\cr
Location: \tab Saxony/Germany\cr
Material: \tab Middle grain quartz measured on aluminium cups on a Risø TL/OSL DA-15 reader.\cr
Reference: \tab unpublished data }

\strong{CW_Curve.BosWallinga2012}

\tabular{ll}{
Lab: \tab Netherlands Centre for Luminescence Dating (NCL)\cr
Lab-Code: \tab NCL-2108077\cr
Location: \tab Guadalentin Basin, Spain\cr
Material: \tab Coarse grain quartz\cr
Reference: \tab Bos & Wallinga (2012) and Baartman et al. (2011)
}
}
\description{
\code{data.frame} containing CW-OSL curve data (time, counts)
}
\examples{

data(ExampleData.CW_OSL_Curve, envir = environment())
plot(ExampleData.CW_OSL_Curve)

}
\references{
Baartman, J.E.M., Veldkamp, A., Schoorl, J.M., Wallinga, J.,
Cammeraat, L.H., 2011. Unravelling Late Pleistocene and Holocene landscape
dynamics: The Upper Guadalentin Basin, SE Spain. Geomorphology, 125,
172-185.

Bos, A.J.J. & Wallinga, J., 2012. How to visualize quartz OSL signal
components. Radiation Measurements, 47, 752-758.
}
\keyword{datasets}
