% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_EmissionSpectra.R
\name{fit_EmissionSpectra}
\alias{fit_EmissionSpectra}
\title{Emission Spectra Deconvoluation}
\usage{
fit_EmissionSpectra(object, verbose = TRUE, plot = TRUE, ...)
}
\arguments{
\item{object}{\linkS4class{RLum.Data.Spectrum}, \link{data.frame}, \link{matrix} (\strong{required}): input
object. Please note that an energy spectrum is expected}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable verbose mode}

\item{plot}{\link{logical} (\emph{with default}): enable/disable plot output}

\item{...}{further arguments to be passed to control the plot output (e.g., \code{main}, \code{xlab}, \code{ylab})}
}
\value{
Returns an \linkS4class{RLum.Results} object ##TODO
}
\description{
Luminescence spectra signal deconvolution on \linkS4class{RLum.Data.Spectrum} objects
}
\details{
##TODO add details

##TODO add tests
}
\note{
Beta version, not uet recommended for productive usage
}
\section{Function version}{
 0.1.0 (2019-01-24 16:22:58)
}

\examples{

##deconvolution of a TL spectrum
##TODO should be modified ... the bg substraction is odd, also the object conversion
\dontrun{

##load example data
data(ExampleData.XSYG, envir = environment())

##subtract background
m <- TL.Spectrum@data
m <- m - m[, 15]
m <- cbind(as.numeric(rownames(m)), m[,5])
m <- convert_Wavelength2Energy(m, order = TRUE)
m[m[,2] < 0 ,2] <- 0
results <- fit_EmissionSpectra(m, main = "TL spectrum")

}

} 

\section{How to cite}{
Kreutzer, S. (2019). fit_EmissionSpectra(): Emission Spectra Deconvoluation. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2019). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0.34. https://CRAN.R-project.org/package=Luminescence
}

\references{
##TODO
}
\seealso{
\linkS4class{RLum.Data.Spectrum}, \linkS4class{RLum.Results}, \link{plot_RLum}, \link{convert_Wavelength2Energy}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\keyword{datagen}
