% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_IsothermalHolding.R
\name{fit_IsothermalHolding}
\alias{fit_IsothermalHolding}
\title{Fit Isothermal Holding Data}
\usage{
fit_IsothermalHolding(data, ITL_model = "GOK", rhop, plot = TRUE, ...)
}
\arguments{
\item{data}{\link{character} or \link{list} (\strong{required}): input data with a table or file path}

\item{ITL_model}{\link{character} (\emph{with default}): ITL data to be fitted}

\item{rhop}{\link{numeric} or \linkS4class{RLum.Results} (\emph{with default}): rhop prime values (one for each sample) or
\linkS4class{RLum.Results} object produced by \link{analyse_FadingMeasurement}}

\item{plot}{\link{logical} (\emph{with default}): enable/disable plot}

\item{...}{further parameters to be passed to the plot output}
}
\value{
An \linkS4class{RLum.Results} object is returned: ##TODO
}
\description{
##TODO
}
\details{
##TODO
}
\section{Function version}{
 0.1.0
}

\examples{
# example code ##TOD


}
\seealso{
\link{analyse_ThermochronometryData}, \link{analyse_FadingMeasurement}
}
\author{
Svenja Riedesel, DTU Risø (Denmark)
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{datagen}
