% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_IsothermalHolding.R
\name{fit_IsothermalHolding}
\alias{fit_IsothermalHolding}
\title{Fit Isothermal Holding Data}
\usage{
fit_IsothermalHolding(
  data,
  rhop,
  ITL_model = "GOK",
  plot = TRUE,
  verbose = TRUE,
  txtProgressBar = TRUE,
  trace = FALSE,
  ...
)
}
\arguments{
\item{data}{\link{character} or \link{data.frame} (\strong{required}): file path or data
frame with 5 columns named "SAMPLE", "TEMP", "TIME", "LxTx", "LxTx_ERROR".}

\item{rhop}{\link{numeric} or \linkS4class{RLum.Results} (\emph{with default}):
a vector of rho prime values (one for each sample) or an \linkS4class{RLum.Results}
object produced by \link{analyse_FadingMeasurement}.}

\item{ITL_model}{\link{character} (\emph{with default}): model to be fitted, either
\code{"GOK"} or \code{"BTS"}.}

\item{plot}{\link{logical} (\emph{with default}): enable/disable the plot output.}

\item{verbose}{\link{logical} (\emph{with default}): enable/disable output to the
terminal.}

\item{txtProgressBar}{\link{logical} (\emph{with default}): enable/disable the
progress bar. Ignored if \code{verbose = FALSE}.}

\item{trace}{\link{logical} (\emph{with default}): enable/disable tracing during
the nls fitting (\link[minpack.lm:nlsLM]{minpack.lm::nlsLM}).}

\item{...}{further arguments and graphical parameters that will be passed
to the \code{plot} function.}
}
\value{
The function returns an \linkS4class{RLum.Results} object and an \emph{optional} plot.
The object returned contains the following elements:

\item{fit}{\link{list} with the fitted models}
\item{coefs}{\link{data.frame} containing the fitted coefficients for the models}
\item{data}{\link{data.frame} containing the data used in the fitting process}

It may return \code{NULL} if no model could be fitted.
}
\description{
##TODO
}
\details{
##TODO
}
\section{Function version}{
 0.1.0
}

\examples{
# example code ##TODO

}
\references{
Li, B., Li, S.H., The effect of band-tail states on the
thermal stability of the infrared stimulated luminescence
from K-feldspar, Journal of Luminescence 136 (2013) 5–10.
doi: 10.1016/j.jlumin.2012.08.043
}
\seealso{
\link{analyse_ThermochronometryData}, \link{analyse_FadingMeasurement}
}
\author{
Svenja Riedesel, DTU Risø (Denmark)\cr
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)
}
\keyword{datagen}
