% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_OSLLifeTimes.R
\name{fit_OSLLifeTimes}
\alias{fit_OSLLifeTimes}
\title{Fitting OSL Lifetime Components}
\usage{
fit_OSLLifeTimes(object, signal_range = NULL, n.components = NULL,
  method_control = list(), plot = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{object}{\linkS4class{RLum.Data.Curve}, \link{data.frame} or \link{matrix} \strong{(required)}:
Input object containing the data to be analysed. All objects can be provided also as list for an automated
processing}

\item{signal_range}{\link{numeric} (\emph{optional}): allows to set a channel range, by default all channels are used, e.g.
\code{signal_range = c(2,100)} considers only channels 2 to 100 and \code{signal_range = c(2)} considers only channels
from channel 2 onwards.}

\item{n.components}{\link{numeric} (\emph{optional}): Fix the number of components. If set the algorithm will try
to fit the number of predefined components. If nothing is set, the algorithm will try to find the best number
of components.}

\item{method_control}{\link{list} (\emph{optonal}): Named to allow a more fine control of the fitting process. See details
for allowed options.}

\item{plot}{\link{logical} (\emph{with default}): Enable/disable plot output}

\item{verbose}{\link{logical} (\emph{with default}): Enable/disable terminal feedback}

\item{...}{parameters passed to \link{plot.default} to control the plot output. Please note that not
all parameters are supported.}
}
\value{
-----------------------------------\cr
\code{[ NUMERICAL OUTPUT ]}\cr
-----------------------------------\cr

\strong{\code{RLum.Results}}-object

\strong{slot:} \strong{\code{@data}}

\tabular{lll}{
\strong{Element} \tab \strong{Type} \tab \strong{Description}\cr
\code{$data} \tab \code{matrix} \tab the final fit matrix \cr
\code{$start_matrix} \tab \code{matrix} \tab the start matrix used for the fitting \cr
\code{$fit} \tab \code{nls} \tab the fit object returned by \link[minpack.lm:nls.lm]{minpack.lm::nls.lm} \cr
}

\strong{slot:} \strong{\code{@info}}

The original function call

------------------------\cr
\code{[ PLOT OUTPUT ]}\cr
------------------------\cr

A plot showing the original data and the fit so far possible
}
\description{
Fitting OSL Lifetime Components
}
\details{
The fitting is optimised to fit the off-time flank of POSL measurements.

\strong{method_control} options

TODO
}
\section{Function version}{
 0.1.0 (2018-06-19 14:06:03)
}

\examples{

print("nothing here so far")

} 

\section{How to cite}{
Kreutzer, S., Schmidt, C. (2018). fit_OSLLifeTimes(): Fitting OSL Lifetime Components. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0. https://CRAN.R-project.org/package=Luminescence
}

\references{
Bluszcz, A., Adamiec, G., 2006. Application of differential evolution to fitting OSL decay curves.
Radiation Measurements 41, 886-891. doi:10.1016/j.radmeas.2006.05.016\cr

\strong{Further reading}

Storn, R., Price, K., 1997. Differential Evolution – A Simple and Efficient Heuristic for Global Optimization over Continuous Spaces.
Journal of Global Optimization 11, 341–359.
}
\seealso{
\link[minpack.lm:nls.lm]{minpack.lm::nls.lm}, \link[DEoptim:DEoptim]{DEoptim::DEoptim}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS-Université Bordeaux Montaigne (France),
Christoph Schmidt, University of Bayreuth (Germany)
\cr R Luminescence Package Team}
