% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SurfaceExposure.R
\name{fit_SurfaceExposure}
\alias{fit_SurfaceExposure}
\title{Nonlinear Least Squares Fit for OSL surface exposure data}
\usage{
fit_SurfaceExposure(data, sigmaphi = NULL, mu = NULL, age = NULL,
  Ddot = NULL, D0 = NULL, weights = FALSE, plot = TRUE, legend = TRUE,
  error_bars = TRUE, coord_flip = FALSE, ...)
}
\arguments{
\item{data}{\link{data.frame} or \link{list} (\strong{required}):}

\item{sigmaphi}{\link{numeric} (\emph{optional}):}

\item{mu}{\link{numeric} (\emph{optional}):}

\item{age}{\link{numeric} (\emph{optional}):}

\item{Ddot}{\link{numeric} (\emph{optional}):}

\item{D0}{\link{numeric} (\emph{optional}):}

\item{weights}{\link{logical} (\emph{optional}):}

\item{plot}{\link{logical} (\emph{optional}):}

\item{legend}{\link{logical} (\emph{optional}):}

\item{error_bars}{\link{logical} (\emph{optional}):}

\item{coord_flip}{\link{logical} (\emph{optional}):}

\item{...}{Further parameters passed to \link{plot}.
Custom parameters include:
\itemize{
\item \code{verbose} (\link{logical}): show or hide console output
\item \code{line_col}: Color of the fitted line
\item \code{line_lty}: Type of the fitted line (see \code{lty} in \code{?par})
\item \code{line_lwd}: Line width of the fitted line (see \code{lwd} in \code{?par})
}}
}
\value{
Function returns results numerically and graphically:

-----------------------------------\cr
\code{[ NUMERICAL OUTPUT ]}\cr
-----------------------------------\cr

\strong{\code{RLum.Results}}-object

\strong{slot:} \strong{\code{@data}}

\tabular{lll}{
\strong{Element} \tab \strong{Type} \tab \strong{Description}\cr
\code{$summary} \tab \code{data.frame} \tab summary of the fitting results \cr
\code{$data} \tab \code{data.frame} \tab the original input data \cr
\code{$fit} \tab \code{nls} \tab the fitting object produced by \link[minpack.lm:nlsLM]{minpack.lm::nlsLM} \cr
\code{$args} \tab \code{character} \tab arguments of the call \cr
\code{$call} \tab \code{call} \tab the original function call \cr
}

\strong{slot:} \strong{\code{@info}}

Currently unused.

------------------------\cr
\code{[ PLOT OUTPUT ]}\cr
------------------------\cr

A scatter plot of the provided depth-intensity OSL surface exposure data
with the fitted model.
}
\description{
This function determines the (weighted) least-squares estimates of the
parameters of either eq. 1 in \emph{Sohbati et al. (2012a)} or eq. 12 in
\emph{Sohbati et al. (2012b)} for a given OSL surface exposure data set.
}
\details{
\strong{Dose rate}

TODO

\strong{Global fitting}

TODO
}
\section{Function version}{
 0.1.0
}

\examples{
# none yet

}
\references{
Sohbati, R., Murray, A.S., Chapot, M.S., Jain, M., Pederson, J., 2012a.
Optically stimulated luminescence (OSL) as a chronometer for surface exposure
dating. Journal of Geophysical Research 117, B09202. doi:10.1029/2012JB009383

Sohbati, R., Jain, M., Murray, A.S., 2012b. Surface exposure dating of
non-terrestial bodies using optically stimulated luminescence: A new method.
Icarus 221, 160-166.
}
\seealso{
\link[minpack.lm:nlsLM]{minpack.lm::nlsLM}
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\keyword{datagen}
