% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ThermalQuenching.R
\name{fit_ThermalQuenching}
\alias{fit_ThermalQuenching}
\title{Fitting Thermal Quenching Data}
\usage{
fit_ThermalQuenching(data, start_param = list(),
  method_control = list(), n.MC = 100, verbose = TRUE, plot = TRUE,
  ...)
}
\arguments{
\item{data}{\link{data.frame} (\strong{required}): input data with three columns, the first column contains
temperature values in deg. C, colmns 2 and 3 the dependent values with its error}

\item{start_param}{\link{list} (optional): option to provide own start parameters for the fitting, see
detalis}

\item{method_control}{\link{list} (optianl): further options to fine tune the fitting, see details for
further information}

\item{n.MC}{\link{numeric} (\emph{with default}): number of Monte Carlo runs for the error estimation}

\item{verbose}{\link{logical} (\emph{with default}): enables/disables terminal output}

\item{plot}{\link{logical} (\emph{with default}): enables/disables plot output}

\item{...}{further arguments that can be passed to control the plotting, support are \code{main}, \code{pch},
\code{col_fit}, \code{col_points}, \code{lty}, \code{lwd}, \code{xlab}, \code{ylab}, \code{xlim}, \code{ylim}}
}
\value{
##TODO
}
\description{
Applying a nls-fitting to thermal quenching data
}
\section{Function version}{
 0.1.0 (2018-10-05 19:52:50)
}

\examples{

##create short example dataset
data <- data.frame(
  T = c(25, 40, 50, 60, 70, 80, 90, 100, 110),
  V = c(0.06, 0.058, 0.052, 0.051, 0.041, 0.034, 0.035, 0.033, 0.032),
  V_X = c(0.012, 0.009, 0.008, 0.008, 0.007, 0.006, 0.005, 0.005, 0.004))

##fit
fit_ThermalQuenching(
 data = data,
 n.MC = NULL)

} 

\section{How to cite}{
Kreutzer, S. (2018). fit_ThermalQuenching(): Fitting Thermal Quenching Data. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0. https://CRAN.R-project.org/package=Luminescence
}

\references{
Wintle, A.G., 1975. Thermal Quenching of Thermoluminescence in Quartz. Geophys. J. R. astr. Soc. 41, 107–113.
}
\seealso{
\link[minpack.lm:nlsLM]{minpack.lm::nlsLM}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR5060, CNRS - Université Bordeaux Montaigne (Frange)
\cr R Luminescence Package Team}
