% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_Quote.R
\name{get_Quote}
\alias{get_Quote}
\title{Function to return essential quotes}
\usage{
get_Quote(ID, author)
}
\arguments{
\item{ID}{\code{\link{character}}, qoute ID to be returned.}

\item{author}{\code{\link{character}}, all quotes by specified author.}
}
\value{
Returns a character with quote and respective (false) author.
}
\description{
This function returns one of the collected essential quotes in the
growing library. If called without any parameters, a random quote is
returned.
}
\section{Function version}{
 0.1.1 (2015-07-23 17:09:29)
}
\examples{
## ask for an arbitrary qoute
get_Quote()
}
\author{
Michael Dietze, GFZ Potsdam (Germany)
\cr R Luminescence Package Team}

