% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Quote.R
\name{get_Quote}
\alias{get_Quote}
\title{Function to return essential quotes}
\usage{
get_Quote(ID, author, separated = FALSE)
}
\arguments{
\item{ID}{\code{\link{character}}, qoute ID to be returned.}

\item{author}{\code{\link{character}}, all quotes by specified author.}

\item{separated}{\code{\link{logical}}, return result in separated form.}
}
\value{
Returns a character with quote and respective (false) author.
}
\description{
This function returns one of the collected essential quotes in the
growing library. If called without any parameters, a random quote is
returned.
}
\section{Function version}{
 0.1.1 (2017-01-24 21:10:47)
}

\examples{

## ask for an arbitrary qoute
get_Quote()

}
\author{
Michael Dietze, GFZ Potsdam (Germany)
\cr R Luminescence Package Team} 

\section{How to cite}{
Dietze, M. (2017). get_Quote(): Function to return essential quotes. Function version 0.1.1. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.7.0. https://CRAN.R-project.org/package=Luminescence
}

