% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Quote.R
\name{get_Quote}
\alias{get_Quote}
\title{Function to return essential quotes}
\usage{
get_Quote(ID, separated = FALSE)
}
\arguments{
\item{ID}{\link{character} (\emph{optional}): qoute ID to be returned.}

\item{separated}{\link{logical} (\emph{with default}): return result in separated form.}
}
\value{
Returns a character with quote and respective (false) author.
}
\description{
This function returns one of the collected essential quotes in the
growing library. If called without any parameters, a random quote is
returned.
}
\section{Function version}{
 0.1.2 (2018-01-21 17:22:38)
}

\examples{

## ask for an arbitrary qoute
get_Quote()

}
\author{
Michael Dietze, GFZ Potsdam (Germany)
\cr R Luminescence Package Team} 

\section{How to cite}{
Dietze, M. (2018). get_Quote(): Function to return essential quotes. Function version 0.1.2. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.4. https://CRAN.R-project.org/package=Luminescence
}

