% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Quote.R
\name{get_Quote}
\alias{get_Quote}
\title{Function to return essential quotes}
\usage{
get_Quote(ID, separated = FALSE)
}
\arguments{
\item{ID}{\link{character} (\emph{optional}): quote ID to be returned.}

\item{separated}{\link{logical} (\emph{with default}): return result in separated form.}
}
\value{
Returns a character with quote and respective (false) author.
}
\description{
This function returns one of the collected essential quotes in the
growing library. If called without any parameters, a random quote is
returned.
}
\section{Function version}{
 0.1.5
}

\examples{

## ask for an arbitrary quote
get_Quote()

}
\author{
Quote credits: Michael Dietze, GFZ Potsdam (Germany), Sebastian Kreutzer, Geography & Earth Science, Aberystwyth University (United Kingdom), Dirk Mittelstraß, TU Dresden (Germany), Jakob Wallinga (Wageningen University, Netherlands)
, RLum Developer Team} 

\section{How to cite}{
Dietze, M., Kreutzer, S., 2023. get_Quote(): Function to return essential quotes. Function version 0.1.5. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2023. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.23. https://CRAN.R-project.org/package=Luminescence
}

