% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_rightAnswer.R
\name{get_rightAnswer}
\alias{get_rightAnswer}
\title{Function to get the right answer}
\usage{
get_rightAnswer(...)
}
\arguments{
\item{...}{you can pass an infinite number of further arguments}
}
\value{
Returns the right answer
}
\description{
This function returns the right answer
}
\section{Function version}{
 0.1.0 (2015-09-19 16:10:25)
}
\examples{
## you really want to know?
get_rightAnswer()
}
\author{
NA
\cr R Luminescence Package Team}

