% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rightAnswer.R
\name{get_rightAnswer}
\alias{get_rightAnswer}
\title{Function to get the right answer}
\usage{
get_rightAnswer(...)
}
\arguments{
\item{...}{you can pass an infinite number of further arguments}
}
\value{
Returns the right answer
}
\description{
This function returns just the right answer
}
\section{Function version}{
 0.1.0 (2015-11-29 17:27:48)
}
\examples{

## you really want to know?
get_rightAnswer()

}
\author{
inspired by R.G.
\cr R Luminescence Package Team}

