\name{get_Risoe.BINfileData}
\alias{get_Risoe.BINfileData}
\title{get Risoe BINfileData}

\usage{get_Risoe.BINfileData(object)}
\arguments{
  \item{object}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





