% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R
\name{get_Risoe.BINfileData}
\alias{get_Risoe.BINfileData}
\title{General accessor function for Risoe.BINfileData objects}
\usage{
get_Risoe.BINfileData(object, ...)
}
\arguments{
\item{object}{\linkS4class{Risoe.BINfileData} (\strong{required}):
S4 object of class \code{Risoe.BINfileData}}

\item{...}{further arguments passed to the specific class method}
}
\value{
An object of the same type as the input object provided.
}
\description{
The function provides a generalised access point for specific
\linkS4class{Risoe.BINfileData} objects. Depending on the input object, the
corresponding function will be selected.
Allowed arguments can be found in the documentations of the corresponding
\linkS4class{Risoe.BINfileData} class.
}
\section{Function version}{
 0.1.0
}

\seealso{
\linkS4class{Risoe.BINfileData}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2025. get_Risoe.BINfileData(): General accessor function for Risoe.BINfileData objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., Boer, A.d., 2025. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.0.1. https://r-lum.github.io/Luminescence/
}

\keyword{internal}
\keyword{utilities}
