% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Risoe.BINfileData.R
\name{get_Risoe.BINfileData}
\alias{get_Risoe.BINfileData}
\title{General accessor function for RLum S4 class objects}
\usage{
get_Risoe.BINfileData(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{Risoe.BINfileData}} (\bold{required}): S4 object of
class \code{RLum}}

\item{\dots}{further arguments that one might want to pass to the specific
get function}
}
\value{
Return is the same as input objects as provided in the list.
}
\description{
Function calls object-specific get functions for RisoeBINfileData S4 class objects.
}
\details{
The function provides a generalised access point for specific
\code{\linkS4class{Risoe.BINfileData}} objects.\cr Depending on the input object, the
corresponding get function will be selected. Allowed arguments can be found
in the documentations of the corresponding \code{\linkS4class{Risoe.BINfileData}} class.
}
\section{Function version}{
 0.1.0 (2017-06-29 18:40:14)
}

\seealso{
\code{\linkS4class{Risoe.BINfileData}}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
\cr R Luminescence Package Team} 

\section{How to cite}{
Kreutzer, S. (2017). get_Risoe.BINfileData(): General accessor function for RLum S4 class objects. Function version 0.1.0. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.7.5. https://CRAN.R-project.org/package=Luminescence
}

\keyword{utilities}
