\name{Luminescence-package}
\alias{Luminescence-package}
\alias{Luminescence}
\docType{package}
\title{Collection of functions for Luminescence data analysis at the Luminescence Laboratories
Giessen, Bayreuth, Cologne and Freiberg (Germany)}
\description{
This package provides various functions developed for the purpose
of Luminescence Dating data analysis. 
}
\details{
\tabular{ll}{
Package: \tab Luminescence\cr
Type: \tab Package\cr
Version: \tab 0.1.7\cr
Date: \tab 2012-05-24\cr
License: \tab GPL-3\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Authors: Sebastian Kreutzer, Christoph Schmidt, Margret C. Fuchs, Michael Dietze, Manfred Fischer, 
Markus Fuchs \cr

Maintainer: Sebastian Kreutzer, JLU Giessen, Germany, \cr
\email{sebastian.kreutzer@geogr.uni-giessen.de}
}

\keyword{package}
