\name{Luminescence-package}
\alias{Luminescence-package}
\alias{Luminescence}
\docType{package}
\title{Collection of functions for luminescence data analysis at the luminescence laboratories
Giessen, Bayreuth, Cologne and Freiberg (Germany)}
\description{
This package provides various functions developed for the purpose
of Luminescence Dating data analysis. 
}
\details{
\tabular{ll}{
Package: \tab Luminescence\cr
Type: \tab Package\cr
Version: \tab 0.1.8\cr
Date: \tab 2012-08-12\cr
License: \tab GPL-3\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Authors: Sebastian Kreutzer, Christoph Schmidt, Margret C. Fuchs, Michael Dietze, Manfred Fischer, Christoph 
Burow, Markus Fuchs \cr

Maintainer: Sebastian Kreutzer, JLU Giessen, Germany, \cr
\email{sebastian.kreutzer@geogr.uni-giessen.de}
}

\references{
Kreutzer, S., Schmidt, C., Fuchs, M.C., Dietze, M., Fischer, M., Fuchs, M. (2012): Introducing an R package for luminescence dating analysis. Ancient TL, 30, 1-8.
}

\keyword{package}
