\name{Luminescence-package}
\alias{Luminescence-package}
\alias{Luminescence}
\docType{package}
\title{Collection of functions for luminescence dating data analysis}
\description{
This package provides various functions developed for the purpose
of Luminescence Dating data analysis. 
}
\details{
\tabular{ll}{
Package: \tab Luminescence\cr
Type: \tab Package\cr
Version: \tab 0.3.2\cr
Date: \tab 2014-01-07\cr
License: \tab GPL-3\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\bold{Authors} 

\tabular{ll}{ 
Christoph Burow \tab University of Cologne, Germany \cr
Michael Dietze \tab GFZ Helmholtz Centre Potsdam, Germany \cr
Manfred Fischer\tab University of Bayreuth, Germany \cr 
Margret C. Fuchs \tab TU Bergakademie Freiberg, Germany\cr
Sebastian Kreutzer \tab Justus-Liebig-University Giessen, Germany\cr
Christoph Schmidt \tab University of Bayreuth, Germany
}

\bold{Supervisor}

Markus Fuchs, Justus-Liebig-University Giessen, Germany

\bold{Support contact}

\url{http://forum.r-luminescence.de}\cr
\email{team@r-luminescence.de} \cr

\bold{Bug reporting}\cr

\email{bugtracker@r-luminescence.de} \cr

\bold{Project website}

\url{http://www.r-luminescence.de}

\bold{Package maintainer}

Sebastian Kreutzer, Justus-Liebig-University Giessen, Germany, \cr
\email{sebastian.kreutzer@geogr.uni-giessen.de}
}

\references{
Dietze, M., Kreutzer, S., Fuchs, M.C., Burow, C., Fischer, M., Schmidt, C., 2013. A practical guide to the R package Luminescence.  Ancient TL, 31, pp. 11-18.

Kreutzer, S., Schmidt, C., Fuchs, M.C., Dietze, M., Fischer, M., Fuchs, M., 2012. Introducing an R package for luminescence dating analysis. Ancient TL, 30, pp. 1-8.
}

\keyword{package}
