\name{Luminescence-package}
\alias{Luminescence-package}
\alias{Luminescence}
\docType{package}
\title{Collection of functions for luminescence dating data analysis}
\description{
This package provides various functions developed for the purpose
of Luminescence Dating data analysis. 
}
\details{
\tabular{ll}{
Package: \tab Luminescence\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2012-12-15\cr
License: \tab GPL-3\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\bold{Authors} 

\tabular{ll}{
Christoph Burow \tab University of Cologne \cr
Michael Dietze \tab TU Dresden \cr
Manfred Fischer\tab University of Bayreuth \cr 
Margret C. Fuchs \tab TU Bergakademie Freiberg\cr
Sebastian Kreutzer \tab JLU Giessen\cr
Christoph Schmidt \tab University of Bayreuth
}

\bold{Supervisor}

Markus Fuchs, JLU Giessen, Germany

\bold{Support contact}

\email{team@r-luminescence.de}

\bold{Package maintainer}

Sebastian Kreutzer, JLU Giessen, Germany, \cr
\email{sebastian.kreutzer@geogr.uni-giessen.de}
}

\references{
Kreutzer, S., Schmidt, C., Fuchs, M.C., Dietze, M., Fischer, M. & Fuchs, M., 2012. Introducing an R package for luminescence dating analysis. Ancient TL, 30, pp. 1-8.
}

\keyword{package}
