% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R
\name{melt_RLum}
\alias{melt_RLum}
\alias{melt_RLum,list-method}
\title{Melt RLum-class objects into a flat data.frame}
\usage{
melt_RLum(object, ...)

\S4method{melt_RLum}{list}(object, ...)
}
\arguments{
\item{object}{\linkS4class{RLum} (\strong{required}):
S4 object of class \code{RLum}}

\item{...}{further arguments passed to the specific class method}
}
\value{
A flat \link{data.frame}.
}
\description{
Melt RLum-class objects into a flat data.frame
}
\section{Functions}{
\itemize{
\item \code{melt_RLum(list)}: Returns a list a single \link{data.frame}

}}
\section{Function version}{
 0.1.0
}

\examples{
data(ExampleData.XSYG, envir = environment())
melt_RLum(OSL.SARMeasurement[[2]][[1]])

}
\seealso{
\linkS4class{RLum.Data.Curve}, \linkS4class{RLum.Analysis}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{utilities}
