% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_RLum.Analysis.R
\name{merge_RLum.Analysis}
\alias{merge_RLum.Analysis}
\title{Merge function for RLum.Analysis S4 class objects}
\usage{
merge_RLum.Analysis(object, ...)
}
\arguments{
\item{object}{\link{list} of \link[=RLum.Analysis-class]{RLum.Analysis} (\strong{required}):
list of S4 objects of class \code{RLum.Analysis}. Furthermore other objects of
class \link[=RLum-class]{RLum} can be added, see details.}

\item{...}{currently not used.}
}
\value{
Returns an \link[=RLum.Analysis-class]{RLum.Analysis} object ordered according to the order
provided with the input list.
}
\description{
This function simply allows to merge \link[=RLum.Analysis-class]{RLum.Analysis}
objects. Moreover, other \link[=RLum.Data-class]{RLum.Data} objects can be added
to an existing \link[=RLum.Analysis-class]{RLum.Analysis} object. Supported objects
to be added are: \link[=RLum.Data.Curve-class]{RLum.Data.Curve},
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} and
\link[=RLum.Data.Image-class]{RLum.Data.Image}.
}
\note{
The information for the slot 'protocol' is taken from the first
\link[=RLum.Analysis-class]{RLum.Analysis} object in the input list. Therefore at
least one object of type \link[=RLum.Analysis-class]{RLum.Analysis} has to be provided.
}
\section{Function version}{
 0.2.3
}

\examples{

##merge different RLum objects from the example data
data(ExampleData.RLum.Analysis, envir = environment())
data(ExampleData.BINfileData, envir = environment())

object <- Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos=1)
curve <- get_RLum(object)[[2]]

temp.merged <- merge_RLum.Analysis(list(curve, IRSAR.RF.Data, IRSAR.RF.Data))

}
\seealso{
\link{merge_RLum}, \link[=RLum.Analysis-class]{RLum.Analysis},
\link[=RLum.Data.Curve-class]{RLum.Data.Curve},
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum},
\link[=RLum.Data.Image-class]{RLum.Data.Image}, \link[=RLum-class]{RLum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{internal}
\keyword{utilities}
