% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_RLum.Data.Curve.R,
%   R/merge_RLum.Data.Spectrum.R
\name{merge_RLum.Data.Curve}
\alias{merge_RLum.Data.Curve}
\alias{merge_RLum.Data.Spectrum}
\title{Merge RLum.Data objects}
\usage{
merge_RLum.Data.Curve(
  object,
  merge.method = c("mean", "median", "sum", "sd", "var", "max", "min", "append", "-",
    "*", "/"),
  method.info = NULL,
  ...
)

merge_RLum.Data.Spectrum(
  object,
  merge.method = c("mean", "median", "sum", "sd", "var", "min", "max", "append", "-",
    "*", "/"),
  method.info = NULL,
  max.temp.diff = 0.1,
  ...
)
}
\arguments{
\item{object}{\link{list} of \link[=RLum.Data.Curve-class]{RLum.Data.Curve} or \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} (\strong{required}):
list of objects to be merged.}

\item{merge.method}{\link{character} (\strong{required}):
method for combining of the objects, e.g. \code{'mean'} (default), \code{'median'},
\code{'sum'}, see details for
further information and allowed methods.}

\item{method.info}{\link{numeric} (\emph{optional}):
allows to specify how info elements of the input objects are combined,
e.g. \code{1} means that just the elements from the first object are kept,
\code{2} keeps only the info elements from the 2 object etc.
If set to \code{NULL} (default), all elements are combined.}

\item{...}{currently not used.}

\item{max.temp.diff}{\link{numeric} (\emph{with default}):
maximum difference in the time/temperature values between the spectra to
be merged: when differences exceed this threshold value, the merging
occurs but a warning is raised.
Only used for \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects.}
}
\value{
Returns an \link[=RLum.Data.Curve-class]{RLum.Data.Curve} or
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} object, depending on the input.
}
\description{
This function allows to merge \link[=RLum.Data.Curve-class]{RLum.Data.Curve} and
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects in different ways without
modifying the original objects. For \code{RLum.Data.Curve} objects, merging is
always applied on the 2nd column of the object's data matrix.

\strong{Supported merge operations are:}

\code{"mean"} (default)

The mean over the count/cell values is calculated using \link{rowMeans}.

\code{"median"}

The median over the count/cell values is calculated using
\link[matrixStats:rowMedians]{matrixStats::rowMedians}.

\code{"sum"}

All count/cell values will be summed up using \link{rowSums}.

\code{"sd"}

The standard deviation over the count/cell values is calculated using
\link[matrixStats:rowSds]{matrixStats::rowSds}.

\code{"var"}

The variance over the count/cell values is calculated using
\link[matrixStats:rowVars]{matrixStats::rowVars}.

\code{"min"}

The min values from the count/cell values is calculated using
\link[matrixStats:rowMins]{matrixStats::rowMins}.

\code{"max"}

The max values from the count/cell values is calculated using
\link[matrixStats:rowMins]{matrixStats::rowMins}.

\code{"append"}

Appends count/cell values of all objects to one combined data object. The
channel width is automatically re-calculated, but requires a constant channel
width of the original data. Note: for \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}
objects, this method is only available when all objects have the same number
of columns.

\code{"-"}

The row sums of the last objects are subtracted from the first object.

\code{"*"}

The row sums of the last objects are multiplied with the first object.

\code{"/"}

Values of the first object are divided by row sums of the last objects.
}
\note{
The information from the slot \code{recordType} is taken from the first
object in the input list. The slot
'curveType' is filled with the name \code{merged}.
}
\section{S3-generic support}{


These functions are fully operational via S3-generics:
\code{+}, \code{-}, \code{/}, \code{*}, \code{merge}
}

\section{Function version}{
 0.2.3
}

\examples{

## load example data
data(ExampleData.XSYG, envir = environment())

## grep first and third TL curves
TL.curves  <- get_RLum(OSL.SARMeasurement$Sequence.Object, recordType = "TL (UVVIS)")
TL.curve.1 <- TL.curves[[1]]
TL.curve.3 <- TL.curves[[3]]

## plot single curves
plot_RLum(TL.curve.1)
plot_RLum(TL.curve.3)

## subtract the 1st curve from the 3rd and plot
TL.curve.merged <- merge_RLum.Data.Curve(list(TL.curve.3, TL.curve.1),
                                         merge.method = "/")
plot_RLum(TL.curve.merged)

}
\seealso{
\link{merge_RLum},
\link[=RLum.Data.Curve-class]{RLum.Data.Curve},
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
}
\keyword{internal}
\keyword{utilities}
