% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/merge_RLum.Analysis.R, R/merge_RLum.Data.Curve.R,
%   R/merge_RLum.Data.Spectrum.R, R/merge_RLum.R, R/merge_RLum.Results.R
\name{merge_RLum}
\alias{merge_RLum}
\alias{merge_RLum.Analysis}
\alias{merge_RLum.Data.Curve}
\alias{merge_RLum.Data.Spectrum}
\alias{merge_RLum,list-method}
\alias{merge_RLum.Results}
\title{Merge RLum-class objects}
\usage{
merge_RLum(object, ...)

merge_RLum.Analysis(object, ...)

merge_RLum.Data.Curve(
  object,
  merge.method = c("mean", "median", "sum", "sd", "var", "max", "min", "append", "-",
    "*", "/"),
  method.info = NULL,
  ...
)

merge_RLum.Data.Spectrum(
  object,
  merge.method = c("mean", "median", "sum", "sd", "var", "min", "max", "append", "-",
    "*", "/"),
  method.info = NULL,
  max.temp.diff = 0.1,
  ...
)

\S4method{merge_RLum}{list}(object, ...)

merge_RLum.Results(object, flatten = TRUE, ...)
}
\arguments{
\item{object}{\link{list} (\strong{required}):
list of \link[=RLum-class]{RLum} objects to be merged. All elements must
be of the same type, unless at least one \link[=RLum.Analysis-class]{RLum.Analysis}
object is present, in which case also \link[=RLum.Data-class]{RLum.Data} can
be provided.}

\item{...}{further arguments passed to the specific class methods.}

\item{merge.method}{\link{character} (\emph{with default}):
method for combining of the objects, e.g. \code{'mean'} (default), \code{'median'},
\code{'sum'}, see below for further information.
Only used for \link[=RLum.Data.Curve-class]{RLum.Data.Curve} and
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects.}

\item{method.info}{\link{numeric} (\emph{optional}):
allows to specify how info elements of the input objects are combined,
e.g. \code{1} keeps only the info elements from the first object, \code{2} keeps only
those from the second object, etc. If set to \code{NULL} (default), all elements
are combined.
Only used for \link[=RLum.Data.Curve-class]{RLum.Data.Curve} and
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects.}

\item{max.temp.diff}{\link{numeric} (\emph{with default}):
maximum difference in the time/temperature values between the spectra to
be merged: when differences exceed this threshold value, the merging
occurs but a warning is raised.
Only used for \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects.}

\item{flatten}{\link{logical} (\emph{with default}):
whether list elements should be flattened before merging.
Only used for \link[=RLum.Results-class]{RLum.Results} objects.}
}
\value{
Returns an object of the same class as the input elements.
}
\description{
These functions merge \link[=RLum-class]{RLum} objects of the same class
without modifying the original objects. Empty list elements (\code{NULL}) are
automatically removed from the input list. The dispatch is based on the
class of the input objects:
\itemize{
\item \link[=RLum.Analysis-class]{RLum.Analysis}: when at least one input is an
\code{RLum.Analysis} object, a single \code{RLum.Analysis} object is returned
containing the supplied \link[=RLum.Data-class]{RLum.Data} objects and
\code{RLum.Analysis} records, in the order given.
\item \link[=RLum.Data.Curve-class]{RLum.Data.Curve} and
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}: merging is done element-wise
on the data values using one of several supported methods (see below).
\item \link[=RLum.Results-class]{RLum.Results}: data elements are appended or
combined depending on their type.
}
}
\note{
\itemize{
\item For \link[=RLum.Analysis-class]{RLum.Analysis} objects, the information for the
\code{protocol} slot is taken from the first \link[=RLum.Analysis-class]{RLum.Analysis}
object in the input list.
\item For \link[=RLum.Data.Curve-class]{RLum.Data.Curve} and
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects, the information from the
\code{recordType} slot is taken from the first object in the input list. The
\code{curveType} slot is set to \code{"merged"}.
\item For \link[=RLum.Data.Image-class]{RLum.Data.Image}, no merging is supported.
\item For \link[=RLum.Results-class]{RLum.Results} objects, the \code{originator} is taken
from the first element and not reset to \code{"merge_RLum"}.
}
}
\section{Supported merge operations}{


The following values for the \code{merge.method} argument are supported when
merging \link[=RLum.Data.Curve-class]{RLum.Data.Curve} and
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects:
\itemize{
\item \code{"mean"} (default): the mean over the count/cell values is calculated
using \link{rowMeans}.
\item \code{"median"}: the median over the count/cell values is calculated using
\link[matrixStats:rowMedians]{matrixStats::rowMedians}.
\item \code{"sum"}: all count/cell values are summed up using \link{rowSums}.
\item \code{"sd"}: the standard deviation over the count/cell values is calculated
using \link[matrixStats:rowSds]{matrixStats::rowSds}.
\item \code{"var"}: the variance over the count/cell values is calculated using
\link[matrixStats:rowVars]{matrixStats::rowVars}.
\item \code{"min"}: the min values from the count/cell values is calculated using
\link[matrixStats:rowMins]{matrixStats::rowMins}.
\item \code{"max"}: the max values from the count/cell values is calculated using
\link[matrixStats:rowMaxs]{matrixStats::rowMaxs}.
\item \code{"append"}: appends count/cell values of all objects to one combined data
object. The channel width is automatically re-calculated, but requires a
constant channel width of the original data.
\strong{Note:} For \link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum} objects, this method
is only available when all objects have the same number of columns.
\item \code{"-"}: the row sums of the last objects are subtracted from the first
object.
\item \code{"*"}: the row sums of the last objects are multiplied by the first
object.
\item \code{"/"}: values of the first object are divided by row sums of the last
objects.
}
}

\section{S3 generic support}{


Merging of \code{RLum.Data.Curve} and \code{RLum.Data.Spectrum} objects can also be
performed via S3-generics: \code{+}, \code{-}, \code{/}, \code{*}.
}

\section{Function version}{
 0.3.0
}

\examples{

## load example data
data(ExampleData.XSYG, envir = environment())
data(ExampleData.DeValues, envir = environment())
data(ExampleData.RLum.Analysis, envir = environment())

## extract the first and third TL curves
TL.curves  <- get_RLum(OSL.SARMeasurement$Sequence.Object,
                       recordType = "TL (UVVIS)")

## ---- RLum.Analysis ----
merged <- merge_RLum(list(TL.curves[[1]], IRSAR.RF.Data, IRSAR.RF.Data))

## ---- RLum.Data.Curve ----
## subtract the 1st curve from the 3rd
TL.merged <- merge_RLum(list(TL.curves[[3]], TL.curves[[1]]),
                        merge.method = "-")

## ---- RLum.Data.Spectrum ----
## sum two copies of the same spectrum
spectrum.merged <- merge_RLum(list(TL.Spectrum, TL.Spectrum),
                              merge.method = "sum")

## ---- RLum.Results ----
res <- calc_CentralDose(ExampleData.DeValues$CA1)
res.merged <- merge_RLum(list(res, res))

}
\seealso{
\link[=RLum.Analysis-class]{RLum.Analysis},
\link[=RLum.Data.Curve-class]{RLum.Data.Curve},
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum},
\link[=RLum.Results-class]{RLum.Results}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
}
\keyword{utilities}
