% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_RLum.Results.R
\name{merge_RLum.Results}
\alias{merge_RLum.Results}
\title{Merge function for RLum.Results S4-class objects}
\usage{
merge_RLum.Results(object, flatten = TRUE)
}
\arguments{
\item{object}{\link{list} (\strong{required}):
a list of \link[=RLum.Results-class]{RLum.Results} objects}

\item{flatten}{\link{logical} (\emph{with default}):
whether list elements should be flattened before merging.}
}
\description{
Function merges objects of class \link[=RLum.Results-class]{RLum.Results}. The slots in the objects
are combined depending on the object type, e.g., for \link{data.frame} and \link{matrix}
rows are appended.
}
\details{
Elements are appended where possible and attributes are preserved if
not of similar name as the default attributes of, e.g., a \link{data.frame}
}
\note{
The \code{originator} is taken from the first element and not reset to \link{merge_RLum}
}
\section{Function version}{
 0.3.1
}

\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{internal}
\keyword{utilities}
