\name{merge_RLum.Results}
\alias{merge_RLum.Results}
\title{merge RLum Results}

\usage{merge_RLum.Results(object.list)}
\arguments{
  \item{object.list}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





