% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_RLum.R
\name{merge_RLum.Results}
\alias{merge_RLum.Results}
\title{General accessor function for RLum S4 class objects}
\usage{
merge_RLum.Results(object.list)
}
\arguments{
\item{object.list}{\code{\linkS4class{RLum}} (\bold{required}): a list of S4
objects of class \code{RLum}}
}
\value{
Return is the same as input objects as provided in the list.
}
\description{
Function calls object-specific get functions for RLum S4 class objects.
}
\details{
The function provides a generalised access point for specific
\code{\linkS4class{RLum}} objects.\cr Depending on the input object, the
corresponding merge function will be selected. Allowed arguments can be found
in the documentations of the corresponding \code{\linkS4class{RLum}} class.
}
\section{Function version}{
 0.1
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
\cr R Luminescence Package Team}
\seealso{
\code{\linkS4class{RLum.Data.Curve}},
\code{\linkS4class{RLum.Data.Image}},
\code{\linkS4class{RLum.Data.Spectrum}},
\code{\linkS4class{RLum.Analysis}},
\code{\linkS4class{RLum.Results}}
}
\keyword{utilities}

