% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_RLum.Results.R
\name{merge_RLum.Results}
\alias{merge_RLum.Results}
\title{Merge function for RLum.Results S4-class objects}
\usage{
merge_RLum.Results(objects)
}
\arguments{
\item{objects}{\code{\link{list}} (required): a list of \code{\linkS4class{RLum.Results}} objects}
}
\description{
Function merges objects of class \code{\linkS4class{RLum.Results}}. The slots in the objects
are combined depending on the object type, e.g., for \code{\link{data.frame}} and \code{\link{matrix}}
rows are appended.
}
\note{
The originator is taken from the first element and not reset to \code{merge_RLum}
}
\section{Function version}{
 0.2.0 (2016-05-02 09:36:06)
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
\cr R Luminescence Package Team}

