% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_RLum.Results.R
\name{merge_RLum.Results}
\alias{merge_RLum.Results}
\title{Merge function for RLum.Results S4-class objects}
\usage{
merge_RLum.Results(objects)
}
\arguments{
\item{objects}{\link{list} (\strong{required}):
a list of \linkS4class{RLum.Results} objects}
}
\description{
Function merges objects of class \linkS4class{RLum.Results}. The slots in the objects
are combined depending on the object type, e.g., for \link{data.frame} and \link{matrix}
rows are appended.
}
\note{
The originator is taken from the first element and not reset to \code{merge_RLum}
}
\section{Function version}{
 0.2.0
}

\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2019. merge_RLum.Results(): Merge function for RLum.Results S4-class objects. Function version 0.2.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., 2019. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.6. https://CRAN.R-project.org/package=Luminescence
}

\keyword{internal}
