% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_RLum.R
\name{methods_RLum}
\alias{methods_RLum}
\alias{plot.list}
\alias{plot.RLum.Results}
\alias{plot.RLum.Analysis}
\alias{plot.RLum.Data.Curve}
\alias{plot.RLum.Data.Spectrum}
\alias{plot.RLum.Data.Image}
\alias{plot.Risoe.BINfileData}
\alias{subset.Risoe.BINfileData}
\alias{subset.RLum.Analysis}
\alias{bin}
\alias{bin.RLum.Data.Curve}
\alias{bin.RLum.Data.Spectrum}
\alias{length.RLum.Results}
\alias{length.RLum.Analysis}
\alias{length.RLum.Data.Curve}
\alias{length.Risoe.BINfileData}
\alias{rep.RLum}
\alias{names.RLum.Data.Curve}
\alias{names.RLum.Data.Spectrum}
\alias{names.RLum.Data.Image}
\alias{names.RLum.Analysis}
\alias{names.RLum.Results}
\alias{names.Risoe.BINfileData}
\alias{row.names.RLum.Data.Spectrum}
\alias{as.data.frame.RLum.Data}
\alias{as.data.frame.Risoe.BINfileData}
\alias{as.list.RLum}
\alias{as.matrix.RLum.Data}
\alias{merge.RLum}
\alias{unlist.RLum.Analysis}
\alias{+.RLum.Data}
\alias{-.RLum.Data}
\alias{*.RLum.Data}
\alias{/.RLum.Data}
\alias{[.RLum.Data}
\alias{[.RLum.Analysis}
\alias{[.RLum.Results}
\alias{[<-.RLum.Data.Curve}
\alias{[[.RLum.Analysis}
\alias{[[.RLum.Results}
\alias{$.RLum.Data}
\alias{$.RLum.Analysis}
\alias{$.RLum.Results}
\title{Methods for S3-generics implemented for the package 'Luminescence'.}
\usage{
\method{plot}{list}(x, y, ...)

\method{plot}{RLum.Results}(x, y, ...)

\method{plot}{RLum.Analysis}(x, y, ...)

\method{plot}{RLum.Data.Curve}(x, y, ...)

\method{plot}{RLum.Data.Spectrum}(x, y, ...)

\method{plot}{RLum.Data.Image}(x, y, ...)

\method{plot}{Risoe.BINfileData}(x, y, ...)

\method{subset}{Risoe.BINfileData}(x, subset, records.rm = TRUE, ...)

\method{subset}{RLum.Analysis}(x, subset = NULL, ...)

bin(x, ...)

\method{bin}{RLum.Data.Curve}(x, bin_size = 2, ...)

\method{bin}{RLum.Data.Spectrum}(x, bin_size.row = 1, bin_size.col = 1, ...)

\method{length}{RLum.Results}(x, ...)

\method{length}{RLum.Analysis}(x, ...)

\method{length}{RLum.Data.Curve}(x, ...)

\method{length}{Risoe.BINfileData}(x, ...)

\method{rep}{RLum}(x, ...)

\method{names}{RLum.Data.Curve}(x, ...)

\method{names}{RLum.Data.Spectrum}(x, ...)

\method{names}{RLum.Data.Image}(x, ...)

\method{names}{RLum.Analysis}(x, ...)

\method{names}{RLum.Results}(x, ...)

\method{names}{Risoe.BINfileData}(x)

\method{row.names}{RLum.Data.Spectrum}(x, ...)

\method{as.data.frame}{RLum.Data}(x, row.names = NULL, optional = FALSE, ...)

\method{as.data.frame}{Risoe.BINfileData}(x, row.names = NULL, optional = FALSE, ...)

\method{as.list}{RLum}(x, ...)

\method{as.matrix}{RLum.Data}(x, ...)

\method{merge}{RLum}(x, y, ...)

\method{unlist}{RLum.Analysis}(x, recursive = TRUE, ...)

\method{+}{RLum.Data}(x, y)

\method{-}{RLum.Data}(x, y)

\method{*}{RLum.Data}(x, y)

\method{/}{RLum.Data}(x, y)

\method{[}{RLum.Data}(x, y, z, drop = TRUE)

\method{[}{RLum.Analysis}(x, i, drop = FALSE)

\method{[}{RLum.Results}(x, i, drop = TRUE)

\method{[}{RLum.Data.Curve}(x, i, j) <- value

\method{[[}{RLum.Analysis}(x, i)

\method{[[}{RLum.Results}(x, i)

\method{$}{RLum.Data}(x, i)

\method{$}{RLum.Analysis}(x, i)

\method{$}{RLum.Results}(x, i)
}
\arguments{
\item{x}{\link[=RLum-class]{RLum} or \link[=Risoe.BINfileData-class]{Risoe.BINfileData} (\strong{required}):
input object}

\item{y}{\link{integer} (\emph{optional}):
the row index of the matrix, data.frame}

\item{...}{further arguments that can be passed to the method}

\item{subset}{\verb{[subset]} \link{expression} (\strong{required}):
logical expression indicating elements or rows to keep, this function works
in \link[=Risoe.BINfileData-class]{Risoe.BINfileData} objects like \link{subset.data.frame}, but takes care
of the object structure. Works also on \link[=RLum.Analysis-class]{RLum.Analysis} objects.}

\item{records.rm}{\link{subset} \link{logical} (\emph{with default}):
whether records should be removed from the data set. If \code{FALSE}, the \code{SEL}
column marks the selected records.}

\item{row.names}{\link{logical} (\emph{with default}):
enable/disable row names (\code{as.data.frame}).}

\item{optional}{\link{logical} (\emph{with default}):
logical. If TRUE, setting row names and converting column names
(to syntactic names: see make.names) is optional (see \link[base:as.data.frame]{base::as.data.frame})}

\item{recursive}{\link{logical} (\emph{with default}):
enable/disable further sub-setting (\code{unlist}).}

\item{z}{\link{integer} (\emph{optional}):
the column index of the matrix, data.frame}

\item{drop}{\link{logical} (\emph{with default}):
keep object structure or drop it}

\item{i}{\link{character} (\emph{optional}):
name of the wanted record type or data object or row in the
\link[=RLum.Data.Curve-class]{RLum.Data.Curve} object}

\item{j}{\link{integer} (\emph{optional}):
column of the data matrix in the \link[=RLum.Data.Curve-class]{RLum.Data.Curve} object}

\item{value}{\link{numeric} (\strong{required}):
numeric value which replace the value in the \link[=RLum.Data.Curve-class]{RLum.Data.Curve} object}
}
\description{
This document summarises all implemented S3-generics. The name of the function
is given before the first dot, after the dot the name of the object that is
supported by this method is given, e.g. \code{plot.RLum.Data.Curve} can be called
by \code{plot(object, ...)}, where \code{object} is the \code{RLum.Data.Curve} object.
}
\details{
The term S3-generics sounds complicated, however, it just means that something
has been implemented in the package to increase the usability for users new
in R and who are not familiar with the underlying \code{RLum}-object structure of
the package. The practical outcome is that operations and functions presented
in standard books on R can be used without knowing the specifics of the R
package \code{'Luminescence'}. For examples see the example section.
}
\note{
\code{methods_RLum} are not really new functions, everything given here are mostly just
surrogates for existing functions in the package.
}
\examples{

##load example data
data(ExampleData.RLum.Analysis, envir = environment())


## combine curves in various ways
curve1 <- IRSAR.RF.Data[[1]]
curve2 <-  IRSAR.RF.Data[[1]]
curve1 + curve2
curve1 - curve2
curve1 / curve2
curve1 * curve2


##`$` access curves
IRSAR.RF.Data$RF

}
\keyword{internal}
