% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_RLum.R
\name{names_RLum}
\alias{names_RLum}
\title{S4-names function for RLum S4 class objects}
\usage{
names_RLum(object)
}
\arguments{
\item{object}{\linkS4class{RLum} (\strong{required}):
S4 object of class \code{RLum}}
}
\value{
Returns a \link{character}
}
\description{
Function calls object-specific names functions for RLum S4 class objects.
}
\details{
The function provides a generalised access point for specific
\linkS4class{RLum} objects.\cr
Depending on the input object, the corresponding 'names' function will be
selected. Allowed arguments can be found in the documentations of the
corresponding \linkS4class{RLum} class.
}
\section{Function version}{
 0.1.0 (2017-07-30 13:10:59)
}

\seealso{
\linkS4class{RLum.Data.Curve}, \linkS4class{RLum.Data.Image},
\linkS4class{RLum.Data.Spectrum}, \linkS4class{RLum.Analysis}, \linkS4class{RLum.Results}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team} 

\section{How to cite}{
Kreutzer, S. (2017). names_RLum(): S4-names function for RLum S4 class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence
}

\keyword{utilities}
