% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_RLum.R
\name{names_RLum}
\alias{names_RLum}
\title{S4-names function for RLum S4 class objects}
\usage{
names_RLum(object)
}
\arguments{
\item{object}{\code{\linkS4class{RLum}} (\bold{required}): S4 object of
class \code{RLum}}
}
\value{
Returns a \code{\link{character}}
}
\description{
Function calls object-specific names functions for RLum S4 class objects.
}
\details{
The function provides a generalised access point for specific
\code{\linkS4class{RLum}} objects.\cr Depending on the input object, the
corresponding 'names' function will be selected. Allowed arguments can be found
in the documentations of the corresponding \code{\linkS4class{RLum}} class.
}
\section{Function version}{
 0.1.0 (2015-11-29 17:27:48)
}

\seealso{
\code{\linkS4class{RLum.Data.Curve}},
\code{\linkS4class{RLum.Data.Image}},
\code{\linkS4class{RLum.Data.Spectrum}},
\code{\linkS4class{RLum.Analysis}},
\code{\linkS4class{RLum.Results}}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
\cr R Luminescence Package Team} 

\section{How to cite}{
Kreutzer, S. (2017). names_RLum(): S4-names function for RLum S4 class objects. Function version 0.1.0. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.7.0. https://CRAN.R-project.org/package=Luminescence
}

\keyword{utilities}
