% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data.Curve-class.R,
%   R/RLum.Data.Image-class.R, R/RLum.Data.Spectrum-class.R
\name{normalise_RLum}
\alias{normalise_RLum}
\alias{normalise_RLum,list-method}
\alias{normalise_RLum,RLum.Analysis-method}
\alias{normalise_RLum,RLum.Data.Curve-method}
\alias{normalise_RLum,RLum.Data.Image-method}
\alias{normalise_RLum,RLum.Data.Spectrum-method}
\title{Normalise of RLum.Data-class objects}
\usage{
normalise_RLum(object, norm = TRUE, ...)

\S4method{normalise_RLum}{list}(object, norm = TRUE, ...)

\S4method{normalise_RLum}{RLum.Analysis}(object, norm = TRUE, ...)

\S4method{normalise_RLum}{RLum.Data.Curve}(object, norm = TRUE)

\S4method{normalise_RLum}{RLum.Data.Image}(object, norm = TRUE, global = TRUE)

\S4method{normalise_RLum}{RLum.Data.Spectrum}(object, norm = TRUE)
}
\arguments{
\item{object}{\linkS4class{RLum} (\strong{required}):
S4 object of class \code{RLum}}

\item{norm}{\link{logical} \link{character} (\strong{required}):
if logical, whether curve normalisation should occur; alternatively, one
of \code{"max"} (used with \code{TRUE}), \code{"min"}, \code{"last"},  \code{"huot"} or an arbitary
real number (e.g., 2.2)}

\item{...}{further arguments passed to the specific class method}

\item{global}{\link{logical} (*with default): this defines whether the normalisation
is apply globally (same to all) or locally, means, each frame has is
own normalisation. If \code{global = TRUE} the arguments for \code{norm  = 'first'} and
\code{norm = 'last'} work as expected and consider either the first or the last
frame for the normalisation.

Normalise \linkS4class{RLum.Data.Image} objects to value set via
the argument \code{norm}}
}
\value{
An object of the same type as the input object provided.
}
\description{
The function provides a generalised access point for specific
\linkS4class{RLum} objects. Depending on the input object, the corresponding
function will be selected. The normalisation is performed in the internal
function \code{.normalise_curve()}.
}
\details{
The argument \code{norm} normalises all count values. To date the following
options are supported:

\code{norm = TRUE} or \code{norm = "max"}: Curve values are normalised to the highest
count value in the curve

\code{norm = "min"}: Curve values are normalised to the smallest count value
in the curve

\code{norm = 2.2}: Curve values are normalised to 2.2, while this can be any
real number

\code{norm = "first"}: Curve values are normalised to the first count value.

\code{norm = "last"}: Curve values are normalised to the last count value
(this can be useful in particular for radiofluorescence curves)

\code{norm = "huot"}: Curve values are normalised as suggested by Sébastien Huot
via GitHub:
\deqn{
y = (observed - median(background)) / (max(observed) - median(background))
}

The background of the curve is defined as the last 20\% of the count values
of a curve.
}
\section{Functions}{
\itemize{
\item \code{normalise_RLum(list)}: Returns a list of \linkS4class{RLum.Data} objects that had been passed to
\link{normalise_RLum}

\item \code{normalise_RLum(RLum.Analysis)}: Normalisation of \code{RLum.Data} records contained in the input object.

\item \code{normalise_RLum(RLum.Data.Curve)}: Normalise \linkS4class{RLum.Data.Curve} objects to value set via
the argument \code{norm}

\item \code{normalise_RLum(RLum.Data.Image)}: 

\item \code{normalise_RLum(RLum.Data.Spectrum)}: Normalise \linkS4class{RLum.Data.Spectrum} objects to value set via
the argument \code{norm}

}}
\note{
Currently only \code{RLum} objects of class \code{RLum.Data.Curve} and \code{RLum.Analysis}
(with curve data) are supported.
}
\section{Function version}{
 0.1.0
}

\examples{

## load example data
data(ExampleData.CW_OSL_Curve, envir = environment())

## create RLum.Data.Curve object from this example
curve <-
  set_RLum(
      class = "RLum.Data.Curve",
      recordType = "OSL",
      data = as.matrix(ExampleData.CW_OSL_Curve)
  )

## plot data without and with smoothing
plot_RLum(curve)
plot_RLum(normalise_RLum(curve))

}
\seealso{
\linkS4class{RLum.Data.Curve}, \linkS4class{RLum.Analysis}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{utilities}
