% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AbanicoPlot.R
\name{plot_AbanicoPlot}
\alias{plot_AbanicoPlot}
\title{Function to create an abanico plot}
\usage{
plot_AbanicoPlot(
  data,
  log.z = TRUE,
  z.0 = c("mean.weighted", "mean", "median"),
  dispersion = c("qr", "sd", "2sd"),
  plot.ratio = 0.75,
  rotate = FALSE,
  mtext = "",
  summary = c("n", "in.2s"),
  summary.pos = "sub",
  summary.method = c("MCM", "weighted", "unweighted"),
  legend = NULL,
  legend.pos = "topleft",
  stats = NULL,
  rug = FALSE,
  kde = TRUE,
  hist = FALSE,
  dots = FALSE,
  boxplot = FALSE,
  y.axis = TRUE,
  error.bars = FALSE,
  bar = NULL,
  bar.col = NULL,
  polygon.col = NULL,
  line = NULL,
  line.col = NULL,
  line.lty = NULL,
  line.label = NULL,
  grid.col = NULL,
  frame = 1,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{data}{\link{data.frame} or \link[=RLum.Results-class]{RLum.Results} object (\strong{required}):
for \code{data.frame} two columns: De (\code{data[,1]}) and De error (\code{data[,2]}).
To plot several data sets in one plot the data sets must be provided as
\code{list}, e.g. \code{list(data.1, data.2)}. Rows with \code{NA} values will be removed
prior to plotting.\cr
For some \link[=RLum.Results-class]{RLum.Results} objects, one or more lines (and
corresponding labels) are drawn automatically:
\itemize{
\item \link{calc_AverageDose} (ADM)
\item \link{calc_CentralDose} (CDM)
\item \link{calc_MaxDose} (MDM)
\item \link{calc_MinDose} (MAM)
\item \link{calc_FiniteMixture}
Alternative labels can be set via the \code{line.label} option. This behaviour
can be suppressed altogether by setting \code{line = NA}.
}}

\item{log.z}{\link{logical} (\emph{with default}):
display the z-axis in logarithmic scale (\code{TRUE} by default). The setting is
automatically reset to \code{FALSE} if any zero values appear in the De column.}

\item{z.0}{\link{character} or \link{numeric} (\emph{with default}):
User-defined central value used for centring of data. One of \code{"mean.weighted"}
(default), \code{"mean"}and \code{"median"}, or a single numeric value (not its
logarithm).}

\item{dispersion}{\link{character} (\emph{with default}):
measure of dispersion used to draw the scatter polygon. Can be one of the
following:
\itemize{
\item \code{"qr"} (quartile range, default)
\item \code{"sd"} (standard deviation)
\item \code{"2sd"} (2 standard deviations)
\item \code{"pNN"} (symmetric percentile range, with \code{NN} being the lower percentile,
e.g. \code{"p05"} indicates the range between 5 and 95 \%, or \code{"p10"} indicates
the range between 10 and 90 \%)
}

The default is \code{"qr"}. Note that \code{"sd"} and \code{"2sd"} are only meaningful in
combination with \code{"z.0 = 'mean'"} because the unweighted mean is used to
centre the polygon.}

\item{plot.ratio}{\link{numeric} (\emph{with default}):
Relative space, given to the radial versus the cartesian plot part,
default is 0.75.}

\item{rotate}{\link{logical} (\emph{with default}):
Option to turn the plot by 90 degrees.}

\item{mtext}{\link{character} (\emph{with default}):
additional text below the plot title.}

\item{summary}{\link{character} (\emph{with default}):
add statistic measures of centrality and dispersion to the plot.
Can be one or more of several keywords. See details for available keywords.
Results differ depending on the log-option for the z-scale (see details).}

\item{summary.pos}{\link{numeric} or \link{character} (\emph{with default}):
optional position coordinates or keyword (e.g. \code{"topright"}) for the
statistical summary. Alternatively, the keyword \code{"sub"} may be
specified to place the summary below the plot header. However, this latter
option in only possible if \code{mtext} is not used.}

\item{summary.method}{\link{character} (\emph{with default}):
keyword indicating the method used to calculate the statistic summary.
One of \code{"MCM"} (default), \code{"weighted"} or \code{"unweighted"}.
See \link{calc_Statistics} for details.}

\item{legend}{\link{character} vector (\emph{optional}):
legend content to be added to the plot.}

\item{legend.pos}{\link{numeric} or \link{character} (\emph{with default}):
optional position coordinates or keyword (e.g. \code{"topright"})
for the legend to be plotted.}

\item{stats}{\link{character}:
additional labels of statistically important values in the plot. Can be one
or more of the following:
\itemize{
\item \code{"min"}
\item \code{"max"}
\item \code{"median"}
}}

\item{rug}{\link{logical} (\emph{with default}):
Option to add a rug to the KDE part, to indicate the location of individual values.}

\item{kde}{\link{logical} (\emph{with default}):
Option to add a KDE plot to the dispersion part (\code{TRUE} by default). The
smoothing bandwidth can be controlled by setting the \code{bw} argument (\code{"SJ"}
by default; see \link[stats:density]{stats::density} for details).}

\item{hist}{\link{logical} (\emph{with default}):
Option to add a histogram to the dispersion part. Only meaningful when only
one data set is plotted. The number of cells in the histogram can be
controlled by setting the \code{breaks} argument (see \link[graphics:hist]{graphics::hist} for
details).}

\item{dots}{\link{logical} (\emph{with default}):
Option to add a dot plot to the dispersion part. If number of dots exceeds
space in the dispersion part, a square indicates this.}

\item{boxplot}{\link{logical} (\emph{with default}):
Option to add a boxplot to the dispersion part, default is \code{FALSE}.}

\item{y.axis}{\link{logical} (\emph{with default}):
option to hide standard y-axis labels and show 0 only, useful for data with
small scatter. To suppress the y-axis entirely, use \code{yaxt == 'n'} (the
standard \link[graphics:par]{graphics::par} setting) instead.}

\item{error.bars}{\link{logical} (\emph{with default}):
Option to show De-errors as error bars on De-points. Useful in combination
with \verb{y.axis = FALSE, bar = FALSE}.}

\item{bar}{\link{numeric} or \link{logical} (\emph{with default}):
option to add one or more dispersion bars (i.e., bar showing the 2-sigma range)
centred at the defined values. By default a bar is drawn according to \code{"z.0"}.
To omit the bar set \code{"bar = FALSE"}.}

\item{bar.col}{\link{character} or \link{numeric} (\emph{with default}):
colour of the dispersion bar. Default is \code{"grey60"}.}

\item{polygon.col}{\link{character} or \link{numeric} (\emph{with default}):
colour of the polygon showing the data scatter. Sometimes this
polygon may be omitted for clarity. To disable it use \code{FALSE} or
\code{polygon = FALSE}. Default is \code{"grey80"}.}

\item{line}{\link{numeric} or \link[=RLum.Results-class]{RLum.Results}:
numeric values of the additional lines to be added. This can be set to \code{NA}
to suppress the line added automatically by some \code{RLum.Results} objects.}

\item{line.col}{\link{character} or \link{numeric}:
colour of the additional lines.}

\item{line.lty}{\link{integer}:
line type of additional lines.}

\item{line.label}{\link{character}:
labels for the additional lines.}

\item{grid.col}{\link{character} or \link{numeric} (\emph{with default}):
colour of the grid lines (originating at \verb{[0,0]} and stretching to
the z-scale). To disable grid lines use \code{FALSE}. Default is \code{"grey"}.}

\item{frame}{\link{numeric} (\emph{with default}):
the plot frame type, one of the following:
\itemize{
\item 0: no frame
\item 1: frame originates at 0,0 and runs along min/max isochrons (default)
\item 2: frame embraces the 2-sigma bar
\item 3: frame embraces the entire plot as a rectangle
}}

\item{interactive}{\link{logical} (\emph{with default}):
create an interactive abanico plot (requires the \code{'plotly'} package).}

\item{...}{further arguments and graphical parameters to control the plot
output. Supported are: \code{main}, \code{sub}, \code{ylab}, \code{xlab}, \code{zlab}, \code{xlim},
\code{ylim}, \code{zlim}, \code{cex}, \code{pt.cex} (point size), \code{lty}, \code{lwd}, \code{pch}, \code{col},
\code{at}, \code{bw}, and \code{breaks}. \code{xlab} must be a vector of length two, specifying
the upper and lower x-axis labels.

Please note that in the interactive mode, if you are using an expression,
the \code{zlab} must use HTML tags, such as \verb{D<sub>e</sub>} for \code{D[e]}.}
}
\value{
Returns a plot object and, optionally, a list with plot calculus data.
}
\description{
The abanico plot provides a comprehensive presentation of data precision
and its dispersion around a central value, and can also display a
kernel density estimate, histogram and/or dot plot of the dose values.
}
\details{
The abanico plot is a combination of the classic radial plot
\link{plot_RadialPlot} and a kernel density estimate plot (e.g
\link{plot_KDE}. It allows straightforward visualisation of data precision,
error scatter around a user-defined central value and the combined
distribution of the values, on the actual scale of the measured data (e.g.
seconds, equivalent dose, years). The principle of the plot is shown in
Galbraith & Green (1990). The function authors are thankful for the
thought-provoking figure in this article.

The semi circle (z-axis) of the classic radial plot is bent to a straight
line here, which actually is the basis for combining this polar (radial)
part of the plot with any other Cartesian visualisation method
(KDE, histogram, PDF and so on). Note that the plot allows displaying
two measures of distribution. One is the 2-sigma
bar, which illustrates the spread in value errors, and the other is the
polygon, which stretches over both parts of the abanico plot (polar and
Cartesian) and illustrates the actual spread in the values themselves.

Since the 2-sigma-bar is a polygon, it can be (and is) filled with shaded
lines. To change density (lines per inch, default is 15) and angle (default
is 45 degrees) of the shading lines, specify these parameters. See
\link[graphics:polygon]{graphics::polygon} for further help.

The proportion of the polar part and the cartesian part of the abanico plot
can be modified for display reasons (\code{plot.ratio = 0.75}). By default,
the polar part spreads over 75 \% and leaves 25 \% for the part that
shows the KDE graph.

The abanico plot supports other than the weighted mean as measure of
centrality. When it is obvious that the data
is not (log-)normally distributed, the mean (weighted or not) cannot be a
valid measure of centrality and hence central dose. Accordingly, the median
and the weighted median can be chosen as well to represent a proper measure
of centrality (e.g. \code{centrality = "median.weighted"}). Also
user-defined numeric values (e.g. from the central age model) can be used if
this appears appropriate.
\subsection{Statistics summary}{

A statistic summary, i.e. a collection of statistic measures of
centrality and dispersion (and further measures) can be added by specifying
one or more of the following keywords in the \code{summary} argument:
\itemize{
\item \code{"n"} (number of samples)
\item \code{"mean"} (mean De value)
\item \code{"median"} (median of the De values)
\item \code{"sd.abs"} (absolute standard deviation)
\item \code{"sd.rel"} (relative standard deviation in percent)
\item \code{"se.abs"} (absolute standard error)
\item \code{"se.rel"} (relative standard error)
\item \code{"in.2s"} (percent of samples in 2-sigma range)
\item \code{"q.25"} (25\% percentile)
\item \code{"q.75"} (75\% percentile)
\item \code{"skewness"} (skewness)
\item \code{"kurtosis"} (kurtosis)
}

\strong{Note:} the input data for the statistic summary is sent to the function
\link{calc_Statistics} depending on the log-option for the z-scale. If
\code{"log.z = TRUE"}, the summary is based on the logarithms of the input
data. If \code{"log.z = FALSE"} the linearly scaled data is used.

\strong{Note:} \link{calc_Statistics} calculates these statistic
measures in three different ways: \code{unweighted}, \code{weighted} and
\code{MCM-based} (i.e., based on Monte Carlo Methods). By default, the
MCM-based version is used. If you wish to use another method, indicate this
with the appropriate keyword using the argument \code{summary.method}.
}

\subsection{Other arguments}{

The optional parameter \code{layout} allows more sophisticated ways to modify
the entire plot. Each element of the plot can be addressed and its properties
can be defined. This includes font type, size and decoration, colours and
sizes of all plot items. To infer the definition of a specific layout style
cf. \link{get_Layout} or type, e.g. for the layout type \code{"journal"}.
A layout type can be modified by the user by
assigning new values to the list object.

It is possible for the z-scale to specify where ticks are to be drawn
by using the parameter \code{at}, e.g. \code{at = seq(80, 200, 20)}, cf. function
documentation of \code{axis}. Specifying tick positions manually overrides a
\code{zlim}-definition.
}
}
\section{Function version}{
 0.1.25
}

\examples{

## load example data and recalculate to Gray
data(ExampleData.DeValues, envir = environment())
ExampleData.DeValues <- ExampleData.DeValues$CA1

## plot the example data straightforward
plot_AbanicoPlot(data = ExampleData.DeValues)

## now with linear z-scale
plot_AbanicoPlot(data = ExampleData.DeValues,
                 log.z = FALSE)

## now with output of the plot parameters
plot1 <- plot_AbanicoPlot(data = ExampleData.DeValues)
str(plot1)
plot1$zlim

## now with adjusted z-scale limits
plot_AbanicoPlot(data = ExampleData.DeValues,
                 zlim = c(10, 200))

## now with adjusted x-scale limits
plot_AbanicoPlot(data = ExampleData.DeValues,
                 xlim = c(0, 20))

## now with rug to indicate individual values in KDE part
plot_AbanicoPlot(data = ExampleData.DeValues,
                 rug = TRUE)

## now with a smaller bandwidth for the KDE plot
plot_AbanicoPlot(data = ExampleData.DeValues,
                 bw = 0.04)

## now with a histogram instead of the KDE plot
plot_AbanicoPlot(data = ExampleData.DeValues,
                 hist = TRUE,
                 kde = FALSE)

## now with a KDE plot and histogram with manual number of bins
plot_AbanicoPlot(data = ExampleData.DeValues,
                 hist = TRUE,
                 breaks = 20)

## now with a KDE plot and a dot plot
plot_AbanicoPlot(data = ExampleData.DeValues,
                 dots = TRUE)

## now with user-defined plot ratio
plot_AbanicoPlot(data = ExampleData.DeValues,
                 plot.ratio = 0.5)
## now with user-defined central value
plot_AbanicoPlot(data = ExampleData.DeValues,
                 z.0 = 70)

## now with median as central value
plot_AbanicoPlot(data = ExampleData.DeValues,
                 z.0 = "median")

## now with the 17-83 percentile range as definition of scatter
plot_AbanicoPlot(data = ExampleData.DeValues,
                 z.0 = "median",
                 dispersion = "p17")

## now with user-defined green line for minimum age model
CAM <- calc_CentralDose(ExampleData.DeValues,
                        plot = FALSE)
plot_AbanicoPlot(data = CAM,
                 line.col = "darkseagreen")

## now create plot with legend, colour, different points and smaller scale
plot_AbanicoPlot(data = ExampleData.DeValues,
                 legend = "Sample 1",
                 col = "tomato4",
                 bar.col = "peachpuff",
                 pch = "R",
                 cex = 0.8)

## now without 2-sigma bar, polygon, grid lines and central value line
plot_AbanicoPlot(data = ExampleData.DeValues,
                 bar.col = FALSE,
                 polygon.col = FALSE,
                 grid.col = FALSE,
                 y.axis = FALSE,
                 lwd = 0)

## now with direct display of De errors, without 2-sigma bar
plot_AbanicoPlot(data = ExampleData.DeValues,
                 bar.col = FALSE,
                 ylab = "",
                 y.axis = FALSE,
                 error.bars = TRUE)

## now with user-defined axes labels
plot_AbanicoPlot(data = ExampleData.DeValues,
                 xlab = c("Data error (\%)",
                          "Data precision"),
                 ylab = "Scatter",
                 zlab = "Equivalent dose [Gy]")

## now with minimum, maximum and median value indicated
plot_AbanicoPlot(data = ExampleData.DeValues,
                 stats = c("min", "max", "median"))

## now with a brief statistical summary as subheader
plot_AbanicoPlot(data = ExampleData.DeValues,
                 summary = c("n", "in.2s"))

## now with another statistical summary
plot_AbanicoPlot(data = ExampleData.DeValues,
                 summary = c("mean.weighted", "median"),
                 summary.pos = "topleft")

## now a plot with two 2-sigma bars for one data set
plot_AbanicoPlot(data = ExampleData.DeValues,
                 bar = c(30, 100))

## now the data set is split into sub-groups, one is manipulated
data.1 <- ExampleData.DeValues[1:30,]
data.2 <- ExampleData.DeValues[31:62,] * 1.3

## now a common dataset is created from the two subgroups
data.3 <- list(data.1, data.2)

## now the two data sets are plotted in one plot
plot_AbanicoPlot(data = data.3)

## now with some graphical modification
plot_AbanicoPlot(data = data.3,
                 z.0 = "median",
                 col = c("steelblue4", "orange4"),
                 bar.col = c("steelblue3", "orange3"),
                 polygon.col = c("steelblue1", "orange1"),
                 pch = c(2, 6),
                 angle = c(30, 50),
                 summary = c("n", "in.2s", "median"))

## create Abanico plot with predefined layout definition
plot_AbanicoPlot(data = ExampleData.DeValues,
                 layout = "journal")

## now with predefined layout definition and further modifications
plot_AbanicoPlot(
 data = data.3,
 z.0 = "median",
 layout = "journal",
 col = c("steelblue4", "orange4"),
 bar.col = adjustcolor(c("steelblue3", "orange3"),
                         alpha.f = 0.5),
 polygon.col = c("steelblue3", "orange3"))

## for further information on layout definitions see documentation
## of function get_Layout()

## now with manually added plot content
## create empty plot with numeric output
AP <- plot_AbanicoPlot(data = ExampleData.DeValues,
                       pch = NA)

## identify data in 2 sigma range
in_2sigma <- AP$data[[1]]$data.in.2s

## restore function-internal plot parameters
par(AP$par)

## add points inside 2-sigma range
points(x = AP$data[[1]]$precision[in_2sigma],
       y = AP$data[[1]]$std.estimate.plot[in_2sigma],
       pch = 16)

## add points outside 2-sigma range
points(x = AP$data[[1]]$precision[!in_2sigma],
       y = AP$data[[1]]$std.estimate.plot[!in_2sigma],
       pch = 1)

} 

\section{How to cite}{
Dietze, M., Kreutzer, S., Colombo, M., 2026. plot_AbanicoPlot(): Function to create an abanico plot. Function version 0.1.25. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\references{
Galbraith, R. & Green, P., 1990. Estimating the component ages
in a finite mixture. International Journal of Radiation Applications and
Instrumentation. Part D. Nuclear Tracks and Radiation Measurements 17 (3),
197-206.

Dietze, M., Kreutzer, S., Burow, C., Fuchs, M.C., Fischer, M., Schmidt, C., 2015.
The abanico plot: visualising chronometric data with individual standard errors.
Quaternary Geochronology. doi:10.1016/j.quageo.2015.09.003
}
\seealso{
\link{plot_RadialPlot}, \link{plot_KDE},
\link{plot_Histogram}, \link{plot_ViolinPlot}
}
\author{
Michael Dietze, GFZ Potsdam (Germany)\cr
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
Inspired by a plot introduced by Galbraith & Green (1990)
, RLum Developer Team}
