% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DRCSummary.R
\name{plot_DRCSummary}
\alias{plot_DRCSummary}
\title{Create a Dose-Response Curve Summary Plot}
\usage{
plot_DRCSummary(
  object,
  source_dose_rate = NULL,
  sel_curves = NULL,
  show_dose_points = FALSE,
  show_natural = FALSE,
  n = 51L,
  ...
)
}
\arguments{
\item{object}{\link[=RLum.Results-class]{RLum.Results} (\strong{required}): input object created by
\link{analyse_SAR.CWOSL}. The input object can be provided as \link{list}.}

\item{source_dose_rate}{\link{numeric} (\emph{optional}): allows to modify the axis
and show values in Gy, instead seconds. Only a single numerical value is
allowed.}

\item{sel_curves}{\link{numeric} (\emph{optional}): id of the curves to be plotted in
its occurring order. A sequence can be provided for selecting, e.g., only
every 2nd curve from the input object.}

\item{show_dose_points}{\link{logical} (\emph{with default}): enable/disable plotting
of dose points in the graph.}

\item{show_natural}{\link{logical} (\emph{with default}): enable/disable the plot
of the natural \code{Lx/Tx} values.}

\item{n}{\link{integer} (\emph{with default}): number of x-values used to evaluate
one curve object. Large numbers slow down the plotting process and are
usually not needed.}

\item{...}{further arguments and graphical parameters to control the plot
output. Supported are: \code{main}, \code{xlab}, \code{ylab}, \code{xlim}, \code{ylim}, \code{lty}, \code{lwd},
\code{pch}, \code{col.pch}, \code{col.lty}, \code{mtext}, and \code{pt.cex} (point size).}
}
\value{
An \link[=RLum.Results-class]{RLum.Results} object is returned:

Slot: \strong{@data}\cr

\tabular{lll}{
\strong{OBJECT} \tab \strong{TYPE} \tab \strong{COMMENT}\cr
\code{results} \tab \link{data.frame} \tab with dose and LxTx values \cr
\code{data} \tab \link[=RLum.Results-class]{RLum.Results} \tab original input data \cr
}

Slot: \strong{@info}\cr

\tabular{lll}{
\strong{OBJECT} \tab \strong{TYPE} \tab \strong{COMMENT} \cr
\code{call} \tab \code{call} \tab the original function call \cr
\code{args} \tab \code{list} \tab arguments of the original function call \cr
}

\emph{Note: If the input object is a \link{list} a list of \link[=RLum.Results-class]{RLum.Results} objects is returned.}
}
\description{
While analysing OSL SAR or pIRIR-data the view on the data is usually
limited to one dose-response curve (DRC) at the time for one aliquot. This
function overcomes this limitation by plotting all DRCs from an
\link[=RLum.Results-class]{RLum.Results} object created by \link{analyse_SAR.CWOSL} in one single
plot.

If you want plot your DRC on an energy scale (dose in Gy), you can either
use option \code{source_dose_rate} or perform your SAR analysis with the dose
points in Gy (better axis scaling).
}
\section{Function version}{
 0.2.5
}

\examples{

#load data example data
data(ExampleData.BINfileData, envir = environment())

#transform the values from the first position in a RLum.Analysis object
object <- Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos=1)

## perform SAR analysis
results <- analyse_SAR.CWOSL(
    object,
    signal_integral = 1:2,
    background_integral = 900:1000,
    plot = FALSE
)

##plot only DRC
plot_DRCSummary(results)

}
\seealso{
\link[=RLum.Results-class]{RLum.Results}, \link{analyse_SAR.CWOSL}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany) \cr
Christoph Burow, University of Cologne (Germany)
}
