% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DRCSummary.R
\name{plot_DRCSummary}
\alias{plot_DRCSummary}
\title{Create a Dose-Response Curve Summary Plot}
\usage{
plot_DRCSummary(object, sel_curves = NULL, show_dose_points = FALSE,
  show_natural = FALSE, n = 51L, ...)
}
\arguments{
\item{object}{\linkS4class{RLum.Results} object (\strong{required}): input object created by the function \link{analyse_SAR.CWOSL}}

\item{sel_curves}{\link{numeric} (optional): id of the curves to be plotting in its occuring order. A sequence can
be provided for selecting, e.g., only every 2nd curve from the input object}

\item{show_dose_points}{\link{logical} (with default): enable or disable plot of dose points in the graph}

\item{show_natural}{\link{logical} (with default): enable or disable the plot of the natural Lx/Tx values}

\item{n}{\link{integer} (with default): the number of x-values used to evaluate one curve object. Large numbers slow
down the plotting process and are usually not needed}

\item{...}{Further arguments and graphical parameters to be passed.}
}
\value{
An \linkS4class{RLum.Results} object is returned:

Slot: \strong{@data}\cr

\tabular{lll}{
\strong{OBJECT} \tab \strong{TYPE} \tab \strong{COMMENT}\cr
\code{results} \tab \link{data.frame} \tab with dose and LxTx values \cr
\code{data} \tab \linkS4class{RLum.Results} \tab original input data \cr
}

Slot: \strong{@info}\cr

\tabular{lll}{
\strong{OBJECT} \tab \strong{TYPE} \tab \strong{COMMENT} \cr
\code{call} \tab \code{call} \tab the original function call \cr
\code{args} \tab \code{list} \tab arguments of the original function call \cr
}
}
\description{
While analysing OSL SAR or pIRIR data the view on the data is limited usually to one
dose-response curve (DRC) at the time for one aliquot. This function overcomes this limitation
by plotting all DRC from an \linkS4class{RLum.Results} object created by the function \link{analyse_SAR.CWOSL}
in one single plot.
}
\details{
If you want plot your DRC on an energy scale (dose in Gy), you should run the SAR analysis with the
dose points in Gy. If you need more options than provided with this function, please have
a look into the source code to create and modify an own R script.
}
\section{Function version}{
 0.1.1 (2018-04-27 17:44:10)
}

\examples{

#load data example data
data(ExampleData.BINfileData, envir = environment())

#transform the values from the first position in a RLum.Analysis object
object <- Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos=1)

results <- analyse_SAR.CWOSL(
  object = object,
  signal.integral.min = 1,
  signal.integral.max = 2,
   background.integral.min = 900,
   background.integral.max = 1000,
   plot = FALSE
 )

##plot only DRC
plot_DRCSummary(results)

}
\seealso{
\linkS4class{RLum.Results}, \link{analyse_SAR.CWOSL}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Université Bordeaux Montaigne (France) \cr
Christoph Burow, University of Cologne
\cr R Luminescence Package Team} 

\section{How to cite}{
Kreutzer, S., Burow, C. (2018). plot_DRCSummary(): Create a Dose-Response Curve Summary Plot. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0. https://CRAN.R-project.org/package=Luminescence
}

