% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Crosstalk.R
\name{plot_MoranScatterplot}
\alias{plot_MoranScatterplot}
\title{Moran scatter plot: visualizing spatial dependency}
\usage{
plot_MoranScatterplot(
  object,
  df_neighbour = get_Neighbour(object = object),
  bo_show_legend = TRUE,
  str_y_def = "mean_neighbours",
  pch = 20,
  cex = 0.8,
  xlab = "Value central grain",
  ylab = ifelse(str_y_def == "weighted_sum", yes = "Weighted sum neighbouring grains", no
    = "Plain mean neighbour grains"),
  log = "",
  ...
)
}
\arguments{
\item{object}{\linkS4class{RLum.Results} or \link{numeric} (\strong{required}): containing a numerical vector of length 100,
representing one or more measurement discs ("positions") in a reader.
Each element in the vector represents one grain hole location on a disc.}

\item{df_neighbour}{\link{data.frame} (\emph{with default}) Dataframe, indicating
which borders to consider, and their respective weights.
Defaults to \code{get_Neighbour(object = vn_values)}.}

\item{bo_show_legend}{\link{logical} (\emph{with default}) Single Boolean. If \code{TRUE}
(default), a small legend explaining the different lines is plotted.}

\item{str_y_def}{\link{character} (\emph{with default}) Calculation of y position. Defaults to \code{"mean_neighbours"}
which is the plain mean of all neighbour values and the most illustrative. The other option is \code{"weighted_sum"},
which means the sum of the border weights times the neighbour values, which is actually closer
to the way Moran's I is by default calculated in this package.}

\item{pch}{\link{character}, \link{integer} (\emph{with default}): Accepts options
\code{"show_location_ids"} (plots grain location id's), \code{"show_n_neighbours"}
(indicates numbers of neighbours) and the normal base plot \code{pch} options.}

\item{cex}{FIXME(mcol)}

\item{xlab}{FIXME(mcol)}

\item{ylab}{FIXME(mcol)}

\item{log}{FIXME(mcol)}

\item{...}{Other parameters to be forwarded to the base R plot functions. Note that \code{xlab} (x axis label),
\code{ylab} (y axis label) and \code{cex} (scaling value) are given default values. Because of sometimes large value
differences, \code{log = "x"}, \code{log = "y"} and \code{log = "xy"}are supported. In case of negative values and logarithmic
plotting, values are increased so the smallest value to plot is 1. Summary elements such as means, least
square line etc. will still be based on the linear case.}
}
\value{
Dataframe with plotting coordinates and grain location id's, for creating user-defined
plots. Invisible.
}
\description{
Scatter plot, with on the x axis the original grain signal and on the y axis
the weighted mean of the neighbor grain signals. The plot area is divided into four quadrants,
and also a least square line (which slopes indicates, but not exactly represents, Moran's I) and
an 1:1 line (which indicates a Moran's I of around 1).
}
\details{
Note that this function plots on the y-axis the mean of the neighbours, while the function
\code{calc_MoransI()} by default will for its global calculation weight every border the same. So, grain locations
with 1, 2 or 3 neighbours will appear higher on the y-axis than their influence on Moran's I justify -- apart
from scaling, this explains a part of the differences of Moran's scatter plots between different packages.
Also note that island' grain locations (=those not bordering other grains) are left out of these plots but
might still influence Moran's I calculations.
}
\author{
Anna-Maartje de Boer, Luc Steinbuch, Wageningen University & Research, 2025
}
