% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLum.R
\name{plot_RLum}
\alias{plot_RLum}
\title{General plot function for RLum S4 class objects}
\usage{
plot_RLum(object, ...)
}
\arguments{
\item{object}{\linkS4class{RLum} (\strong{required}):
object of class \linkS4class{RLum} or a list of such objects. If a
list is provided, the function tries to plot every object in the list
according to its \code{RLum} class, after removing non-RLum objects.}

\item{...}{further arguments and graphical parameters to pass to the
specific plot functions. The only arguments that are supported directly are
\code{main} (plot title) and \code{mtext} (plot subtitle). Here \code{main} can be provided
as a list and the arguments in the
list will dispatched to the plots if \code{object} is of type \code{list} as well.}
}
\value{
Produces a plot depending on the input object.
}
\description{
The function calls object specific plot functions for RLum S4 class objects.
}
\details{
The function provides a generalised access point for plotting various
\linkS4class{RLum} objects. Depending on the input object, the
corresponding plot function will be selected. Allowed arguments can be
found in the documentations of each plot function.

\tabular{lll}{
\strong{object} \tab \strong{corresponding plot function} \cr
\linkS4class{RLum.Data.Curve} \tab \link{plot_RLum.Data.Curve} \cr
\linkS4class{RLum.Data.Spectrum} \tab \link{plot_RLum.Data.Spectrum}\cr
\linkS4class{RLum.Data.Image} \tab \link{plot_RLum.Data.Image}\cr
\linkS4class{RLum.Analysis} \tab \link{plot_RLum.Analysis}\cr
\linkS4class{RLum.Results} \tab \link{plot_RLum.Results}
}
}
\section{Function version}{
 0.5
}

\examples{
#load Example data
data(ExampleData.CW_OSL_Curve, envir = environment())

#transform data.frame to RLum.Data.Curve object
temp <- as(ExampleData.CW_OSL_Curve, "RLum.Data.Curve")

#plot RLum object
plot_RLum(temp)

}
\seealso{
\link{plot_RLum.Analysis},
\link{plot_RLum.Data.Curve},
\link{plot_RLum.Data.Spectrum},
\link{plot_RLum.Data.Image},
\link{plot_RLum.Results}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
}
\keyword{dplot}
