% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROI.R
\name{plot_ROI}
\alias{plot_ROI}
\title{Create Regions of Interest (ROI) Graphic}
\usage{
plot_ROI(object, plot = TRUE, ...)
}
\arguments{
\item{object}{\linkS4class{RLum.Analysis} or a \link{list} of such objects (\strong{required}):
data input. Please note that to avoid function errors, only input created
by the function \link{read_RF2R} is accepted}

\item{plot}{\link{logical} (\emph{with default}): enable or disable plot output to use
the function only to extract the ROI data}

\item{...}{further parameters to manipulate the plot. On top of all arguments of
\link[graphics:plot.default]{graphics::plot.default} the following arguments are supported: \code{lwd.ROI}, \code{lty.ROI},
\code{col.ROI}, \code{col.pixed}, \code{text.labels}, \code{text.offset}, \code{grid} (\code{TRUE/FALSE}), \code{legend} (\code{TRUE/FALSE}),
\code{legend.text}, \code{legend.pos}}
}
\value{
An ROI plot and an \linkS4class{RLum.Results} object with a matrix containing
the extracted ROI data.
}
\description{
Create Regions of Interest (ROI) Graphic
}
\section{Function version}{
 0.1.0
}

\examples{
file <- system.file("extdata", "RF_file.rf", package = "Luminescence")
temp <- read_RF2R(file)
plot_ROI(temp)

}
\seealso{
\link{read_RF2R}, \link{analyse_IRSAR.RF}
}
\author{
Sebastian Kreutzer, Department of Geography & Earth Sciences, Aberystwyth University
(United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. plot_ROI(): Create Regions of Interest (ROI) Graphic. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., 2020. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.8.9000-85. https://CRAN.R-project.org/package=Luminescence
}

\keyword{datagen}
\keyword{plot}
