% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSL2Risoe.BINfileData.R
\name{PSL2Risoe.BINfileData}
\alias{PSL2Risoe.BINfileData}
\title{Convert portable OSL data to an Risoe.BINfileData object}
\usage{
PSL2Risoe.BINfileData(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{RLum.Analysis}} (\bold{required}):
\code{RLum.Analysis} object produced by \code{\link{read_PSL2R}}}

\item{...}{currently not used.}
}
\value{
Returns an S4 \code{\linkS4class{Risoe.BINfileData}} object that can
be used to write a BIN file using \code{\link{write_R2BIN}}.
}
\description{
Converts an RLum.Analysis object produced by the function 'read_PSL2R()' to
an Risoe.BINfileData object
}
\details{
<placeholder>
}
\section{Function version}{
 0.0.1
}
\examples{
# none available yet

}
\author{
Christoph Burow, University of Cologne (Germany)
}
\seealso{
\code{\linkS4class{RLum.Analysis}}, \code{\linkS4class{RLum.Data.Curve}},
\code{\linkS4class{Risoe.BINfileData}}
}
\keyword{IO}

