% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_PSL2Risoe.BINfileData.R
\name{PSL2Risoe.BINfileData}
\alias{PSL2Risoe.BINfileData}
\title{Convert portable OSL data to a Risoe.BINfileData object}
\usage{
PSL2Risoe.BINfileData(...)
}
\arguments{
\item{...}{Unused.}
}
\description{
This function is defunct, use \link{convert_PSL2Risoe.BINfileData} instead.
}
