% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Daybreak2R.R
\name{read_Daybreak2R}
\alias{read_Daybreak2R}
\title{Import Daybreak ASCII dato into R}
\usage{
read_Daybreak2R(file, verbose = TRUE, txtProgressBar = TRUE)
}
\arguments{
\item{file}{\code{\link{character}} or \code{\link{list}} (\bold{required}): path and file name of the
file to be imported. Alternatively a list of file names can be provided or just the path a folder
containing measurement data. Please note that the specific, common, file extension (txt) is likely
leading to function failures during import when just a path is provided.}

\item{verbose}{\code{\link{logical}} (with default): enables or disables terminal feedback}

\item{txtProgressBar}{\code{\link{logical}} (with default): enables or disables
\code{\link{txtProgressBar}}.}
}
\value{
A list of \code{\linkS4class{RLum.Analysis}} objects (each per position) is provided.
}
\description{
Import a *.txt (ASCII) file produced by a Daybreak reader into R.
}
\note{
\bold{[BETA VERSION]} This function version still needs to be properly tested.
}
\section{Function version}{
 0.2.1 (2016-05-02 10:24:12)
}
\examples{

## This function has no example yet.

}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)\cr Based on a suggestion by Willian Amidon and Andrew Louis Gorin.
\cr R Luminescence Package Team} 

\section{How to cite}{
Kreutzer, S. (2016). read_Daybreak2R(): Import Daybreak ASCII dato into R. Function version 0.2.1. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2016). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.7.0. https://CRAN.R-project.org/package=Luminescence
}

\references{
-
}
\seealso{
\code{\linkS4class{RLum.Analysis}}, \code{\linkS4class{RLum.Data.Curve}}
}
\keyword{IO}

