% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_Daybreak2R.R
\name{read_Daybreak2R}
\alias{read_Daybreak2R}
\title{Import Daybreak ASCII dato into R}
\usage{
read_Daybreak2R(file, txtProgressBar = TRUE)
}
\arguments{
\item{file}{\code{\link{character}} (\bold{required}): path and file name of the
file to import}

\item{txtProgressBar}{\link{logical} (with default): enables or disables
\code{\link{txtProgressBar}}.}
}
\value{
A list of \code{\linkS4class{RLum.Analysis}} objects (each per position) is provided.
}
\description{
Import a *.txt produced by a Daybreak reader into R
}
\note{
[BETA VERSION]
}
\section{Function version}{
 0.1.0 (2015-08-10 22:04:05)
}
\examples{
##No example yet
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
\cr R Luminescence Package Team}
\references{
-
}
\seealso{
\code{\linkS4class{RLum.Analysis}}, \code{\linkS4class{RLum.Data.Curve}}
}
\keyword{IO}

