% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HeliosOSL2R.R
\name{read_HeliosOSL2R}
\alias{read_HeliosOSL2R}
\title{Import Luminescence Data from Helios Luminescence Reader}
\usage{
read_HeliosOSL2R(file, verbose = TRUE, ...)
}
\arguments{
\item{file}{\link{character}, \link{list} (\strong{required}):
name of one or multiple \code{.osl} files (URLs are supported); it can be the
path to a directory, in which case the function tries to detect and import
all \code{.osl} files found in the directory.}

\item{verbose}{\link{logical} (\emph{with default}):
enable/disable output to the terminal.}

\item{...}{not in use, for compatibility reasons only}
}
\value{
A \linkS4class{RLum.Analysis} object. Results are returned as a
list when multiple files are processed or \code{file} is a list.
}
\description{
Import of files with \code{.osl} extension produced by the zero rad Helios
luminescence reader and conversion to \linkS4class{RLum.Analysis}
objects.
}
\note{
Thanks to Krzysztof Maternicki for providing example data.
}
\section{Function version}{
 0.1.1
}

\examples{
file <- system.file("extdata/HeliosOSL_Example.osl", package = "Luminescence")
read_HeliosOSL2R(file)

}
\seealso{
\linkS4class{RLum.Data.Curve}, \linkS4class{RLum.Analysis}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{IO}
