% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HeliosOSL2R.R
\name{read_HeliosOSL2R}
\alias{read_HeliosOSL2R}
\title{Import Luminescence Data from Helios Luminescence Reader}
\usage{
read_HeliosOSL2R(file, verbose = TRUE, ...)
}
\arguments{
\item{file}{\link{character} (\strong{required}): path to file to be imported. Can be a \link{list}
for further processing}

\item{verbose}{\link{logical}: enable/disable terminal feedback}

\item{...}{not in use, for compatibility reasons only}
}
\value{
\linkS4class{RLum.Analysis} object
}
\description{
Straightforward import of files with the ending \code{.osl} produced
by the zero rad Helios luminescence reader and conversion to \linkS4class{RLum.Analysis} objects.
}
\section{Function version}{
 0.1.0
}

\examples{
##TODO

}
\seealso{
\linkS4class{RLum.Data.Curve}, \linkS4class{RLum.Analysis}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
\keyword{IO}
