% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PSL2R.R
\name{read_PSL2R}
\alias{read_PSL2R}
\title{Import PSL files to R}
\usage{
read_PSL2R(file, drop.bg = FALSE, as.decay.curve = TRUE, ...)
}
\arguments{
\item{file}{\code{\link{character}} or \code{\link{list}} (\bold{required}): path and file name of the
PSL file. If input is a \code{list} it should comprise only \code{character}s representing
valid paths and PSL file names.
Alternatively the input character can be just a directory (path). In this case the
the function tries to detect and import all PSL files found in the directory.}

\item{drop.bg}{\code{\link{logical}} (with default):  \code{TRUE} to automatically 
remove all non-OSL/IRSL curves.}

\item{as.decay.curve}{\code{\link{logical}} (with default): Portable OSL Reader curves
are often given as cumulative light sum curves. Use \code{TRUE} (default) to convert
the curves to the more usual decay form.}

\item{...}{currently not used.}
}
\value{
Returns an S4 \code{\linkS4class{RLum.Analysis}} object.
}
\description{
Imports PSL files produced by a SUERC portable OSL reader into R
}
\details{
<placeholder>
}
\section{Function version}{
 0.0.1
}
\examples{
# none available yet

}
\author{
Christoph Burow, University of Cologne (Germany)
}
\seealso{
\code{\linkS4class{RLum.Analysis}},
\code{\linkS4class{RLum.Data.Curve}}
}
\keyword{IO}

