% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_RF2R.R
\name{read_RF2R}
\alias{read_RF2R}
\title{Import RF-files to R}
\usage{
read_RF2R(file, verbose = TRUE, ...)
}
\arguments{
\item{file}{\link{character} (\strong{required}):
name of one or multiple RF files (URLs are supported); it can be the path
to a directory, in which case the function tries to detect and import all
RF files found in the directory.}

\item{verbose}{\link{logical} (\emph{with default}):
enable/disable output to the terminal.}

\item{...}{not used, only for compatible reasons}
}
\value{
Returns an S4 \link[=RLum.Analysis-class]{RLum.Analysis} object containing
\link[=RLum.Data.Curve-class]{RLum.Data.Curve} objects for each curve. Results are
returned as a list when multiple files are processed or \code{file} is a list.
The header may contain \code{key = "value"} pairs, which will be stored in the
\code{info} slot; incomplete pairs are silently dropped.
}
\description{
Import files produced by the IR-RF 'ImageJ' macro (\code{SR-RF.ijm};
Mittelstraß and Kreutzer, 2021) into R and create a list of
\link[=RLum.Analysis-class]{RLum.Analysis} objects.
}
\details{
The results of spatially resolved IR-RF data are summarised in
so-called RF-files (Mittelstraß and Kreutzer, 2021).
The output of the function can be passed to function \link{analyse_IRSAR.RF}.
}
\section{Function version}{
 0.1.4
}

\examples{
file <- system.file("extdata", "RF_file.rf", package = "Luminescence")
object <- read_RF2R(file)

} 

\section{How to cite}{
Kreutzer, S., Colombo, M., 2026. read_RF2R(): Import RF-files to R. Function version 0.1.4. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\references{
Mittelstraß, D., Kreutzer, S., 2021. Spatially resolved infrared radiofluorescence:
single-grain K-feldspar dating using CCD imaging. Geochronology 3, 299–319. \doi{10.5194/gchron-3-299-2021}
}
\seealso{
\link[=RLum.Analysis-class]{RLum.Analysis}, \link[=RLum.Data.Curve-class]{RLum.Data.Curve},
\link{analyse_IRSAR.RF}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)\cr
, RLum Developer Team}
\keyword{IO}
