% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_RF2R.R
\name{read_RF2R}
\alias{read_RF2R}
\title{Import RF files to R}
\usage{
read_RF2R(file)
}
\arguments{
\item{file}{\link{character} (\strong{required}): path and file name of the RF file}
}
\value{
Returns an S4 \linkS4class{RLum.Analysis} object containing
\linkS4class{RLum.Data.Curve} objects for each curve.
}
\description{
TODO
}
\details{
TODO
}
\section{Function version}{
 0.0.1 (2018-11-22 19:16:08)
}

\examples{

##Import
file <- system.file("extdata", "RF_file.rf", package = "Luminescence")
temp <- read_RF2R(file)

}
\seealso{
\linkS4class{RLum.Analysis}, \linkS4class{RLum.Data.Curve}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS-Université Bordeaux Montaigne (France)
\cr R Luminescence Package Team} 

\section{How to cite}{
Kreutzer, S. (2018). read_RF2R(): Import RF files to R. Function version 0.0.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0. https://CRAN.R-project.org/package=Luminescence
}

\keyword{IO}
