% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TIFF2R.R
\name{read_TIFF2R}
\alias{read_TIFF2R}
\title{Import TIFF Image Data into R}
\usage{
read_TIFF2R(file)
}
\arguments{
\item{file}{\link{character} (\strong{required}): file name}
}
\value{
\linkS4class{RLum.Data.Image} object
}
\description{
Simple wrapper around \link[tiff:readTIFF]{tiff::readTIFF} to import TIFF images
and TIFF image stacks to be further processed within the package \code{'Luminescence'}
}
\section{Function version}{
 0.1.1
}

\examples{

\dontrun{
file <- file.choose()
image <- read_TIFF2R(file)

}

}
\seealso{
\link[tiff:readTIFF]{tiff::readTIFF}, \linkS4class{RLum.Data.Image}
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2022. read_TIFF2R(): Import TIFF Image Data into R. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2022. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.19. https://CRAN.R-project.org/package=Luminescence
}

\keyword{IO}
