% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TIFF2R.R
\name{read_TIFF2R}
\alias{read_TIFF2R}
\title{Import TIFF Image Data into R}
\usage{
read_TIFF2R(file)
}
\arguments{
\item{file}{[character] (**required**): file name}
}
\value{
[RLum.Image.Data-class] object
}
\description{
Simple wrapper around [tiff::readTIFF] to import TIFF images
and TIFF image stacks to be further processed within the package 'Luminescence'
}
\section{Function version}{
 0.1.0
}

\examples{

\dontrun{
file <- file.choose()
image <- read_TIFF2R(file)

}

}
\seealso{
[tiff::readTIFF], [RLum.Data.Image-class], [raster::raster]
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2022. read_TIFF2R(): Import TIFF Image Data into R. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2022. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.17. https://CRAN.R-project.org/package=Luminescence
}

\keyword{IO}
