% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R
\name{remove_RLum}
\alias{remove_RLum}
\alias{remove_RLum,list-method}
\alias{remove_RLum,RLum.Analysis-method}
\title{Strips records from RLum-class objects}
\usage{
remove_RLum(object, ...)

\S4method{remove_RLum}{list}(object, ...)

\S4method{remove_RLum}{RLum.Analysis}(object, ...)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
object with records to be removed.}

\item{...}{parameters to be passed to \link{get_RLum}. Only the
following named arguments are supported: \code{record.id}, \code{recordType},
\code{curveType}, \code{RLum.type}, \code{protocol}, \code{info.object} and \code{subset}.
Arguments \code{get.index} and \code{drop} are preset to \code{TRUE} and \code{FALSE},
respectively, and cannot be overridden. Other named arguments are ignored,
while unnamed arguments are not supported.}
}
\value{
An object of the same type as the input provided with records removed; it
can result in an empty object.
}
\description{
Remove records from an RLum-class object in a convenient way using
\link{get_RLum} for the selection.
}
\section{Methods (by class)}{
\itemize{
\item \code{remove_RLum(list)}: Returns a list of \link[=RLum-class]{RLum} objects where the selected records are stripped

\item \code{remove_RLum(RLum.Analysis)}: Method to remove records from an \link[=RLum.Analysis-class]{RLum.Analysis} object.

}}
\section{Function version}{
 0.1.0
}

\examples{
## load example data
data(ExampleData.XSYG, envir = environment())
sar <- OSL.SARMeasurement$Sequence.Object[1:9]

## strop only OSL curves
sar <- remove_RLum(sar, recordType = "OSL")
sar

}
\seealso{
\link[=RLum.Analysis-class]{RLum.Analysis}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2026. remove_RLum(): Strips records from RLum-class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{utilities}
