% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R, R/RLum-class.R
\name{replicate_RLum}
\alias{replicate_RLum}
\alias{replicate_RLum,RLum-method}
\title{General replication function for RLum-class objects}
\usage{
replicate_RLum(object, times = 1)

\S4method{replicate_RLum}{RLum}(object, times = 1)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
an \link[=RLum-class]{RLum} object}

\item{times}{\link{integer} (\emph{optional}):
number for times each element should be repeated.}
}
\value{
A \link{list} with the object repeated.
}
\description{
The function replicates \link[=RLum-class]{RLum} objects and returns a list of such
objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{replicate_RLum(RLum)}: Replication method for \link[=RLum-class]{RLum} objects.

}}
\section{Function version}{
 0.1.0
}

\seealso{
\link[=RLum-class]{RLum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2026. replicate_RLum(): General replication function for RLum-class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{utilities}
