% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R, R/RLum-class.R
\name{replicate_RLum}
\alias{replicate_RLum}
\alias{replicate_RLum,RLum-method}
\title{General replication function for RLum-class objects}
\usage{
replicate_RLum(object, times = 1)

\S4method{replicate_RLum}{RLum}(object, times = 1)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
an \link[=RLum-class]{RLum} object}

\item{times}{\link{integer} (\emph{optional}):
number for times each element should be repeated.}
}
\value{
A \link{list} with the object repeated.
}
\description{
The function replicates \link[=RLum-class]{RLum} objects and returns a list of such
objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{replicate_RLum(RLum)}: Replication method for \link[=RLum-class]{RLum} objects.

}}
\section{Function version}{
 0.1.0
}

\seealso{
\link[=RLum-class]{RLum}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
}
\keyword{utilities}
