% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R, R/RLum-class.R
\name{replicate_RLum}
\alias{replicate_RLum}
\alias{replicate_RLum,RLum-method}
\title{General replication function for RLum-class objects}
\usage{
replicate_RLum(object, times = 1)

\S4method{replicate_RLum}{RLum}(object, times = 1)
}
\arguments{
\item{object}{\linkS4class{RLum} (\strong{required}):
an \linkS4class{RLum} object}

\item{times}{\link{integer} (\emph{optional}):
number for times each element should be repeated.}
}
\value{
A \link{list} with the object repeated.
}
\description{
The function replicates \linkS4class{RLum} objects and returns a list of such
objects.
}
\section{Functions}{
\itemize{
\item \code{replicate_RLum(RLum)}: Replication method for \linkS4class{RLum} objects.

}}
\section{Function version}{
 0.1.0
}

\seealso{
\linkS4class{RLum}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{utilities}
