% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R
\name{replicate_RLum}
\alias{replicate_RLum}
\title{General replication function for RLum S4 class objects}
\usage{
replicate_RLum(object, times = NULL)
}
\arguments{
\item{object}{\linkS4class{RLum} (\strong{required}):
an \linkS4class{RLum} object}

\item{times}{\link{integer} (\emph{optional}):
number for times each element is repeated element}
}
\value{
A \link{list} with the object repeated.
}
\description{
The function replicates RLum S4 class objects and returns a list of such
objects.
}
\section{Function version}{
 0.1.0
}

\seealso{
\linkS4class{RLum}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{utilities}
