% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_RLum.R
\name{replicate_RLum}
\alias{replicate_RLum}
\title{General replication function for RLum S4 class objects}
\usage{
replicate_RLum(object, times = NULL)
}
\arguments{
\item{object}{\linkS4class{RLum} (\strong{required}):
an \linkS4class{RLum} object}

\item{times}{\link{integer} (\emph{optional}):
number for times each element is repeated element}
}
\value{
Returns a \link{list} of the object to be repeated
}
\description{
Function replicates RLum S4 class objects and returns a list for this objects
}
\section{Function version}{
 0.1.0 (2018-01-21 17:22:38)
}

\seealso{
\linkS4class{RLum}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team} 

\section{How to cite}{
Kreutzer, S. (2018). replicate_RLum(): General replication function for RLum S4 class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.2. https://CRAN.R-project.org/package=Luminescence
}

\keyword{utilities}
