% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_RLum.R
\name{replicate_RLum}
\alias{replicate_RLum}
\title{General replication function for RLum S4 class objects}
\usage{
replicate_RLum(object, times = NULL)
}
\arguments{
\item{object}{\linkS4class{RLum} (\strong{required}):
an \linkS4class{RLum} object}

\item{times}{\link{integer} (\emph{optional}):
number for times each element is repeated element}
}
\value{
Returns a \link{list} of the object to be repeated
}
\description{
Function replicates RLum S4 class objects and returns a list for this objects
}
\section{Function version}{
 0.1.0
}

\seealso{
\linkS4class{RLum}
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. replicate_RLum(): General replication function for RLum S4 class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Riedesel, S., Autzen, M., Mittelstrass, D., 2020. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.10. https://CRAN.R-project.org/package=Luminescence
}

\keyword{utilities}
