% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_RLum.R
\name{report_RLum}
\alias{report_RLum}
\title{Create HTML report of (RLum) objects}
\usage{
report_RLum(object, file, timestamp = TRUE, clean = TRUE, ...)
}
\arguments{
\item{object}{an object}

\item{file}{filename}

\item{timestamp}{prefix for the filename}

\item{clean}{remove intermediate files}

\item{...}{currently not used}
}
\value{
Writes a HTML and .Rds file.
}
\description{
This function creates a HTML report for a given object, listing its complete
structure and content. The object itself is saved as a serialised .Rds file.
}
\details{
Provide some details here.
}
\section{Function version}{
 0.1.0
}
\examples{
# coming soon
}
\author{
Christoph Burow, University of Cologne (Germany) \cr
}
\seealso{
\code{\link[rmarkdown]{render}}
}

