% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risoe.BINfileData2RLum.Analysis.R
\name{Risoe.BINfileData2RLum.Analysis}
\alias{Risoe.BINfileData2RLum.Analysis}
\title{Convert Risoe.BINfileData object to an RLum.Analysis object}
\usage{
Risoe.BINfileData2RLum.Analysis(object, pos = NULL, grain = NULL,
  run = NULL, set = NULL, ltype = NULL, dtype = NULL,
  protocol = "unknown", keep.empty = TRUE, txtProgressBar = FALSE)
}
\arguments{
\item{object}{[Risoe.BINfileData-class] (\bold{required}):
\code{Risoe.BINfileData} object}

\item{pos}{[numeric] (optional): position number of the
\code{Risoe.BINfileData} object for which the curves are stored in the
\code{RLum.Analysis} object. If \code{length(position)>1} a list of \code{RLum.Analysis} objects
is returned. If nothing is provided every position will be converted. If the position is not valid \code{NA} is
returned.}

\item{grain}{[vector}, \link{numeric] (optional): grain number from
the measurement to limit the converted data set (e.g., \code{grain =
c(1:48)}). Please be aware that this option may lead to unwanted effects, as the output
is strictly limited to the choosen grain number for all position numbers}

\item{run}{[vector}, \link{numeric] (optional): run number from
the measurement to limit the converted data set (e.g., \code{run =
c(1:48)}).}

\item{set}{[vector}, \link{numeric] (optional): set number from
the measurement to limit the converted data set (e.g., \code{set =
c(1:48)}).}

\item{ltype}{[vector}, \link{character] (optional): curve type
to limit the converted data. Commonly allowed values are: \code{IRSL}, \code{OSL},
\code{TL}, \code{RIR}, \code{RBR} and \code{USER} (see also [Risoe.BINfileData-class])}

\item{dtype}{[vector}, \link{character] (optional): data type to
limit the converted data. Commonly allowed values are listed in [Risoe.BINfileData-class]}

\item{protocol}{[character] (optional): sets protocol type for
analysis object. Value may be used by subsequent analysis functions.}

\item{keep.empty}{[logical] (with default): If \code{TRUE} (default)
an \code{RLum.Analysis} object is returned even if it does not contain any
records. Set to \code{FALSE} to discard all empty objects.}

\item{txtProgressBar}{\link{logical} (with default): enables or disables
[txtProgressBar].}
}
\value{
Returns an [RLum.Analysis-class] object.
}
\description{
Converts values from one specific position of a Risoe.BINfileData S4-class
object to an RLum.Analysis object.
}
\details{
The [RLum.Analysis-class] object requires a set of curves for
specific further protocol analyses. However, the
[Risoe.BINfileData-class] usually contains a set of curves for
different aliquots and different protocol types that may be mixed up.
Therefore, a conversion is needed.
}
\note{
The \code{protocol} argument of the [RLum.Analysis-class]
object is set to 'unknown' if not stated otherwise.
}
\section{Function version}{
 0.4.2
}

\examples{

##load data
data(ExampleData.BINfileData, envir = environment())

##convert values for position 1
Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos = 1)

}
\seealso{
[Risoe.BINfileData-class], [RLum.Analysis-class], [read_BIN2R]
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
}
\keyword{manip}
