% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risoe.BINfileData2RLum.Analysis.R
\name{Risoe.BINfileData2RLum.Analysis}
\alias{Risoe.BINfileData2RLum.Analysis}
\title{Convert Risoe.BINfileData object to an RLum.Analysis object}
\usage{
Risoe.BINfileData2RLum.Analysis(object, pos, grain, run, set, ltype,
  protocol = "unknown", txtProgressBar = FALSE)
}
\arguments{
\item{object}{\code{\linkS4class{Risoe.BINfileData}} (\bold{required}):
\code{Risoe.BINfileData} object}

\item{pos}{\code{\link{numeric}} (optional): position number of the
\code{Risoe.BINfileData} object for which the curves are stored in the
\code{RLum.Analysis} object. If \code{length(position)>1} a list of \code{RLum.Analysis} objects
is returned. If nothing is provided every position will be converted. If the position is not valid \code{NA} is
returned.}

\item{grain}{\code{\link{vector}, \link{numeric}} (optional): grain number from
the measurement to limit the converted data set (e.g., \code{grain =
c(1:48)}). Please be aware that this option may lead to unwanted effects, as the output
is strictly limited to the choosen grain number for all position numbers}

\item{run}{\code{\link{vector}, \link{numeric}} (optional): run number from
the measurement to limit the converted data set (e.g., \code{run =
c(1:48)}).}

\item{set}{\code{\link{vector}, \link{numeric}} (optional): set number from
the measurement to limit the converted data set (e.g., \code{set =
c(1:48)}).}

\item{ltype}{\code{\link{vector}, \link{character}} (optional): curve type
to limit the converted data. Allowed values are: \code{IRSL}, \code{OSL},
\code{TL}, \code{RIR}, \code{RBR} and \code{USER}}

\item{protocol}{\code{\link{character}} (optional): sets protocol type for
analysis object. Value may be used by subsequent analysis functions.}

\item{txtProgressBar}{\link{logical} (with default): enables or disables
\code{\link{txtProgressBar}}.}
}
\value{
Returns an \code{\linkS4class{RLum.Analysis}} object.
}
\description{
Converts values from one specific position of a Risoe.BINfileData S4-class
object to an RLum.Analysis object. Internally the function uses the function
\code{\link{Risoe.BINfileData2RLum.Data.Curve}} to recalculate the curves.
}
\details{
The \code{\linkS4class{RLum.Analysis}} object requires a set of curves for
specific further protocol analyses. However, the
\code{\linkS4class{Risoe.BINfileData}} usually contains a set of curves for
different aliquots and different protocol types that may be mixed up.
Therefore, a conversion is needed.
}
\note{
The \code{protocol} argument of the \code{\linkS4class{RLum.Analysis}}
object is set to 'unknown' if not stated otherwise.
}
\section{Function version}{
 0.2.2 (2015-12-05 15:52:49)
}
\examples{

##load data
data(ExampleData.BINfileData, envir = environment())

##convert values for position 1
Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos = 1)

}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
\cr R Luminescence Package Team}
\references{
#
}
\seealso{
\code{\linkS4class{Risoe.BINfileData}}, \code{\link{Risoe.BINfileData2RLum.Data.Curve}}
\code{\linkS4class{RLum.Analysis}}, \code{\link{read_BIN2R}}
}
\keyword{manip}

