% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\docType{class}
\name{RLum-class}
\alias{RLum-class}
\title{Class \code{"RLum"}}
\description{
Abstract class for data in the package Luminescence
}
\note{
\code{RLum} is a virtual class.
}
\section{Objects from the Class}{
 A virtual Class: No objects can be created
from it.
}
\examples{
showClass("RLum")
}
\author{
Sebastian Kreutzer, 2013 (Freiberg Instruments/JLU Giessen, Germany)
}
\references{
#
}
\seealso{
\code{\linkS4class{RLum.Data}}, \code{\linkS4class{RLum.Analysis}}
}
\keyword{classes}

