% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RLum-class.R
\docType{class}
\name{RLum-class}
\alias{RLum-class}
\title{Class \code{"RLum"}}
\description{
Abstract class for data in the package Luminescence
}
\note{
\code{RLum} is a virtual class.
}
\section{Objects from the Class}{
 A virtual Class: No objects can be created
from it.
}

\section{Class version}{
 0.2.0
}
\examples{
showClass("RLum")
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
}
\seealso{
\code{\linkS4class{RLum.Data}}, \code{\linkS4class{RLum.Analysis}}
}
\keyword{classes}

