\name{RLum-class}
\Rdversion{1.1}
\docType{class}
\alias{RLum-class}

\title{Class \code{"RLum"}}
\description{
Abstract class for data in the package Luminescence
}
\section{Objects from the Class}{A virtual Class: No objects can be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"}}
  }
}
\section{Methods}{
No methods defined with class "RLum" in the signature.
}
\references{
#
}
\author{Sebastian Kreutzer, 2013 (Freiberg Instruments/JLU Giessen, Germany)
}

\section{Version}{0.1 (2013-01-18)}

\note{
\code{RLum} is a virtual class.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\linkS4class{RLum.Data}}, \code{\linkS4class{RLum.Analysis}}
}
\examples{
showClass("RLum")
}
\keyword{classes}
