% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RLum.Analysis-class.R
\docType{class}
\name{RLum.Analysis-class}
\alias{RLum.Analysis-class}
\title{Class \code{"RLum.Analysis"}}
\description{
Object class containing analysis data for protocol analysis.
}
\note{
The method \code{get_structure.RLum.Analysis} is currently just
avaiblable for objects containing \code{\linkS4class{RLum.Data.Curve}}.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RLum.Analysis", ...)}.
}
\examples{
showClass("RLum.Analysis")

## usage of get_RLum.Analysis() with returning an RLum.Analysis object
#  get_RLum.Analysis(object, keep.object = TRUE)
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
}
\seealso{
\code{\link{Risoe.BINfileData2RLum.Analysis}},
\code{\linkS4class{Risoe.BINfileData}}, \code{\linkS4class{RLum}}
}

