% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.Data-class.R
\docType{class}
\name{RLum.Data-class}
\alias{RLum.Data-class}
\title{Class \code{"RLum.Data"}}
\description{
Generalized virtual data class for luminescence data.
}
\note{
Just a virtual class.
}
\section{Objects from the Class}{
 A virtual Class: No objects can be created
from it.
}

\section{Class version}{
 0.2.1
}

\examples{

showClass("RLum.Data")

}
\seealso{
[RLum-class], [RLum.Data.Curve-class],
[RLum.Data.Spectrum-class]
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
}
\keyword{classes}
