% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\docType{class}
\name{RLum.Data-class}
\alias{RLum.Data-class}
\title{Class \code{"RLum.Data"}}
\description{
Generalized virtual data class for luminescence data.
}
\note{
Just a virtual class.
}
\section{Objects from the Class}{
 A virtual Class: No objects can be created
from it.
}
\examples{
showClass("RLum.Data")
}
\author{
Sebastian Kreutzer, 2013 (Freiberg Instruments/JLU Giessen, Germany)
}
\references{
#
}
\seealso{
\code{\linkS4class{RLum}}, \code{\linkS4class{RLum.Data.Curve}},
\code{\linkS4class{RLum.Data.Spectrum}}
}
\keyword{classes}

