\name{RLum.Data-class}
\Rdversion{1.1}
\docType{class}
\alias{RLum.Data-class}

\title{Class \code{"RLum.Data"}}
\description{
Generalized virtual data class for luminescence data.
}
\section{Objects from the Class}{A virtual Class: No objects can be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"}}
  }
}
\section{Methods}{
No methods defined with class "RLum.Data" in the signature.
}
\references{
#
}
\author{Sebastian Kreutzer, 2013 (Freiberg Instruments/JLU Giessen, Germany)
}

\section{Version}{0.1 (2013-01-18)}

\note{
Just a virtual class. 
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\linkS4class{RLum}}, 
\code{\linkS4class{RLum.Data.Curve}}, 
\code{\linkS4class{RLum.Data.Spectrum}}
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("RLum.Data")
}
\keyword{classes}
