% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RLum.Data.Curve-class.R
\docType{class}
\name{RLum.Data.Curve-class}
\alias{RLum.Data.Curve-class}
\title{Class \code{"RLum.Data.Curve"}}
\description{
Class for luminescence curve data.
}
\note{
The class should only contain data for a single curve. For additional
elements the slot \code{info} can be used (e.g. providing additional heating
ramp curve).
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RLum.Data.Curve", ...)}.
}
\examples{
showClass("RLum.Data.Curve")
}
\author{
Sebastian Kreutzer Freiberg Instruments/JLU Giessen (Germany)
}
\references{
#
}
\seealso{
\code{\linkS4class{RLum}}, \code{\linkS4class{RLum.Data}},
\code{\link{plot_RLum}}
}
\keyword{classes}

