% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RLum.Data.Image-class.R
\docType{class}
\name{RLum.Data.Image-class}
\alias{RLum.Data.Image-class}
\title{Class \code{"RLum.Data.Image"}}
\description{
Class for luminescence image data (TL/OSL/RF).
}
\note{
The class should only contain data for a set of images. For additional
elements the slot \code{info} can be used.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RLum.Data.Image", ...)}.
}
\examples{
showClass("RLum.Data.Image")

##so far no further example available
}
\author{
Sebastian Kreutzer, Universite Bordeaux Montaigne (France)
}
\references{
#
}
\seealso{
\code{\linkS4class{RLum}}, \code{\linkS4class{RLum.Data}},
\code{\link{plot_RLum}}
}
\keyword{classes}

