% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RLum.Data.Spectrum-class.R
\docType{class}
\name{RLum.Data.Spectrum-class}
\alias{RLum.Data.Spectrum-class}
\title{Class \code{"RLum.Data.Spectrum"}}
\description{
Class for luminescence spectra data (TL/OSL/RF).
}
\note{
The class should only contain data for a single spectra data set. For
additional elements the slot \code{info} can be used.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RLum.Data.Spectrum", ...)}.
}
\examples{
showClass("RLum.Data.Spectrum")

##show example data (uncomment for usage)
# data(ExampleData.XSYG, envir = environment())
# TL.Spectrum
}
\author{
Sebastian Kreutzer, JLU Giessen (Germany)
}
\references{
#
}
\seealso{
\code{\linkS4class{RLum}}, \code{\linkS4class{RLum.Data}},
\code{\link{plot_RLum}}
}
\keyword{classes}

