% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RLum.Results-class.R
\docType{class}
\name{RLum.Results-class}
\alias{RLum.Results-class}
\alias{get_RLum,RLum.Results-method}
\alias{merge_RLum.Results,list-method}
\alias{set_RLum,RLum.Results-method}
\title{Class \code{"RLum.Results"}}
\usage{
\S4method{set_RLum}{RLum.Results}(class, originator, data)

\S4method{get_RLum}{RLum.Results}(object, data.object)

\S4method{merge_RLum.Results}{list}(object.list)
}
\arguments{
\item{class}{x}

\item{originator}{x}

\item{data}{x}

\item{object}{x}

\item{data.object}{name of the slot to be returned}

\item{object.list}{list}
}
\description{
Object class contains results data from functions.
}
\section{Methods (by generic)}{
\itemize{
\item \code{set_RLum}: Construction method for \code{RLum.Results} object.
The slot \code{originator} is optional and predefined as the function
that calls the function \code{set_RLum.Results}.

\item \code{get_RLum}: The argument \code{data.object} allows directly accessing
objects delivered within the slot \code{data}. If no \code{data.object} is
specified, a preselected object is returned. The default return
object depends on the object originator (e.g. \code{fit_LMCurve})

\item \code{merge_RLum.Results}: The argument \code{data.object} allows directly accessing
objects delivered within the slot \code{data}. If no \code{data.object} is
specified, a preselected object is returned. The default return
object depends on the object originator (e.g. \code{fit_LMCurve})
}}
\note{
The class is intended to store results from functions to be used by
other functions. The data in the object should always be accessed by the
method \code{get_RLum.Results}.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RLum.Results", ...)}.
}
\examples{
showClass("RLum.Results")
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
}
\seealso{
\code{\linkS4class{RLum}}
}
\keyword{classes}
\keyword{methods}

