% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Luminescence-package.R
\docType{class}
\name{RLum.Results-class}
\alias{RLum.Results-class}
\alias{get_RLum.Results}
\alias{get_RLum.Results,RLum.Results-method}
\alias{merge_RLum.Results}
\alias{merge_RLum.Results,list-method}
\alias{merge_RLum.Results-methods}
\alias{set_RLum.Results}
\alias{set_RLum.Results,ANY,list-method}
\alias{set_RLum.Results,RLum.Results-method}
\alias{show,RLum.Results-method}
\alias{validObject,RLum.Results-method}
\title{Class \code{"RLum.Results"}}
\description{
Object class contains results data from functions.
}
\note{
The class is intended to store results from functions to be used by
other functions. The data in the object should always be accessed by the
method \code{get_RLum.Results}.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("RLum.Results", ...)}.
}
\examples{
showClass("RLum.Results")
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
}
\references{
#
}
\seealso{
\code{\linkS4class{RLum}}
}
\keyword{classes}
\keyword{methods}

