% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_GammaDose.R
\name{scale_GammaDose}
\alias{scale_GammaDose}
\title{A short title}
\usage{
scale_GammaDose(data, conversion_factors = c("Guerinetal2011",
  "AdamiecAitken1998", "Liritzisetal2013")[1],
  fractional_gamma_dose = c("Aitken1985")[1], verbose = TRUE,
  plot = TRUE, plot_single = TRUE, ...)
}
\arguments{
\item{data}{\link{data.frame} (\strong{required}):
A table containing all relevant information for each individual layer. The
table must have the following named columns:
\itemize{
\item \code{id} (\link{character}): an arbitrary id or name of each layer
\item \code{thickness} (\link{numeric}): vertical extent of each layer in cm
\item \code{sample_offset} (\link{logical}): distance of the sample in cm,
\strong{measured from the BOTTOM OF THE TARGET LAYER}. Except for the target layer
all values must be \code{NA}.
\item \code{K} (\link{numeric}): K nuclide content in %
\item \code{K_se} (\link{numeric}): error on the K content
\item \code{Th} (\link{numeric}): Th nuclide content in ppm
\item \code{Th_se} (\link{numeric}): error on the Th content
\item \code{U} (\link{numeric}): U nuclide content in ppm
\item \code{U_se} (\link{numeric}): error on the U content
\item \code{water_content} (\link{numeric}): water content of each layer in %
\item \code{water_content_se} (\link{numeric}):  error on the water content
\item \code{density} (\link{numeric}): bulk density of each layer in g/cm^-3
}}

\item{conversion_factors}{\link{logical} (\emph{optional}):
\itemize{
\item \code{"Guerinetal2011"} (default)
\item \code{"AdamiecAitken1998"}
\item \code{"Liritzisetal2013"}
}}

\item{fractional_gamma_dose}{\link{logical} (\emph{optional}):
\itemize{
\item \code{"Aitken1985"} (default): Table H1 in the appendix
}}

\item{verbose}{\link{logical} (\emph{optional}):
Show or hide console output (defaults to \code{TRUE}).}

\item{plot}{\link{logical} (\emph{optional}):
Show or hide the plot (defaults to \code{TRUE}).}

\item{plot_single}{\link{logical} (\emph{optional}):
Show all plots in one panel (defaults to \code{TRUE}).}

\item{...}{Further parameters passed to \link{barplot}.}
}
\value{
Function returns results numerically and graphically:

-----------------------------------\cr
\code{[ NUMERICAL OUTPUT ]}\cr
-----------------------------------\cr

\strong{\code{RLum.Results}}-object

\strong{slot:} \strong{\code{@data}}

\tabular{lll}{
\strong{Element} \tab \strong{Type} \tab \strong{Description}\cr
\code{$summary} \tab \code{data.frame} \tab summary of the model results \cr
\code{$data} \tab \code{data.frame} \tab the original input data \cr
\code{$dose_rates} \tab \code{list} \tab two \code{data.frames} for the scaled and infinite matrix dose rates \cr
\code{$tables} \tab \code{list} \tab several \code{data.frames} containing intermediate results \cr
\code{$args} \tab \code{character} \tab arguments of the call \cr
\code{$call} \tab \code{call} \tab the original function call \cr
}

\strong{slot:} \strong{\code{@info}}

Currently unused.

------------------------\cr
\code{[ PLOT OUTPUT ]}\cr
------------------------\cr

A barplot visualising the contribution of each layer to the total dose rate
received by the sample in the target layer.
}
\description{
Two or three sentence description
}
\details{
An exhaustive explanation on the methodology
}
\note{
\strong{This function has BETA status. If possible, results should be}
\strong{cross-checked.}
}
\section{Function version}{
 0.1.0 (2018-08-22 19:00:29)
}

\examples{

# Load example data
data("ExampleData.ScaleGammaDose", envir = environment())
x <- ExampleData.ScaleGammaDose

# Scale gamma dose rate
results <- scale_GammaDose(data = x,
                           conversion_factors = "Liritzisetal2013",
                           fractional_gamma_dose = "Aitken1985",
                           verbose = TRUE,
                           plot = TRUE)

get_RLum(results)


} 

\section{How to cite}{
Riedesel, S., Autzen, M., Burow, C. (2018). scale_GammaDose(): A short title. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0. https://CRAN.R-project.org/package=Luminescence
}

\references{
Guerin, G., Mercier, N., Adamiec, G., 2011. Dose-rate conversion
factors: update. Ancient TL, 29, 5-8.

Adamiec, G., Aitken, M.J., 1998. Dose-rate conversion factors: update.
Ancient TL 16, 37-46.

Liritzis, I., Stamoulis, K., Papachristodoulou, C., Ioannides, K., 2013.
A re-evaluation of radiation dose-rate conversion factors. Mediterranean
Archaeology and Archaeometry 13, 1-15.

Aitken, M.J., 1985. Thermoluminescence Dating. Academic Press, London.
}
\seealso{
\link{ExampleData.ScaleGammaDose}, \link{approx}, \link{barplot}
}
\author{
Svenja Riedesel, Aberystwyth University (United Kingdom) \cr
Martin Autzen, DTU NUTECH Center for Nuclear Technologies (Denmark) \cr
Christoph Burow, University of Cologne (Germany) \cr
Based on an excel spreadsheet and accompanying macro written by Ian Bailiff.
\cr R Luminescence Package Team}
\keyword{datagen}
