% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Second2Gray.R
\name{Second2Gray}
\alias{Second2Gray}
\title{Converting equivalent dose values from seconds (s) to Gray (Gy)}
\usage{
Second2Gray(...)
}
\arguments{
\item{...}{Unused.}
}
\description{
This function is defunct, use \link{convert_Second2Gray} instead.
}
