% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Second2Gray.R
\name{Second2Gray}
\alias{Second2Gray}
\title{Converting equivalent dose values from seconds (s) to gray (Gy)}
\usage{
Second2Gray(data, dose.rate, error.propagation = "omit")
}
\arguments{
\item{data}{\code{\link{data.frame}} (\bold{required}): input values,
structure: data (\code{values[,1]}) and data error (\code{values [,2]}) are
required}

\item{dose.rate}{\code{\linkS4class{RLum.Results}} or \code{\link{data.frame}} or \code{\link{numeric}}
(\bold{required}): \code{RLum.Results} needs to be orginated from the
function \code{\link{calc_SourceDoseRate}}, for \code{vector} dose rate in
Gy/s and dose rate error in Gy/s}

\item{error.propagation}{\code{\link{character}} (with default): error propagation method used for error
calculation (\code{omit}, \code{gaussian} or \code{absolute}), see details for further
information}
}
\value{
Returns a \link{data.frame} with converted values.
}
\description{
Conversion of absorbed radiation dose in seconds (s) to the SI unit gray
(Gy) including error propagation. Normally used for equivalent dose data.
}
\details{
Calculation of De values from seconds (s) to gray (Gy) \deqn{De [Gy] = De
[s] * Dose Rate [Gy/s])} \cr

Provided calculation error propagation methods for error calculation (with 'se' as the standard error
and 'DR' of the dose rate of the beta-source):\cr

\bold{(1) \code{omit}} (default)\cr

\deqn{se(De) [Gy] = se(De) [s] * DR [Gy/s]}

In this case the standard error of the dose rate of the beta-source is treated as systematic
(i.e. non-random), it error propagation is omitted. However, the error must be considered during
calculation of the final age. (cf. Aitken, 1985, pp. 242). This approach can be seen as
method (2) (gaussian) for the case the (random) standard error of the beta-source calibration is
0. Which particular method is requested depends on the situation and cannot be prescriptive.

\bold{(2) \code{gaussian}} error propagation \cr

\deqn{se(De) [Gy] = \sqrt((DR [Gy/s] * se(De) [s])^2 + (De [s] * se(DR) [Gy/s])^2)}

Applicable under the assumption that errors of De and se are uncorrelated.

\bold{(3) \code{absolute}} error propagation \cr

\deqn{se(De) [Gy]= abs(DR [Gy/s] * se(De) [s]) + abs(De [s] * se(DR) [Gy/s])}

Applicable under the assumption that errors of De and se are not uncorrelated.
}
\note{
If no or a wrong error propagation method is given, the execution of the function is
stopped. Furthermore, if a \code{data.frame} is provided for the dose rate values is has to
be of the same length as the data frame provided with the argument \code{data}
}
\section{Function version}{
 0.6.0 (2015-11-29 17:27:48)
}

\examples{


##(A) for known source dose rate at date of measurement
## - load De data from the example data help file
data(ExampleData.DeValues, envir = environment())
## - convert De(s) to De(Gy)
Second2Gray(ExampleData.DeValues$BT998, c(0.0438,0.0019))





##(B) for source dose rate calibration data
## - calculate source dose rate first
dose.rate <-  calc_SourceDoseRate(measurement.date = "2012-01-27",
                                  calib.date = "2014-12-19",
                                  calib.dose.rate = 0.0438,
                                  calib.error = 0.0019)
# read example data
data(ExampleData.DeValues, envir = environment())

# apply dose.rate to convert De(s) to De(Gy)
Second2Gray(ExampleData.DeValues$BT998, dose.rate)

} 

\section{How to cite}{
Kreutzer, S., Dietze, M., Fuchs, M.C., Fuchs, M. (2017). Second2Gray(): Converting equivalent dose values from seconds (s) to gray (Gy). Function version 0.6.0. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.7.3. https://CRAN.R-project.org/package=Luminescence
}

\references{
Aitken, M.J., 1985. Thermoluminescence dating. Academic Press.
}
\seealso{
\code{\link{calc_SourceDoseRate}}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France),\cr Michael Dietze, GFZ Potsdam (Germany),\cr Margret C. Fuchs, HZDR,
Helmholtz-Institute Freiberg for Resource Technology
(Germany)
\cr R Luminescence Package Team}
\keyword{manip}
