\name{set_Risoe.BINfileData}
\alias{set_Risoe.BINfileData}
\title{set Risoe BINfileData}

\usage{set_Risoe.BINfileData(METADATA, DATA)}
\arguments{
  \item{METADATA}{
}
  \item{DATA}{
}
}



\author{Sebastian Kreutzer [aut, trl, cre],
Michael Dietze [aut],
Christoph Burow [aut],
Margret C. Fuchs [aut],
Christoph Schmidt [ctb],
Manfred Fischer [aut, trl],
Markus Fuchs [ths]}





