% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_Risoe.BINfileData.R
\name{set_Risoe.BINfileData}
\alias{set_Risoe.BINfileData}
\title{General accessor function for RLum S4 class objects}
\usage{
set_Risoe.BINfileData(METADATA, DATA, .RESERVED)
}
\arguments{
\item{METADATA}{x}

\item{DATA}{x}

\item{.RESERVED}{x}
}
\value{
Return is the same as input objects as provided in the list.
}
\description{
Function calls object-specific get functions for RisoeBINfileData S4 class objects.
}
\details{
The function provides a generalised access point for specific
\code{\linkS4class{Risoe.BINfileData}} objects.\cr Depending on the input object, the
corresponding get function will be selected. Allowed arguments can be found
in the documentations of the corresponding \code{\linkS4class{Risoe.BINfileData}} class.
}
\section{Function version}{
 0.1 (2015-07-23 17:09:29)
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
\cr R Luminescence Package Team}
\seealso{
\code{\linkS4class{Risoe.BINfileData}}
}
\keyword{utilities}

