% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_RLum.R
\name{set_RLum}
\alias{set_RLum}
\alias{set_RLum.Analysis}
\alias{set_RLum.Data.Curve}
\alias{set_RLum.Data.Image}
\alias{set_RLum.Data.Spectrum}
\alias{set_RLum.Results}
\title{General accessor function for RLum S4 class objects}
\usage{
set_RLum(class, ...)
}
\arguments{
\item{class}{\code{\linkS4class{RLum}} (\bold{required}): name of the S4 class to
create}

\item{\dots}{further arguments that one might want to pass to the specific
get function}
}
\value{
Return is the same as input objects as provided in the list.
}
\description{
Function calls object-specific set functions for RLum S4 class objects.
}
\details{
The function provides a generalised access point for specific
\code{\linkS4class{RLum}} objects.\cr Depending on the input object, the
corresponding set function will be selected.  Allowed arguments can be found
in the documentations of the corresponding \code{\linkS4class{RLum}} class.
}
\section{Function version}{
 0.1 (2015-09-21 15:05:38)
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
\cr R Luminescence Package Team}
\seealso{
\code{\linkS4class{RLum.Data.Curve}},
\code{\linkS4class{RLum.Data.Image}},
\code{\linkS4class{RLum.Data.Spectrum}},
\code{\linkS4class{RLum.Analysis}},
\code{\linkS4class{RLum.Results}}
}
\keyword{utilities}

