\name{sTeve}
\alias{sTeve}
\title{
sTeve - sophisticated Tool for efficient data validation and evaluation
}
\description{
Function provides a sophisticated routine for comprehensive luminescence dating 
data analysis}
\usage{
sTeve (n_frames = 20, t_animation = 3)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n_frames}{\code{\link{integer}} (with default): n frames
}
  \item{t_animation}{\code{\link{integer}} (with default): t animation
}
}
\details{
This amazing sophisticated function validates your data seriously. 
}
\value{
Validates your data.
}
\references{
# 
}
\author{
R Luminescence Team, Dec. 2012 
}
\note{
The function should not be taken too seriously. 
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\link{plot_KDE}
}
\examples{
##no example available
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
}
\keyword{manip}
