% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sTeve}
\alias{sTeve}
\title{sTeve - sophisticated tool for efficient data validation and evaluation}
\usage{
sTeve(n_frames = 10, t_animation = 2, n.tree = 7, type)
}
\arguments{
\item{n_frames}{\link{integer} (\emph{with default}): n frames}

\item{t_animation}{\link{integer} (\emph{with default}): t animation}

\item{n.tree}{\link{integer} (\emph{with default}): How many trees do you
want to cut?}

\item{type}{\link{integer} (\emph{optional}): Make a decision: 1, 2 or 3}
}
\value{
Validates your data.
}
\description{
This function provides a sophisticated routine for comprehensive
luminescence dating data analysis.
}
\details{
This amazing sophisticated function validates your data seriously.
}
\note{
This function should not be taken too seriously.
}
\examples{

##no example available

}
\seealso{
\link{plot_KDE}
}
\author{
R Luminescence Team, 2012-2013
}
\keyword{manip}
