\name{sTeve}
\alias{sTeve}
\title{
sTeve - sophisticated tool for efficient data validation and evaluation
}
\description{
This function provides a sophisticated routine for comprehensive luminescence dating 
data analysis.}
\usage{
sTeve (n_frames = 10, t_animation = 2, n.tree = 7, type)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n_frames}{\code{\link{integer}} (with default): n frames
}
  \item{t_animation}{\code{\link{integer}} (with default): t animation
}
	\item{n.tree}{\code{\link{integer}} (with default): How many trees do 
	you want to cut?
}
	\item{type}{\code{\link{integer}} (optional): Make a decision: 1, 2 or 3
}
}

\details{
This amazing sophisticated function validates your data seriously. 
}
\value{
Validates your data.
}
\references{
# 
}
\author{R Luminescence Team, 2012-2013
}
\note{
This function should not be taken too seriously. 
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\link{plot_KDE}
}
\examples{
##no example available
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
}
\keyword{manip}
