% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sTeve}
\alias{sTeve}
\title{sTeve - sophisticated tool for efficient data validation and evaluation}
\usage{
sTeve(n_frames = 10, t_animation = 2, n.tree = 7, type)
}
\arguments{
\item{n_frames}{\code{\link{integer}} (with default): n frames}

\item{t_animation}{\code{\link{integer}} (with default): t animation}

\item{n.tree}{\code{\link{integer}} (with default): How many trees do you
want to cut?}

\item{type}{\code{\link{integer}} (optional): Make a decision: 1, 2 or 3}
}
\value{
Validates your data.
}
\description{
This function provides a sophisticated routine for comprehensive
luminescence dating data analysis.
}
\details{
This amazing sophisticated function validates your data seriously.
}
\note{
This function should not be taken too seriously.
}
\examples{

##no example available

} 

\section{How to cite}{
NA, NA, ,  (2017). sTeve(): sTeve - sophisticated tool for efficient data validation and evaluation. In: Kreutzer, S., Dietze, M., Burow, C., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2017). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.7.4. https://CRAN.R-project.org/package=Luminescence
}

\references{
#
}
\seealso{
\link{plot_KDE}
}
\author{
R Luminescence Team, 2012-2013
}
\keyword{manip}
