% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R
\name{structure_RLum}
\alias{structure_RLum}
\alias{structure_RLum,list-method}
\alias{structure_RLum,RLum.Analysis-method}
\title{General structure function for RLum-class objects}
\usage{
structure_RLum(object, ...)

\S4method{structure_RLum}{list}(object, ...)

\S4method{structure_RLum}{RLum.Analysis}(object, fullExtent = FALSE)
}
\arguments{
\item{object}{\link[=RLum-class]{RLum} (\strong{required}):
S4 object of class \code{RLum}}

\item{...}{further arguments passed to the specific class method}

\item{fullExtent}{\link{logical} (\emph{with default}):
extends the returned \code{data.frame} to its full extent, i.e. all info elements
are part of the return as well. The default value is \code{FALSE} as the data
frame might become rather big.}
}
\value{
Returns a \link{data.frame} with structure of the object.
}
\description{
The function provides a generalised access point for specific
\link[=RLum-class]{RLum} objects. Depending on the input object, the corresponding
function will be selected.
}
\section{Methods (by class)}{
\itemize{
\item \code{structure_RLum(list)}: Returns a list of data frames containing the structure of each \link[=RLum-class]{RLum}
object in the input list.

\item \code{structure_RLum(RLum.Analysis)}: Returns the structure of an \link[=RLum.Analysis-class]{RLum.Analysis} object.

}}
\section{Function version}{
 0.2.0
}

\examples{

## load example data
data(ExampleData.XSYG, envir = environment())

## show structure
structure_RLum(OSL.SARMeasurement$Sequence.Object)

}
\seealso{
\link[=RLum.Data.Curve-class]{RLum.Data.Curve}, \link[=RLum.Data.Image-class]{RLum.Data.Image},
\link[=RLum.Data.Spectrum-class]{RLum.Data.Spectrum}, \link[=RLum.Analysis-class]{RLum.Analysis},
\link[=RLum.Results-class]{RLum.Results}
}
\author{
Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2026. structure_RLum(): General structure function for RLum-class objects. Function version 0.2.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., de Boer, A., Bluszcz, A., 2026. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.3.1. https://r-lum.github.io/Luminescence/
}

\keyword{utilities}
