% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/template_DRAC.R
\name{template_DRAC}
\alias{template_DRAC}
\title{Create a DRAC input data template (v1.1)}
\usage{
template_DRAC(nrows = 1, notification = TRUE)
}
\arguments{
\item{nrows}{\code{\link{integer}} (with default): specifies the number of rows
of the template (i.e., the number of data sets you want to submit)}

\item{notification}{\code{\link{logical}} (with default): show or hide the
notification}
}
\value{
A list.
}
\description{
This function returns a DRAC input template (v1.1) to be used in conjunction
with the use_DRAC() function
}
\examples{
# create a new DRAC input template
template <- template_DRAC()

# show content of the template
print(template)
print(template$`Project ID`)
print(template[[4]])

# fill in data
template$
}

