% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/use_DRAC.R
\name{use_DRAC}
\alias{use_DRAC}
\title{Use DRAC to calculate dose rate data.}
\usage{
use_DRAC(file, name)
}
\arguments{
\item{file}{\code{\link{character}}: spreadsheet to be passed
to the DRAC website for calculation.}

\item{name}{\code{\link{character}}: Optional user name submitted to DRAC. If
omitted, a random name will be generated}
}
\value{
A results data set from the DRAC website.
}
\description{
The function provides an interface from R to DRAC. An R-object or a
pre-formatted XLS/XLSX file is passed to the DRAC website and the
results are re-imported into R.
}
\section{Function version}{
 0.1.0
}
\examples{
\dontrun{
## re-create DRAC example data set
text <- paste0("DRAC-example Quartz Q AdamiecAitken1998 3.4000 0.5100 14.4700 ",
               "1.6900 1.2000 0.1400 0.0000 0.0000 N X X X X X X X X X X X X ",
               "X X X X X N 90.0000 125.0000 Brennanetal1991 Guerinetal2012-Q ",
               "8.0000 10.0000 Bell1979 0.0000 0.0000 5.0000 2.0000 2.2000 ",
               "0.2200 1.8000 0.1000 30.0000 70.0000 150.0000 X X 20.0000 ",
               "0.2000\\n",
               "DRAC-example Feldspar F AdamiecAitken1998 2.0000 0.2000 8.0000 ",
               "0.4000 1.7500 0.0500 0.0000 0.0000 Y X X X X 12.5000 0.5000 X ",
               "X N X X X X X X X X Y 180.0000 212.0000 Bell1980 Mejdahl1979 ",
               "0.0000 0.0000 Bell1979 0.1500 0.0500 10.0000 3.0000 0.1500 ",
               "0.0150 1.8000 0.1000 60.0000 100.0000 200.0000 X X 15.0000 ",
               "1.5000\\n",
               "DRAC-example Polymineral PM AdamiecAitken1998 4.0000 0.4000 ",
               "12.0000 0.1200 0.8300 0.0800 0.0000 0.0000 Y X X X X 12.5000 ",
               "0.5000 X X N X X 2.5000 0.1500 X X X X Y 4.0000 11.0000 ",
               "Bell1980 Mejdahl1979 0.0000 0.0000 X 0.0860 0.0038 10.0000 ",
               "5.0000 X X X X X X X 0.2000 0.1000 204.4700 2.6900")

## send example data set let DRAC do the calculations
Reply <- use_DRAC(file = text)

# send the actual IO template spreadsheet to DRAC
file <-  "~/DRAC_Input_and_Output_Template-2.xlsx"
use_DRAC(file = file)
}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France), Michael Dietze,
GFZ Potsdam (Germany)\cr
}
\references{
Durcan, J.A., King, G.E., Duller, G.A.T., 2015. DRAC: Dose Rate and Age Calculator for trapped charge dating.
Quaternary Geochronology 28, 54-61. doi:10.1016/j.quageo.2015.03.012
}

