% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R,
%   R/RLum.Analysis-class.R, R/RLum.Data-class.R, R/RLum.Results-class.R
\name{view}
\alias{view}
\alias{view,RLum.Analysis-method}
\alias{view,RLum.Data-method}
\alias{view,RLum.Results-method}
\title{Convenience data visualisation function}
\usage{
view(object, ...)

\S4method{view}{RLum.Analysis}(object, ...)

\S4method{view}{RLum.Data}(object, ...)

\S4method{view}{RLum.Results}(object, element = 1, ...)
}
\arguments{
\item{object}{(\strong{required}) object to view}

\item{...}{further arguments passed to the specific class method}

\item{element}{\link{integer} (\emph{with default}):
index of the element to display.}
}
\value{
\code{NULL} and opens the data viewer.
}
\description{
Invokes the \link[utils:View]{utils::View} function tailored to objects in the package.
If started from RStudio, it uses the RStudio viewer.
}
\section{Functions}{
\itemize{
\item \code{view(RLum.Analysis)}: View method for \linkS4class{RLum.Analysis} objects.

\item \code{view(RLum.Data)}: View method for \linkS4class{RLum.Data} objects.

\item \code{view(RLum.Results)}: View method for \linkS4class{RLum.Results} objects.

}}
\seealso{
\code{\link[utils:View]{utils::View()}}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{utilities}
