% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luminescence-generics.R
\name{view}
\alias{view}
\title{Convenience data visualisation function}
\usage{
view(object, ...)
}
\arguments{
\item{object}{(\strong{required}) object to view}

\item{...}{further arguments passed to the specific class method}
}
\value{
\code{NULL} and opens the data viewer.
}
\description{
Invokes the \link[utils:View]{utils::View} function tailored to objects in the package.
If started from RStudio, it uses the RStudio viewer.
}
\seealso{
\code{\link[utils:View]{utils::View()}}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\keyword{utilities}
