% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_R2BIN.R
\name{write_R2BIN}
\alias{write_R2BIN}
\title{Export Risoe.BINfileData into Risoe BIN-file}
\usage{
write_R2BIN(object, file, version, compatibility.mode = FALSE,
  txtProgressBar = TRUE)
}
\arguments{
\item{object}{\linkS4class{Risoe.BINfileData} (\bold{required}):
input object to be stored in a bin file.}

\item{file}{\link{character} (\bold{required}): file name and path of
the output file\cr \code{[WIN]}: \code{write_R2BIN(object, "C:/Desktop/test.bin")},
\cr \code{[MAC/LINUX]}: \code{write_R2BIN("/User/test/Desktop/test.bin")}}

\item{version}{\link{character} (optional): version number for the
output file. If no value is provided the highest version number from the
\linkS4class{Risoe.BINfileData} is taken automatically.\cr\cr Note:
This argument can be used to convert BIN-file versions.}

\item{compatibility.mode}{\link{logical} (with default): this option
recalculates the position values if necessary and set the max. value to 48.
The old position number is appended as comment (e.g., 'OP: 70). This option
accounts for potential compatibility problems with the Analyst software.}

\item{txtProgressBar}{\link{logical} (with default): enables or disables
\link{txtProgressBar}.}
}
\value{
Write a binary file.
}
\description{
Exports a Risoe.BINfileData object in a *.bin or *.binx file that can be
opened by the Analyst software or other Risoe software.
}
\details{
The structure of the exported binary data follows the data structure
published in the Appendices of the Analyst manual p. 42.\cr\cr If
\code{LTYPE}, \code{DTYPE} and \code{LIGHTSOURCE} are not of type
\link{character}, no transformation into numeric values is done.
}
\note{
The function just roughly checks the data structures. The validity of
the output data depends on the user.\cr\cr The validity of the file path is
not further checked. \cr BIN-file conversions using the argument
\code{version} may be a lossy conversion, depending on the chosen input and
output data (e.g., conversion from version 08 to 07 to 06 to 04 or 03).\cr

\bold{Warning}\cr

Although the coding was done carefully it seems that the BIN/BINX-files
produced by Risoe DA 15/20 TL/OSL readers slightly differ on the byte level.
No obvious differences are observed in the METADATA, however, the
BIN/BINX-file may not fully compatible, at least not similar to the once
directly produced by the Risoe readers!\cr

ROI definitions (introduced in BIN-file version 8) are not supported! There are furthermore
ignored by the function \link{read_BIN2R}.
}
\section{Function version}{
 0.4.2
}

\examples{

##uncomment for usage

##data(ExampleData.BINfileData, envir = environment())
##write_R2BIN(CWOSL.SAR.Data, file="[your path]/output.bin")

}
\references{
DTU Nutech, 2016. The Squence Editor, Users Manual, February, 2016.
\url{http://www.nutech.dtu.dk/english/products-and-services/radiation-instruments/tl_osl_reader/manuals}
}
\seealso{
\link{read_BIN2R}, \linkS4class{Risoe.BINfileData},
\link{writeBin}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne
(France)
}
\keyword{IO}
